Simulation of a single-bubble process

The following function generates a time series which switches from a martingale to a mildly explosive process and then back to a martingale.

sim_dgp1(n, te = 0.4 * n, tf = 0.15 * n + te, c = 1, alpha = 0.6,
  sigma = 6.79)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te
A scalar in (0, tf) specifying the observation in which the bubble originates.

tf
A scalar in (te, n) specifying the observation in which the bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

Value

A numeric vector of length n.

Details

The data generating process is described by the following equation: $$X_t = X_{t-1}1\{t < \tau_e\}+ \delta_T X_{t-1}1\{\tau_e \leq t\leq \tau_f\} + \left(\sum_{k=\tau_f+1}^t \epsilon_k + X^*_{\tau_f}\right) 1\{t > \tau_f\} + \epsilon_t 1\{t \leq \tau_f\}$$

where the autoregressive coefficient \(\delta_T\) is given by:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\) and \(X_{\tau_f} = X_{\tau_e} + X^*\). During the pre- and post- bubble periods, \(N_0 = [1, \tau_e)\), X is a pure random walk process. During the bubble expansion period \(B = [\tau_e, \tau_f]\) is a mildly explosive process with expansion rate given by the autoregressive coefficient \(\delta_T\), and continues its martingale path for the subsequent period \(N_1 = (\tau_f, \tau]\).

For further details the user can refer to Phillips et al. (2015) p. 1054.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination date 40 and termination date 55
sim_dgp1(n = 100)

#>   [1] 100.00000000 102.48781905 110.50374586 109.84584741 103.53318386
#>   [6]  92.75650646  91.43922172 105.09896556 114.65956469 106.90391464
#>  [11] 121.64335334 113.12265785 111.43649253  99.85686819  94.85352743
#>  [16]  95.61237902  99.22517734 103.71173807  97.75352892 106.25995867
#>  [21]  99.68460331  96.70999069  86.77348143  67.79509537  63.72775092
#>  [26]  56.55058999  57.13098617  40.63346185  50.35649955  53.36912993
#>  [31]  40.43471220  47.67516752  59.43637745  70.23029699  65.57903942
#>  [36]  70.83120464  70.32355580  71.74569662  72.54236412  74.58939448
#>  [41]  84.44133252  93.44518892 100.74500767 114.45025984 120.93266578
#>  [46] 120.98904541 138.93980471 147.32201828 167.15739509 179.64888995
#>  [51] 187.01588680 201.42448328 213.46863516 228.88051509 254.77320914
#>  [56]  81.38651343  85.25552687  77.77682191  79.05235282  82.07357644
#>  [61]  76.50082301  71.34568241  77.79962446  65.69450799  48.74886808
#>  [66]  44.03513438  41.46755187  55.08183737  55.44971665  50.11153780
#>  [71]  51.32380720  40.22427673  48.39754893  55.71256032  56.18111020
#>  [76]  59.94012525  68.32894092  64.98184636  66.01923522  69.22559617
#>  [81]  70.13520017  54.30568664  48.67752556  43.97601129  46.84729191
#>  [86]  52.52747930  50.72317555  57.48911537  55.25432240  47.33637622
#>  [91]  40.40054118  38.03772161  39.73414771  40.43839641  31.56704721
#>  [96]  39.19062797  33.71816641  16.83448503  10.55085628   0.04195363

# 200 periods with bubble origination date 80 and termination date 110
sim_dgp1(n = 200)

#>   [1] 100.0000000  94.8109686  95.7508083  91.0853020  78.7655761  67.1169270
#>   [7]  59.9929799  50.9202312  47.5054305  47.4336089  64.8734695  65.3481728
#>  [13]  54.8320511  68.7543504  77.7321725  75.2923815  80.5817209  75.7266419
#>  [19]  70.2713685  67.2093459  70.7352354  69.4655353  66.1901039  71.1324445
#>  [25]  74.1121728  65.9732944  68.0463394  66.6727941  72.1197198  70.4345924
#>  [31]  63.8723137  59.7880338  63.7711199  63.2294207  74.0311798  69.8961844
#>  [37]  75.7786512  75.6034290  72.4205674  89.0421092  94.4686823  98.0501398
#>  [43]  96.9405298  91.4322508  87.8846097  83.9849977  83.2388262  83.0041686
#>  [49]  85.9242302  78.6945857  82.6905954  86.3957917  81.1511888  77.0429329
#>  [55]  66.7814446  77.9376809  64.0941932  68.9821517  72.7556728  63.6560694
#>  [61]  68.4937473  70.6482146  63.7572575  72.2695295  65.1490789  58.4705800
#>  [67]  54.5509440  35.0120254  37.9330693  33.2741079  38.3641353  45.4561939
#>  [73]  44.2107663  52.2008539  46.7293952  40.2399052  44.5097184  45.0993511
#>  [79]  40.3541361  41.7359898  37.0325302  24.5591836  25.6966510  26.1783339
#>  [85]  28.3704353  34.6417487  40.1222313  29.2966962  47.0665136  50.8061998
#>  [91]  51.6155062  41.1494697  44.0342860  48.1025389  37.5318361  30.8723346
#>  [97]  36.5601881  42.3447974  45.3774085  35.3509789  21.1329721  18.4505173
#> [103]  28.8538973  39.9825544  38.6257609  42.5596474  30.2170759  39.4157615
#> [109]  42.7465113  48.9359466  47.9736314  54.7050120  48.2625877  48.4861383
#> [115]  41.5504636  27.6204895  20.7472361  16.7420747  24.1233013  32.7284276
#> [121]  33.7937895  25.3534539  24.9550021  33.8890424  33.4979372  35.6476049
#> [127]  35.1116917  22.5216195  34.2073110  37.5122981  31.9362569  31.3843539
#> [133]  30.2109957  30.8882223  34.7977665  33.3089519  28.4084509  31.3006009
#> [139]  41.9215709  32.9711106  32.6258326  36.6145573  39.7241366  34.8469524
#> [145]  30.5813545  25.8000892  24.3526394  23.1477389  28.7662708  14.8724495
#> [151]   6.5706364   4.5667169  -5.2802135  -8.3242364 -14.6178835 -14.5343494
#> [157] -15.5671129  -8.1849318 -12.1412013  -3.1169723  11.0546514   7.4864063
#> [163]  -3.3158231  -7.3977381  -5.2794363   3.2010819   1.7537948   7.4065878
#> [169]   8.9832503  -4.1689335  -2.8199558   3.7733325  -2.1741256  -5.7011642
#> [175]  -0.2119444   6.8929640   6.8214606  16.6706137  23.5888721  19.6040685
#> [181]  19.1887361  18.9273995  10.5234100  -1.9329559  -3.0761883  -7.7362101
#> [187]  -5.1531496  -8.8459876  -0.4540497   2.3149688  16.3125355  18.5745771
#> [193]  15.0473806   7.2650532   6.2419854  12.4884170  10.8348427  16.1508856
#> [199]  11.9906895   7.1976581

# 200 periods with bubble origination date 100 and termination date 150
sim_dgp1(n = 200, te = 100, tf = 150)

#>   [1] 100.00000  95.19986  92.96358  90.88313  83.54042  74.06464  64.40124
#>   [8]  76.42013  64.82968  64.08985  69.09631  72.41617  61.70249  63.80151
#>  [15]  68.56674  68.89810  64.91712  73.85527  68.84771  59.95276  44.89614
#>  [22]  43.00967  48.80559  43.83843  54.57215  65.01395  75.89136  78.17155
#>  [29]  76.50414  85.64322  84.28265  81.97414  84.91083  93.79351 103.08741
#>  [36] 119.46893 122.41340 113.18218 104.06015 116.02370 114.48470 118.65095
#>  [43] 131.21284 133.95927 116.88653 114.96180 132.98474 141.98421 131.86471
#>  [50] 130.38742 134.77722 144.03342 141.93258 143.65320 134.59632 133.17719
#>  [57] 133.24001 129.44002 142.55877 158.46965 162.90734 160.11319 162.99072
#>  [64] 163.45490 178.39659 168.67792 165.83692 154.65170 155.48194 154.47772
#>  [71] 141.93079 141.39824 141.04697 155.60109 159.25869 144.02811 152.57405
#>  [78] 158.41366 146.49451 143.28301 146.44505 152.65230 148.94745 143.47512
#>  [85] 145.56380 126.08522 120.84730 123.44515 119.21320 123.77761 130.81978
#>  [92] 131.06929 113.64666 106.86580 103.31984 103.73748  88.76237  85.95100
#>  [99]  83.08022  85.63047  92.77573  93.70604  95.99506 103.18399 114.23525
#> [106] 122.69242 129.00746 134.78038 146.04421 149.62431 146.37709 157.25737
#> [113] 156.46869 154.82144 160.30693 176.79837 186.53449 195.00272 210.57170
#> [120] 217.33850 227.78274 227.26691 238.53445 241.41089 252.25124 256.96616
#> [127] 269.18523 273.35422 296.52629 309.23373 314.48870 325.80312 336.79684
#> [134] 336.40560 355.14971 372.38730 384.65553 398.82863 412.31626 428.99515
#> [141] 451.46439 459.86496 471.93811 487.85072 512.08304 530.14077 552.71424
#> [148] 576.63382 589.66646 607.11183  73.17365  76.71908  73.89975  77.52094
#> [155]  86.38343  77.92408  68.30633  61.17396  59.05033  74.15219  82.99230
#> [162]  93.29125  94.80897  90.59394  88.96180  83.27160  82.02603  84.51491
#> [169]  91.67615  91.19220  89.88146  87.41695  87.75165  76.05013  75.97073
#> [176]  77.34148  80.98170  85.16350  93.63926  89.27702  86.44840  83.81331
#> [183]  80.31799  80.29774  75.21761  69.53467  58.23039  59.28063  57.22108
#> [190]  51.95545  40.85064  45.44588  47.91725  50.17971  46.62746  45.27809
#> [197]  54.53631  46.02402  28.37537  28.91897