Simulation of a two-bubble process

The following data generating process is similar to sim_dgp1, with the difference that there are two episodes of midly explosive dynamics.

sim_dgp2(n, te1 = 0.2 * n, tf1 = 0.2 * n + te1, te2 = 0.6 * n,
  tf2 = 0.1 * n + te2, c = 1, alpha = 0.6, sigma = 6.79)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te1
A scalar in (0, n) specifying the observation in which the first bubble originates.

tf1
A scalar in (te1, n) specifying the observation in which the first bubble collapses.

te2
A scalar in (tf1, n) specifying the observation in which the second bubble originates.

tf2
A scalar in (te2, n) specifying the observation in which the second bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

Value

A numeric vector of length n.

Details

The data generating process is described by:

$$X_t = X_{t-1}1\{t \in N_0\}+ \delta_T X_{t-1}1\{t \in B_1 \cup B_2\} + \left(\sum_{k=\tau_{1f}+1}^t \epsilon_k + X^*_{\tau_{1f}}\right) 1\{t \in N_1\} $$

$$ + \left(\sum_{l=\tau_{2f}+1}^t \epsilon_l + X^*_{\tau_{2f}}\right) 1\{t \in N_2\} + \epsilon_t 1\{t \in N_0 \cup B_1 \cup B_2\}$$

where the autoregressive coefficient \(\delta_T\) is given:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\), \(X_{\tau_{1f}} = X_{\tau_{1e}} + X^*\) and \(X_{\tau_{2f}} = X_{\tau_{2e}} + X^*\). We use the notation \(N_0 = [1, \tau_{1e})\), \(B_1 = [\tau_{1e}, \tau_{1f}]\), \(N_1 = (\tau_{1f}, \tau_{2e})\), \(B_2 = [\tau_{2e}, \tau_{2f}]\), \(N_2 = (\tau_{2f}, \tau]\), where \(\tau\) is the last observation of the sample. After the collapse of the first bubble \(X_t\) resumes a martingale path until the time \(\tau_{2e}-1\), and a second episode of exuberance begins at \(\tau_{2e}\). The expansion process lasts until \(\tau_{2f}\) and collapses to a value of \(X^*_{\tau_{2f}}\). The process then continues on a martingale path until the end of the sample period \(\tau\). The expansion duration of the first bubble is assumed to be longer than that of the second bubble, i.e. \(\tau_{1f}-\tau_{1e}>\tau_{2f}-\tau_{2e}\).

For further details the user can refer to Phillips et al., (2015) p. 1055.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination dates 20/60 and termination dates 40/70 respectively
sim_dgp2(n = 100)

#>   [1] 100.00000  90.85669  87.86775  85.63597  80.87427  78.30794  85.47201
#>   [8]  81.78135  81.96633  84.57208  79.14960  87.04492  86.38099  84.35680
#>  [15]  73.12693  77.74505  74.82568  70.39934  62.56337  55.39819  63.68988
#>  [22]  70.63032  60.13077  54.07834  60.55122  72.19307  83.38523  80.38582
#>  [29]  81.18721  84.19972  97.70647 107.40122 111.89200 121.54707 127.11531
#>  [36] 132.81933 137.35275 135.55701 155.61911 169.65063  55.39819  57.41275
#>  [43]  55.68749  60.97608  52.90515  68.41407  66.11359  55.30245  57.70421
#>  [50]  54.93976  58.56754  50.84465  50.26263  65.64299  62.65195  76.15610
#>  [57]  62.43919  65.84506  70.27458  74.65183  72.08573  80.34432  98.15294
#>  [64]  99.68683 110.72506 123.36971 130.56249 146.26916 152.74728 173.61087
#>  [71]  74.65183  82.87316  78.67690  75.57784  79.61052  77.13663  75.40224
#>  [78]  74.41388  69.34567  64.86417  75.08392  70.44837  79.76658  83.77845
#>  [85]  94.19629  92.30336  97.20241  89.26268  89.81245  95.83349  90.68023
#>  [92]  87.06828  86.30557  84.27059  98.06736  93.74788  93.14255  84.06176
#>  [99]  82.56449  97.64233

# 200 periods with bubble origination dates 40/120 and termination dates 80/140 respectively
sim_dgp2(n = 200)

#>   [1] 100.00000 101.76723 104.68624  93.08279  85.75416  96.79226  97.02076
#>   [8] 105.62369 106.87693 107.89475 107.05878 114.42090 115.32274 109.75974
#>  [15] 114.43235 113.64454 127.15105 127.14248 131.02944 127.18501 122.31995
#>  [22] 118.00652 113.98057 111.33911 118.65919 107.49562 101.19268  97.45732
#>  [29] 102.60338 116.47216 121.69406 118.47611 101.34707  89.34532 100.16440
#>  [36]  91.66127  76.09044  73.81021  74.37441  75.09846  89.38115 103.31357
#>  [43]  95.75029 101.65922 112.44443 123.81420 119.90894 115.86226 113.89088
#>  [50] 116.78673 121.73668 121.62386 118.11794 134.69765 134.60442 140.18515
#>  [57] 147.05138 150.06072 169.20844 182.91593 196.85097 218.01328 229.19615
#>  [64] 240.33520 256.10585 258.18556 267.56575 284.82568 289.46761 312.67154
#>  [71] 323.61680 349.66320 358.04140 371.34740 384.77614 390.73404 417.11579
#>  [78] 435.38040 460.23534 487.97333  75.09846  75.06227  72.75494  70.15511
#>  [85]  69.76662  61.11458  59.94851  57.82562  58.16562  54.35343  55.28528
#>  [92]  59.15436  55.77677  45.77593  43.64446  40.28870  55.22465  67.92757
#>  [99]  70.60929  71.86347  65.35662  64.40575  64.23907  70.83081  79.06778
#> [106]  74.04351  70.37098  72.68641  76.48135  69.21744  59.32371  55.61043
#> [113]  56.22911  62.99086  55.37394  65.95790  69.40407  75.37781  76.49328
#> [120]  80.54900  88.13778  91.18422  92.10554 101.85226 108.53099 109.56402
#> [127] 107.08615 107.38498 114.74655 121.39014 122.00985 130.29783 147.24805
#> [134] 148.43866 151.87743 158.63803 173.18243 179.26077 188.14170 203.38790
#> [141]  80.54900  87.56506  72.56183  75.71831  81.69704  75.38416  70.02225
#> [148]  75.41667  84.77282  92.81141  90.57069  89.88632  86.74759  82.92435
#> [155]  76.25128  66.14325  60.21632  54.90959  47.32039  47.50989  48.51306
#> [162]  43.34937  31.64316  32.69377  34.87706  42.06701  50.16746  63.07333
#> [169]  50.32149  47.50524  51.72312  51.11737  48.99548  54.71954  57.77773
#> [176]  57.98208  58.05843  57.68469  58.27994  56.60204  63.78550  61.70924
#> [183]  60.88810  62.76989  65.26332  71.93719  65.24255  64.28904  53.76671
#> [190]  52.93328  63.80618  54.08793  56.75239  43.83842  41.87408  43.03733
#> [197]  50.98305  65.59670  65.50322  69.32216