Get House Price Data

Get House Price Data

ukhp_get(region = "england", item = "housePriceIndex", regexp = TRUE,
  start_date = NULL, end_date = NULL, ...)

Arguments

region
the region to select. If regexp is set to FALSE then the matching should be exact, see ukhp_avail_regions for available regions. If regexp is set to TRUE then partial matching is possible. Furthermore it is set to NULL selects all available regions.

item
the item to select. See ukhp_avail_items for the available categories.

regexp
use regular expression in sparql to search for regions.

start_date
the start date

end_date
the end date

...
query modifiers passed through rdf_modifiers.

See also

Examples

# \donttest{
# This is case sensitive
ukhp_get("england")

#> # A tibble: 1,245 x 3
#>    region  date       housePriceIndex
#>    <fct>   <date>               <dbl>
#>  1 england 1968-04-01            1.68
#>  2 england 1968-04-01            1.68
#>  3 england 1968-04-01            1.68
#>  4 england 1968-07-01            1.74
#>  5 england 1968-07-01            1.74
#>  6 england 1968-07-01            1.74
#>  7 england 1968-10-01            1.73
#>  8 england 1968-10-01            1.73
#>  9 england 1968-10-01            1.73
#> 10 england 1969-01-01            1.76
#> # ... with 1,235 more rows

# However you can use regular expression instead of exact match
ukhp_get("england", regexp = TRUE)

#> # A tibble: 1,245 x 3
#>    region  date       housePriceIndex
#>    <fct>   <date>               <dbl>
#>  1 england 1968-04-01            1.68
#>  2 england 1968-04-01            1.68
#>  3 england 1968-04-01            1.68
#>  4 england 1968-07-01            1.74
#>  5 england 1968-07-01            1.74
#>  6 england 1968-07-01            1.74
#>  7 england 1968-10-01            1.73
#>  8 england 1968-10-01            1.73
#>  9 england 1968-10-01            1.73
#> 10 england 1969-01-01            1.76
#> # ... with 1,235 more rows

# For all available regions
ukhp_avail_regions()

#>   [1] "north-down-and-ards"            "east-riding-of-yorkshire"      
#>   [3] "hounslow"                       "west-berkshire"                
#>   [5] "north-yorkshire"                "derry-and-strabane"            
#>   [7] "runnymede"                      "lewisham"                      
#>   [9] "thanet"                         "erewash"                       
#>  [11] "lisburn-and-castlereagh"        "east-renfrewshire"             
#>  [13] "crawley"                        "outer-london"                  
#>  [15] "vale-of-white-horse"            "ceredigion"                    
#>  [17] "tendring"                       "shepway"                       
#>  [19] "forest-heath"                   "gloucestershire"               
#>  [21] "newry-mourne-and-down"          "brighton-and-hove"             
#>  [23] "north-lanarkshire"              "leicestershire"                
#>  [25] "basingstoke-and-deane"          "chelmsford"                    
#>  [27] "craven"                         "windsor-and-maidenhead"        
#>  [29] "mid-suffolk"                    "west-midlands"                 
#>  [31] "fife"                           "worthing"                      
#>  [33] "north-east-lincolnshire"        "east-hertfordshire"            
#>  [35] "bracknell-forest"               "ryedale"                       
#>  [37] "sunderland"                     "county-durham"                 
#>  [39] "south-bucks"                    "west-devon"                    
#>  [41] "gateshead"                      "sedgemoor"                     
#>  [43] "harrogate"                      "eastleigh"                     
#>  [45] "wychavon"                       "swale"                         
#>  [47] "stratford-on-avon"              "harlow"                        
#>  [49] "liverpool"                      "bolton"                        
#>  [51] "east-lothian"                   "wakefield"                     
#>  [53] "torbay"                         "south-ribble"                  
#>  [55] "newham"                         "richmondshire"                 
#>  [57] "south-ayrshire"                 "east-sussex"                   
#>  [59] "west-midlands-region"           "newcastle-upon-tyne"           
#>  [61] "moray"                          "wigan"                         
#>  [63] "epping-forest"                  "south-holland"                 
#>  [65] "rochdale"                       "woking"                        
#>  [67] "pembrokeshire"                  "south-northamptonshire"        
#>  [69] "three-rivers"                   "surrey"                        
#>  [71] "northern-ireland"               "castle-point"                  
#>  [73] "kirklees"                       "ashfield"                      
#>  [75] "caerphilly"                     "north-kesteven"                
#>  [77] "north-east"                     "worcester"                     
#>  [79] "waltham-forest"                 "sefton"                        
#>  [81] "torfaen"                        "cannock-chase"                 
#>  [83] "derbyshire"                     "winchester"                    
#>  [85] "blackpool"                      "west-oxfordshire"              
#>  [87] "london"                         "southampton"                   
#>  [89] "basildon"                       "south-somerset"                
#>  [91] "test-valley"                    "stevenage"                     
#>  [93] "preston"                        "north-tyneside"                
#>  [95] "east-devon"                     "tandridge"                     
#>  [97] "lincoln"                        "city-of-london"                
#>  [99] "teignbridge"                    "breckland"                     
#> [101] "kettering"                      "hinckley-and-bosworth"         
#> [103] "conwy"                          "st-helens"                     
#> [105] "tameside"                       "rossendale"                    
#> [107] "west-lothian"                   "brent"                         
#> [109] "west-yorkshire"                 "halton"                        
#> [111] "solihull"                       "hammersmith-and-fulham"        
#> [113] "redcar-and-cleveland"           "cherwell"                      
#> [115] "mid-devon"                      "stockton-on-tees"              
#> [117] "knowsley"                       "north-hertfordshire"           
#> [119] "chorley"                        "northumberland"                
#> [121] "angus"                          "east-dunbartonshire"           
#> [123] "inverclyde"                     "merseyside"                    
#> [125] "hastings"                       "cardiff"                       
#> [127] "sheffield"                      "north-ayrshire"                
#> [129] "blaby"                          "harrow"                        
#> [131] "barnet"                         "brentwood"                     
#> [133] "havering"                       "warwickshire"                  
#> [135] "york"                           "kensington-and-chelsea"        
#> [137] "central-bedfordshire"           "st-albans"                     
#> [139] "somerset"                       "south-yorkshire"               
#> [141] "southwark"                      "darlington"                    
#> [143] "leicester"                      "trafford"                      
#> [145] "enfield"                        "tyne-and-wear"                 
#> [147] "bolsover"                       "devon"                         
#> [149] "north-warwickshire"             "south-oxfordshire"             
#> [151] "north-somerset"                 "gwynedd"                       
#> [153] "kent"                           "essex"                         
#> [155] "rhondda-cynon-taf"              "sutton"                        
#> [157] "tewkesbury"                     "fermanagh-and-omagh"           
#> [159] "portsmouth"                     "ribble-valley"                 
#> [161] "luton"                          "south-cambridgeshire"          
#> [163] "merton"                         "poole"                         
#> [165] "city-of-kingston-upon-hull"     "great-britain"                 
#> [167] "waverley"                       "mendip"                        
#> [169] "cambridgeshire"                 "rushcliffe"                    
#> [171] "antrim-and-newtownabbey"        "dover"                         
#> [173] "colchester"                     "tamworth"                      
#> [175] "united-kingdom"                 "wandsworth"                    
#> [177] "gloucester"                     "north-west"                    
#> [179] "worcestershire"                 "west-dorset"                   
#> [181] "cotswold"                       "salford"                       
#> [183] "kingston-upon-thames"           "spelthorne"                    
#> [185] "newcastle-under-lyme"           "cheshire-east"                 
#> [187] "east-staffordshire"             "city-of-dundee"                
#> [189] "allerdale"                      "hyndburn"                      
#> [191] "south-norfolk"                  "lancaster"                     
#> [193] "new-forest"                     "stafford"                      
#> [195] "cornwall"                       "amber-valley"                  
#> [197] "mole-valley"                    "wrexham"                       
#> [199] "cheshire-west-and-chester"      "suffolk-coastal"               
#> [201] "sevenoaks"                      "stoke-on-trent"                
#> [203] "shetland-islands"               "inner-london"                  
#> [205] "lewes"                          "barnsley"                      
#> [207] "derbyshire-dales"               "highland"                      
#> [209] "rushmoor"                       "east-midlands"                 
#> [211] "renfrewshire"                   "waveney"                       
#> [213] "milton-keynes"                  "shropshire"                    
#> [215] "neath-port-talbot"              "croydon"                       
#> [217] "barking-and-dagenham"           "tunbridge-wells"               
#> [219] "dumfries-and-galloway"          "vale-of-glamorgan"             
#> [221] "aberdeenshire"                  "hertfordshire"                 
#> [223] "oxford"                         "stockport"                     
#> [225] "falkirk"                        "warwick"                       
#> [227] "fylde"                          "causeway-coast-and-glens"      
#> [229] "uttlesford"                     "slough"                        
#> [231] "king's-lynn-and-west-norfolk"   "mid-sussex"                    
#> [233] "powys"                          "dudley"                        
#> [235] "st-edmundsbury"                 "scotland"                      
#> [237] "city-of-peterborough"           "orkney-islands"                
#> [239] "mid-and-east-antrim"            "north-norfolk"                 
#> [241] "city-of-bristol"                "south-kesteven"                
#> [243] "bexley"                         "cumbria"                       
#> [245] "selby"                          "yorkshire-and-the-humber"      
#> [247] "bromsgrove"                     "hackney"                       
#> [249] "horsham"                        "northamptonshire"              
#> [251] "exeter"                         "south-tyneside"                
#> [253] "thurrock"                       "taunton-deane"                 
#> [255] "mansfield"                      "south-east"                    
#> [257] "lancashire"                     "surrey-heath"                  
#> [259] "rochford"                       "wirral"                        
#> [261] "braintree"                      "east-northamptonshire"         
#> [263] "scottish-borders"               "great-yarmouth"                
#> [265] "gravesham"                      "hampshire"                     
#> [267] "guildford"                      "forest-of-dean"                
#> [269] "england"                        "harborough"                    
#> [271] "city-of-glasgow"                "pendle"                        
#> [273] "oadby-and-wigston"              "east-cambridgeshire"           
#> [275] "norwich"                        "midlothian"                    
#> [277] "isle-of-wight"                  "ashford"                       
#> [279] "newark-and-sherwood"            "city-of-derby"                 
#> [281] "argyll-and-bute"                "england-and-wales"             
#> [283] "bassetlaw"                      "reigate-and-banstead"          
#> [285] "west-dunbartonshire"            "arun"                          
#> [287] "wales"                          "sandwell"                      
#> [289] "stirling"                       "dorset"                        
#> [291] "east-lindsey"                   "belfast"                       
#> [293] "city-of-westminster"            "hart"                          
#> [295] "north-east-derbyshire"          "greater-manchester"            
#> [297] "rugby"                          "eden"                          
#> [299] "na-h-eileanan-siar"             "weymouth-and-portland"         
#> [301] "hartlepool"                     "gosport"                       
#> [303] "cambridge"                      "redditch"                      
#> [305] "oldham"                         "hillingdon"                    
#> [307] "aylesbury-vale"                 "doncaster"                     
#> [309] "warrington"                     "isle-of-anglesey"              
#> [311] "eastbourne"                     "havant"                        
#> [313] "rother"                         "malvern-hills"                 
#> [315] "northampton"                    "wealden"                       
#> [317] "daventry"                       "birmingham"                    
#> [319] "herefordshire"                  "camden"                        
#> [321] "newport"                        "medway"                        
#> [323] "islington"                      "north-devon"                   
#> [325] "walsall"                        "north-west-leicestershire"     
#> [327] "wolverhampton"                  "merthyr-tydfil"                
#> [329] "east-hampshire"                 "south-staffordshire"           
#> [331] "city-of-aberdeen"               "south-lanarkshire"             
#> [333] "swindon"                        "city-of-edinburgh"             
#> [335] "swansea"                        "west-sussex"                   
#> [337] "staffordshire-moorlands"        "haringey"                      
#> [339] "mid-ulster"                     "copeland"                      
#> [341] "nuneaton-and-bedworth"          "broxbourne"                    
#> [343] "epsom-and-ewell"                "corby"                         
#> [345] "calderdale"                     "north-dorset"                  
#> [347] "lincolnshire"                   "tonbridge-and-malling"         
#> [349] "wyre"                           "ipswich"                       
#> [351] "fenland"                        "broadland"                     
#> [353] "west-lindsey"                   "bridgend"                      
#> [355] "wellingborough"                 "rutland"                       
#> [357] "ealing"                         "oxfordshire"                   
#> [359] "scarborough"                    "clackmannanshire"              
#> [361] "flintshire"                     "city-of-nottingham"            
#> [363] "manchester"                     "high-peak"                     
#> [365] "east-dorset"                    "southend-on-sea"               
#> [367] "wiltshire"                      "north-lincolnshire"            
#> [369] "telford-and-wrekin"             "tower-hamlets"                 
#> [371] "carlisle"                       "east-of-england"               
#> [373] "cheltenham"                     "bedford"                       
#> [375] "perth-and-kinross"              "coventry"                      
#> [377] "chiltern"                       "stroud"                        
#> [379] "lichfield"                      "west-lancashire"               
#> [381] "south-west"                     "monmouthshire"                 
#> [383] "hambleton"                      "babergh"                       
#> [385] "bromley"                        "carmarthenshire"               
#> [387] "south-hams"                     "leeds"                         
#> [389] "canterbury"                     "hertsmere"                     
#> [391] "wycombe"                        "wokingham"                     
#> [393] "elmbridge"                      "middlesbrough"                 
#> [395] "south-lakeland"                 "bradford"                      
#> [397] "staffordshire"                  "bournemouth"                   
#> [399] "norfolk"                        "chichester"                    
#> [401] "east-ayrshire"                  "broxtowe"                      
#> [403] "south-gloucestershire"          "maidstone"                     
#> [405] "purbeck"                        "burnley"                       
#> [407] "barrow-in-furness"              "charnwood"                     
#> [409] "city-of-plymouth"               "dartford"                      
#> [411] "boston"                         "south-derbyshire"              
#> [413] "welwyn-hatfield"                "gedling"                       
#> [415] "christchurch"                   "blackburn-with-darwen"         
#> [417] "denbighshire"                   "bath-and-north-east-somerset"  
#> [419] "maldon"                         "richmond-upon-thames"          
#> [421] "nottinghamshire"                "dacorum"                       
#> [423] "chesterfield"                   "adur"                          
#> [425] "redbridge"                      "buckinghamshire"               
#> [427] "wyre-forest"                    "armagh-banbridge-and-craigavon"
#> [429] "rotherham"                      "greenwich"                     
#> [431] "west-somerset"                  "melton"                        
#> [433] "huntingdonshire"                "reading"                       
#> [435] "fareham"                        "suffolk"                       
#> [437] "watford"                        "blaenau-gwent"                 
#> [439] "bury"                           "lambeth"                       
#> [441] "torridge"                      

# For all available items
ukhp_avail_items()

#>  [1] "averagePrice"                             
#>  [2] "housePriceIndex"                          
#>  [3] "percentageChange"                         
#>  [4] "percentageAnnualChange"                   
#>  [5] "salesVolume"                              
#>  [6] "averagePriceDetached"                     
#>  [7] "housePriceIndexDetached"                  
#>  [8] "percentageChangeDetached"                 
#>  [9] "percentageAnnualChangeDetached"           
#> [10] "averagePriceSemiDetached"                 
#> [11] "housePriceIndexSemiDetached"              
#> [12] "percentageChangeSemiDetached"             
#> [13] "percentageAnnualChangeSemiDetached"       
#> [14] "averagePriceTerraced"                     
#> [15] "housePriceIndexTerraced"                  
#> [16] "percentageChangeTerraced"                 
#> [17] "percentageAnnualChangeTerraced"           
#> [18] "averagePriceFlatMaisonette"               
#> [19] "housePriceIndexFlatMaisonette"            
#> [20] "percentageChangeFlatMaisonette"           
#> [21] "percentageAnnualChangeFlatMaisonette"     
#> [22] "averagePriceCash"                         
#> [23] "housePriceIndexCash"                      
#> [24] "percentageChangeCash"                     
#> [25] "percentageAnnualChangeCash"               
#> [26] "averagePriceMortgage"                     
#> [27] "housePriceIndexMortgage"                  
#> [28] "percentageChangeMortgage"                 
#> [29] "percentageAnnualChangeMortgage"           
#> [30] "averagePriceFirstTimeBuyer"               
#> [31] "housePriceIndexFirstTimeBuyer"            
#> [32] "percentageChangeFirstTimeBuyer"           
#> [33] "percentageAnnualChangeFirstTimeBuyer"     
#> [34] "averagePriceFormerOwnerOccupier"          
#> [35] "housePriceIndexFormerOwnerOccupier"       
#> [36] "percentageChangeFormerOwnerOccupier"      
#> [37] "percentageAnnualChangeFormerOwnerOccupier"
#> [38] "averagePriceNewBuild"                     
#> [39] "housePriceIndexNewBuild"                  
#> [40] "percentageChangeNewBuild"                 
#> [41] "percentageAnnualChangeNewBuild"           
#> [42] "averagePriceExistingProperty"             
#> [43] "housePriceIndexExistingProperty"          
#> [44] "percentageChangeExistingProperty"         
#> [45] "percentageAnnualChangeExistingProperty"   
#> [46] "salesVolumeExistingProperty"              
#> [47] "salesVolumeNewBuild"                      
#> [48] "salesVolumeCash"                          
#> [49] "salesVolumeMortgage"                      

ukhp_get(c("endland", "wales"), item = c("salesVolume", "housePriceIndexDetached"))

#> # A tibble: 590 x 4
#>    region            date       salesVolume housePriceIndexDetached
#>    <fct>             <date>           <dbl>                   <dbl>
#>  1 england-and-wales 1996-01-01       53652                    27.8
#>  2 england-and-wales 1995-01-01       70158                    27.8
#>  3 england-and-wales 1995-10-01       72525                    27.8
#>  4 england-and-wales 1995-01-01       50028                    28.0
#>  5 england-and-wales 1996-01-01       72070                    28  
#>  6 england-and-wales 1996-01-01       50628                    28.0
#>  7 england-and-wales 1995-04-01       66877                    28.1
#>  8 england-and-wales 1995-10-01       66925                    28.1
#>  9 england-and-wales 1995-04-01       59413                    28.2
#> 10 england-and-wales 1995-07-01       70442                    28.2
#> # ... with 580 more rows

# }