Wild Bootstrap Critical values

wb_cv performs the Harvey et al. (2016) wild bootstrap re-sampling scheme, which is asymptotically robust to non-stationary volatility, to generate critical values for the recursive unit root tests.

wb_cv(data, minw, nboot = 1000, dist_rad = FALSE)

Arguments

data
A univariate or multivariate numeric ts object, data.frame or matrix. The estimation process cannot handle NA values.

minw
A positive integer. The minimum window size, which defaults to \((0.01 + 1.8/\sqrt(T))*T\).

nboot
A positive integer indicating the number of bootstraps. Default is 1000 repetitions.

dist_rad
Logical. If TRUE then the Rademacher distribution will be used.

Value

A list that contains the critical values for ADF, BADF, BSADF and GSADF t-statistics.

Details

This approach involves applying a wild bootstrap re-sampling scheme to construct the bootstrap analogue of the PWY test which is asymptotically robust to non-stationary volatility.

References

Harvey, D. I., Leybourne, S. J., Sollis, R., & Taylor, A. M. R. (2016). Tests for explosive financial bubbles in the presence of non-stationary volatility. Journal of Empirical Finance, 38(Part B), 548-574.

See also

Examples

# Simulate bubble processes
dta <- data.frame("dg1" = sim_dgp1(n = 100), "dgp2" = sim_dgp2(n = 100))

# Default minimum window
wb <- wb_cv(dta)

# Change the minimum window and the number of bootstraps
wb <- wb_cv(dta, nboot = 1500,  minw = 20)

# Use parallel computing (utilizing all available cores)
wb <- wb_cv(dta, parallel = TRUE)

#> Error in wb_cv(dta, parallel = TRUE): unused argument (parallel = TRUE)