exuber: Econometric Analysis of Explosive Time Series

Testing for and dating periods of explosive dynamics (exuberance) in time series using the univariate and panel recursive unit root tests proposed by Phillips et al. (2015) doi:10.1111/iere.12132 and Pavlidis et al. (2016) doi:10.1007/s11146-015-9531-2 .The recursive least-squares algorithm utilizes the matrix inversion lemma to avoid matrix inversion which results in significant speed improvements. Simulation of a variety of periodically-collapsing bubble processes. Details can be found in Vasilopoulos et al. (2022) doi:10.18637/jss.v103.i10 .

Package options

exuber.show_progress

Should lengthy operations such as radf_mc_cv() show a progress bar? Default: TRUE

exuber.parallel

Should lengthy operations use parallel computation? Default: TRUE in an interactive session, FALSE otherwise (scripts, knitr, R CMD check), since starting workers costs a few seconds; set it to TRUE in a script to opt in. The worker cluster is started once per session and reused. Honored by the radf_*_cv()/radf_*_distr() simulation engines (radf_mc_cv(), radf_wb_cv(), radf_sb_cv(), radf_recovery_cv(), radf_common_cv()); the standalone tests and monitors run serially regardless.

exuber.ncores

How many cores to use for parallel computation. Default: system cores - 1 (2 in a non-interactive session), capped by the MC_CORES environment variable when it is set.

exuber.global_seed

When chosen automatically feeds into all functions with random-number generation. Default: NA

See also

Author

Maintainer: Kostas Vasilopoulos k.vasilopoulo@gmail.com

Authors:

Kostas Vasilopoulos k.vasilopoulo@gmail.com

Efthymios Pavlidis e.pavlidis@lancaster.ac.uk

Enrique Martínez-García emg.economics@gmail.com

Simon Spavound simon.spavound@googlemail.com