Recursive Augmented Dickey-Fuller Test

radf returns the t-statistics from a recursive Augmented Dickey-Fuller test.

radf(data, minw = NULL, lag = 0L)

Arguments

data
A univariate or multivariate numeric ts object, data.frame or matrix. The estimation process cannot handle NA values.

minw
A positive integer. The minimum window size, which defaults to \((0.01 + 1.8/\sqrt(T))*T\).

lag
A non-negative integer. The lag of the Augmented Dickey-Fuller regression.

Value

A list that contains the t-statistic (sequence) for:

adf

Augmented Dickey-Fuller

badf

Backward Augmented Dickey-Fuller

sadf

Supremum Augmented Dickey-Fuller

bsadf

Backward Supremum Augmented Dickey-Fuller

gsadf

Generalized Supremum Augmented Dickey-Fuller

References

Phillips, P. C. B., Wu, Y., & Yu, J. (2011). Explosive Behavior in The 1990s Nasdaq: When Did Exuberance Escalate Asset Values? International Economic Review, 52(1), 201-226.

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 56(4), 1043-1078.

Examples

# We will use simulated data that are stored as data with the name `sim_data`
sim_data

#> # A tibble: 100 x 5
#>    psy1  psy2  evans div   blan  
#>    <sim> <sim> <sim> <sim> <sim> 
#>  1 100   100   0.5   68.4  0.1   
#>  2 90.3  102   0.563 66    0.0408
#>  3 93.4  110   0.583 78.9  0.121 
#>  4 92.9  116   0.575 75.6  0.0195
#>  5 106   127   0.545 74.8  0.0216
#>  6 102   114   0.577 77.7  0.0117
#>  7 109   120   0.658 80.8  0.0265
#>  8 105   130   0.7   91.4  0.0238
#>  9 112   128   0.684 87.4  0.0332
#> 10 124   131   0.735 72.2  0.0342
#> # ... with 90 more rows

rsim <- radf(sim_data)

str(rsim)

#> List of 7
#>  $ adf        : Named num [1:5] -2.46 -2.86 -5.83 -1.95 -5.15
#>   ..- attr(*, "names")= chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ badf       : num [1:81, 1:5] -2.31 -2.36 -2.49 -2.26 -1.63 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ sadf       : Named num [1:5] 1.95 7.88 5.28 1.11 3.93
#>   ..- attr(*, "names")= chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ bsadf      : num [1:81, 1:5] -2.31 -2.36 -2.49 -2.26 -1.26 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ gsadf      : Named num [1:5] 5.19 7.88 5.99 1.34 10.95
#>   ..- attr(*, "names")= chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ bsadf_panel: num [1:81] -0.0702 -1.5266 -1.2379 -1.101 -0.7169 ...
#>  $ gsadf_panel: num 2.41
#>  - attr(*, "index")= num [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>  - attr(*, "lag")= int 0
#>  - attr(*, "n")= int 100
#>  - attr(*, "minw")= num 19
#>  - attr(*, "series_names")= chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  - attr(*, "class")= chr "radf"

# We would also use data that contain a Date column
sim_data_wdate

#> # A tibble: 100 x 6
#>    psy1  psy2  evans div   blan   date      
#>    <sim> <sim> <sim> <sim> <sim>  <date>    
#>  1 100   100   0.5   68.4  0.1    2000-01-01
#>  2 90.3  102   0.563 66    0.0408 2000-02-01
#>  3 93.4  110   0.583 78.9  0.121  2000-03-01
#>  4 92.9  116   0.575 75.6  0.0195 2000-04-01
#>  5 106   127   0.545 74.8  0.0216 2000-05-01
#>  6 102   114   0.577 77.7  0.0117 2000-06-01
#>  7 109   120   0.658 80.8  0.0265 2000-07-01
#>  8 105   130   0.7   91.4  0.0238 2000-08-01
#>  9 112   128   0.684 87.4  0.0332 2000-09-01
#> 10 124   131   0.735 72.2  0.0342 2000-10-01
#> # ... with 90 more rows

rsim_wdate <- radf(sim_data_wdate)

#> Using `date` as index variable.

str(rsim_wdate)

#> List of 7
#>  $ adf        : Named num [1:5] -2.46 -2.86 -5.83 -1.95 -5.15
#>   ..- attr(*, "names")= chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ badf       : num [1:81, 1:5] -2.31 -2.36 -2.49 -2.26 -1.63 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ sadf       : Named num [1:5] 1.95 7.88 5.28 1.11 3.93
#>   ..- attr(*, "names")= chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ bsadf      : num [1:81, 1:5] -2.31 -2.36 -2.49 -2.26 -1.26 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ gsadf      : Named num [1:5] 5.19 7.88 5.99 1.34 10.95
#>   ..- attr(*, "names")= chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ bsadf_panel: num [1:81] -0.0702 -1.5266 -1.2379 -1.101 -0.7169 ...
#>  $ gsadf_panel: num 2.41
#>  - attr(*, "index")= Date[1:100], format: "2000-01-01" "2000-02-01" ...
#>  - attr(*, "lag")= int 0
#>  - attr(*, "n")= int 100
#>  - attr(*, "minw")= num 19
#>  - attr(*, "series_names")= chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  - attr(*, "class")= chr "radf"

index(rsim_wdate)

#>   [1] "2000-01-01" "2000-02-01" "2000-03-01" "2000-04-01" "2000-05-01"
#>   [6] "2000-06-01" "2000-07-01" "2000-08-01" "2000-09-01" "2000-10-01"
#>  [11] "2000-11-01" "2000-12-01" "2001-01-01" "2001-02-01" "2001-03-01"
#>  [16] "2001-04-01" "2001-05-01" "2001-06-01" "2001-07-01" "2001-08-01"
#>  [21] "2001-09-01" "2001-10-01" "2001-11-01" "2001-12-01" "2002-01-01"
#>  [26] "2002-02-01" "2002-03-01" "2002-04-01" "2002-05-01" "2002-06-01"
#>  [31] "2002-07-01" "2002-08-01" "2002-09-01" "2002-10-01" "2002-11-01"
#>  [36] "2002-12-01" "2003-01-01" "2003-02-01" "2003-03-01" "2003-04-01"
#>  [41] "2003-05-01" "2003-06-01" "2003-07-01" "2003-08-01" "2003-09-01"
#>  [46] "2003-10-01" "2003-11-01" "2003-12-01" "2004-01-01" "2004-02-01"
#>  [51] "2004-03-01" "2004-04-01" "2004-05-01" "2004-06-01" "2004-07-01"
#>  [56] "2004-08-01" "2004-09-01" "2004-10-01" "2004-11-01" "2004-12-01"
#>  [61] "2005-01-01" "2005-02-01" "2005-03-01" "2005-04-01" "2005-05-01"
#>  [66] "2005-06-01" "2005-07-01" "2005-08-01" "2005-09-01" "2005-10-01"
#>  [71] "2005-11-01" "2005-12-01" "2006-01-01" "2006-02-01" "2006-03-01"
#>  [76] "2006-04-01" "2006-05-01" "2006-06-01" "2006-07-01" "2006-08-01"
#>  [81] "2006-09-01" "2006-10-01" "2006-11-01" "2006-12-01" "2007-01-01"
#>  [86] "2007-02-01" "2007-03-01" "2007-04-01" "2007-05-01" "2007-06-01"
#>  [91] "2007-07-01" "2007-08-01" "2007-09-01" "2007-10-01" "2007-11-01"
#>  [96] "2007-12-01" "2008-01-01" "2008-02-01" "2008-03-01" "2008-04-01"

# For lag = 1 and minimum window = 20
rsim_20 <- radf(sim_data, minw = 20, lag = 1)

str(rsim_20)

#> List of 7
#>  $ adf        : Named num [1:5] -2.43 -2.81 -5.04 -2.12 -4.51
#>   ..- attr(*, "names")= chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ badf       : num [1:79, 1:5] -1.902 -1.734 -1.094 -0.267 -0.309 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ sadf       : Named num [1:5] 1.682 5.191 -2.34 0.908 3.49
#>   ..- attr(*, "names")= chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ bsadf      : num [1:79, 1:5] -1.902 -1.734 -1.094 0.377 0.197 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ gsadf      : Named num [1:5] 4.67 5.45 4.2 1.01 8.5
#>   ..- attr(*, "names")= chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  $ bsadf_panel: num [1:79] -1.002 -0.787 -0.48 -0.455 -0.486 ...
#>  $ gsadf_panel: num 1.88
#>  - attr(*, "index")= num [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>  - attr(*, "lag")= num 1
#>  - attr(*, "n")= int 100
#>  - attr(*, "minw")= num 20
#>  - attr(*, "series_names")= chr [1:5] "psy1" "psy2" "evans" "div" ...
#>  - attr(*, "class")= chr "radf"