Simulation of a Blanchard (1979) bubble process

Simulation of a Blanchard (1979) rational bubble process.

sim_blan(n, pi = 0.7, sigma = 0.03, r = 0.05)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

pi
A positive value in (0, 1) which governs the probability of the bubble continuing to grow.

sigma
A positive scalar indicating the standard deviation of the innovations.

r
A positive scalar that determines the growth rate of the bubble process.

Value

A numeric vector of length n.

Details

Blanchard's bubble process has two regimes, which occur with probability \(\pi\) and \(1-\pi\). In the first regime, the bubble grows exponentially, whereas in the second regime, the bubble collapses to a white noise.

With probability \(\pi\): $$B_{t+1} = \frac{1+r}{\pi}B_t+\epsilon_{t+1}$$ With probability \(1 - \pi\): $$B_{t+1} = \epsilon_{t+1}$$

where r is a positive constant and \(\epsilon \sim iid(0, \sigma^2)\).

References

Blanchard, O. J. (1979). Speculative bubbles, crashes and rational expectations. Economics letters, 3(4), 387-389.

See also

Examples

sim_blan(n = 100)

#>   [1] 1.0000000000 1.5044349516 0.0136663035 0.0082074077 0.0036984319
#>   [6] 0.0110823758 0.0218268655 0.0480192436 0.0653238090 0.1131860684
#>  [11] 0.1477189418 0.2040898572 0.3064469662 0.4707605891 0.7014608552
#>  [16] 0.0270895801 0.0300487280 0.0132735518 0.0167726606 0.0082474756
#>  [21] 0.0588150939 0.1275694272 0.0400255584 0.0802283141 0.0318977523
#>  [26] 0.0820676344 0.1275763916 0.1946864926 0.0004148071 0.0254236110
#>  [31] 0.0268990174 0.0732599701 0.0210401738 0.0243010776 0.0513516938
#>  [36] 0.0501854370 0.0847297349 0.0964721421 0.1280507084 0.2104421279
#>  [41] 0.3498599931 0.5819014097 0.8078654069 1.2212987568 1.8550141930
#>  [46] 0.0148626029 0.0058326533 0.0584005421 0.0841480113 0.1476947604
#>  [51] 0.0300152341 0.0106278439 0.0199550637 0.0595145022 0.0850109958
#>  [56] 0.1074097492 0.0220793617 0.0732551373 0.0844697141 0.0376664479
#>  [61] 0.0113018393 0.0693245281 0.1214876608 0.0070681462 0.0102631637
#>  [66] 0.0166109192 0.0780888811 0.0297210124 0.0879137260 0.1474779727
#>  [71] 0.1958249800 0.2524940764 0.0165279649 0.0166633939 0.0309742357
#>  [76] 0.0621325786 0.0269080946 0.0119138702 0.0377455736 0.0008485778
#>  [81] 0.0144447105 0.0033418798 0.0130216780 0.0295093272 0.0032873342
#>  [86] 0.0396630448 0.0791859826 0.0656195357 0.0003563488 0.0372713180
#>  [91] 0.0181835739 0.0175634855 0.0673842989 0.1079299381 0.1818182860
#>  [96] 0.3138950638 0.4815260474 0.0119868896 0.0537220519 0.0112337962