Simulation of a Blanchard (1979) bubble process

Simulation of a Blanchard (1979) rational bubble process.

sim_blan(n, pi = 0.7, sigma = 0.03, r = 0.05)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

pi
A positive value in (0, 1) which governs the probability of the bubble continuing to grow.

sigma
A positive scalar indicating the standard deviation of the innovations.

r
A positive scalar that determines the growth rate of the bubble process.

Value

A numeric vector of length n.

Details

Blanchard's bubble process has two regimes, which occur with probability \(\pi\) and \(1-\pi\). In the first regime, the bubble grows exponentially, whereas in the second regime, the bubble collapses to a white noise.

With probability \(\pi\): $$B_{t+1} = \frac{1+r}{\pi}B_t+\epsilon_{t+1}$$ With probability \(1 - \pi\): $$B_{t+1} = \epsilon_{t+1}$$

where r is a positive constant and \(\epsilon \sim iid(0, \sigma^2)\).

References

Blanchard, O. J. (1979). Speculative bubbles, crashes and rational expectations. Economics letters, 3(4), 387-389.

See also

Examples

sim_blan(n = 100)

#>   [1] 1.000000000 1.516075352 2.306306028 0.009153111 0.050408679 0.080151361
#>   [7] 0.124819086 0.146204812 0.252414530 0.404821352 0.587888315 0.880737814
#>  [13] 1.332991500 1.964330930 2.998251318 4.482236967 6.694291946 0.004030664
#>  [19] 0.041277787 0.032826688 0.012957420 0.014573255 0.040832958 0.007217739
#>  [25] 0.007358378 0.055497406 0.002296540 0.024934514 0.078494416 0.095525614
#>  [31] 0.164465721 0.187079811 0.283400699 0.441729082 0.011561454 0.040733764
#>  [37] 0.045561347 0.044748723 0.033139992 0.058413163 0.091980173 0.179344338
#>  [43] 0.280243354 0.401342852 0.626957781 0.911509527 1.332323278 1.985777492
#>  [49] 2.949498804 0.030199458 0.024490626 0.073394941 0.047644970 0.092024842
#>  [55] 0.078026098 0.102507177 0.158310069 0.002964451 0.001678839 0.015157986
#>  [61] 0.035576044 0.124387118 0.178786563 0.237119236 0.282713773 0.053008530
#>  [67] 0.042401810 0.092039660 0.116174355 0.003268625 0.015450566 0.012478844
#>  [73] 0.024037523 0.028947566 0.024943691 0.031306207 0.038975122 0.083491197
#>  [79] 0.155034999 0.215019557 0.275236093 0.403003833 0.585146089 0.004145786
#>  [85] 0.073194320 0.099602301 0.033184280 0.018706745 0.031464809 0.014111359
#>  [91] 0.061490285 0.090220719 0.014765218 0.011537266 0.038018542 0.005834388
#>  [97] 0.008005625 0.046351716 0.004555563 0.034182781