Simulation of a Blanchard (1979) bubble process

Simulation of a Blanchard (1979) rational bubble process.

sim_blan(n, pi = 0.7, sigma = 0.03, r = 0.05, seed = NULL)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

pi
A positive value in (0, 1) which governs the probability of the bubble continuing to grow.

sigma
A positive scalar indicating the standard deviation of the innovations.

r
A positive scalar that determines the growth rate of the bubble process.

seed
An object specifying if and how the random number generator(rng) shold be initialized. Either NULL or an integer will be used in a call to `set.seed` before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute.

Value

A numeric vector of length n.

Details

Blanchard's bubble process has two regimes, which occur with probability \(\pi\) and \(1-\pi\). In the first regime, the bubble grows exponentially, whereas in the second regime, the bubble collapses to a white noise.

With probability \(\pi\): $$B_{t+1} = \frac{1+r}{\pi}B_t+\epsilon_{t+1}$$ With probability \(1 - \pi\): $$B_{t+1} = \epsilon_{t+1}$$

where r is a positive constant and \(\epsilon \sim iid(0, \sigma^2)\).

References

Blanchard, O. J. (1979). Speculative bubbles, crashes and rational expectations. Economics letters, 3(4), 387-389.

See also

Examples

sim_blan(n = 100)

#>   [1] 1.000000000 1.526658870 2.290384739 3.442337294 0.012204155 0.013442736
#>   [7] 0.010560304 0.007168707 0.013892927 0.021605596 0.014077011 0.037159617
#>  [13] 0.022082039 0.063765518 0.109603232 0.023714181 0.031663328 0.019586432
#>  [19] 0.018434117 0.004616175 0.019318907 0.103403061 0.012626237 0.008213506
#>  [25] 0.010816005 0.028732293 0.024120814 0.070685621 0.030173485 0.110275832
#>  [31] 0.069676696 0.073902343 0.112317947 0.145320262 0.194423340 0.269836919
#>  [37] 0.020456341 0.023789213 0.018171989 0.039363123 0.075545151 0.114168441
#>  [43] 0.198047612 0.285774752 0.446838673 0.003197609 0.040103841 0.073577497
#>  [49] 0.178554888 0.004081746 0.042745863 0.017887886 0.017522525 0.026695961
#>  [55] 0.013372510 0.061152041 0.007939852 0.009115448 0.016789490 0.015237674
#>  [61] 0.075486620 0.131001937 0.227004303 0.344155073 0.050534872 0.044789110
#>  [67] 0.062504365 0.012487824 0.022152624 0.035146499 0.025139781 0.027040059
#>  [73] 0.016628239 0.044988995 0.072139921 0.112070524 0.122113823 0.189241555
#>  [79] 0.010850843 0.058246393 0.098550558 0.015422463 0.016496986 0.050779986
#>  [85] 0.096700781 0.140170773 0.156725217 0.203973459 0.330864620 0.514517338
#>  [91] 0.768110417 0.027993754 0.013152331 0.007652451 0.039389499 0.042954732
#>  [97] 0.050859315 0.013249037 0.014477924 0.037981217
#> attr(,"seed")
#>   [1]       10403         564  2035000578 -1381714395   681381691  -876313188
#>   [7]  1567300003 -1751640244   584022128   523616267 -2099322512  -540274505
#>  [13]   329484621   922114502 -1666800787   983008538    11220598  -302418415
#>  [19]   648919070  1209572953    69194679 -1322373960 -1812214081  -313235672
#>  [25]   880657860 -1101350721  1219496644  1376029027   -94720479 -1247456310
#>  [31]  -812498271  -871851154  2004400050  -634738355  1960921034 -1684573683
#>  [37]  1578520115   703696564 -1252050389  -333458220  1780341528   809102435
#>  [43]   456516968   362057471  2037273717  -148264258   146161333  1803703810
#>  [49]  1055938574  -931056023   671951062  1927799857   730615519  1334791648
#>  [55]  1826246775  1065399488 -1610171012 -1065871129 -1678003156   904319644
#>  [61]  -923573474 -1468489600 -1639041803  1935536047   192708323  -170213455
#>  [67] -1407391964   196039782   882724128   455953266 -1356261173  1635823241
#>  [73]  -420615891 -1895001429  1490759554   423230232  1816435058 -1068922484
#>  [79]  1357675641   125624891  1939805271  1720916413   927767392   419570530
#>  [85]  1168684244   204988814 -2040917729 -1733612283  -131915759  -263509689
#>  [91]  -458380530  2094529540  -332544666   777737656 -1263452723   996137767
#>  [97]  -949464693   506355689   404591868 -1276591202  1097743768   388086794
#> [103]  -265547149  -484037503   -11731723 -1888097549   -36194934  -101384736
#> [109]   504252858  1504889476  -568060847  -139360717 -1080836385  1152536757
#> [115]  1202226392   408511738  -249409364  1191175750   556011559   763582749
#> [121]   529715705 -2049327633 -1370689546 -1372838324   746712526   956831152
#> [127]  1316728197 -1282935841 -2071326029  1267879521  -493696236 -1016661322
#> [133]  -843864464  2077095202  -552499621  2011069625   265642525  -291092901
#> [139]  1636366034  -496738744  -511810590    -8253668   623976777    22877771
#> [145]  1632279783   767617741  1451186992 -1829423950  -900353436  1128265982
#> [151] -1466225233  -393963467  1012562017  1691932183   830566398 -1719952140
#> [157]  1404340950  2107218728  -674962467  1056749847 -1685018725 -1783451239
#> [163]  -966227860  -745094162  1076282792  1119998650  1239387075  -380274959
#> [169] -1688293147  2025648739 -1342981286 -2062295984   478034218 -2133276588
#> [175]  1162336993  1769632899   944735023 -1506031611  1718075560  1860682570
#> [181]  1545424572  -719495434  1986276407  -438298035  1985329801   769680959
#> [187]  1706765542  -642019460   274101886  1475163488  1469077909   696626895
#> [193]   133190531   331084241 -2095267388  1218798534 -1555676608   679695890
#> [199]  -224850453   623097449   875829325  -397150133   -92377758   801868280
#> [205] -1476633326  1309192300   447659737 -1093350821   956788599  1507823005
#> [211]  1015328192 -1879381310 -1060908556  -876009938   -37295169  1613228261
#> [217]   649965041 -1826508249 -1485178514 -1987128860  -401184954   421717656
#> [223]   591582381  2065680455 -1583999317   -96847735  1945094172    60800958
#> [229]  1309724984   866580330 -1548872685  1083909729 -1327296107   660550227
#> [235]  -226586454  1365724800  1563006746  -380078620 -2044925903   744739411
#> [241]  1319971263  -704012779   914470072  1488626522  -306012404  1923513190
#> [247]   581312519  1077793981 -1480180519  -903317873 -2051628650   195775404
#> [253] -1274341714    74778512   271285925  -847322369  -352065069  -537253631
#> [259]  -228874444  1018517014 -1449745392   764960834 -1660420485  1630385177
#> [265]  2132446525   -71121925  -198812878  1161588392   895431682   424816252
#> [271]  1524231209  1954536427 -1812456185   172883757  2042058512  1748291602
#> [277] -1454481404   500316190   495155023   849236245   729906241  1147264375
#> [283]   118748638 -1446304172  -647122506  -934837880     6852506  -418540176
#> [289] -1513172567  -404375371   685519365  1640085015 -1519379764 -1611469684
#> [295]  1472255600   999503770   364084283 -1998977541  1159911375  -291783523
#> [301]  -113308710  1171207126   989983614  2070106508 -1560709547   766645889
#> [307]  1766622425   960542715  1450497024  1064261048   788677564   896511230
#> [313]   245434479  -616454001  -562198821  -308647623  -475670602  -259994694
#> [319]  1576315554  -704946920  1397601473 -1664892339  1540682781  -110487153
#> [325]   427056948    79911700 -1028812952  1471235666  1326509411 -1969255101
#> [331]  -176488921 -1480003995   -80138318  -377616002 -1813683690   314043044
#> [337]  1622471101  -332697111  -816194911 -2095646781   134005880  2034649136
#> [343]  1309574692 -1436799770  1887661063  1175376583  1969945731  1358157745
#> [349] -1572759426  -466488078  -877749046  1911394048   190417593  1267361861
#> [355]   -28589483 -1822658009  1619823740  -803434756 -1326551168 -1043204246
#> [361]  2087944299 -1382705941   944734239   898036205 -1726542710   376034310
#> [367]    -8528626  -155550404 -1399705563  1498937521  1725591913   930688203
#> [373]  1654234832   872620808 -1194674004  -311966994  1087338911 -2095460641
#> [379] -2047163381   159445769  1943182022   624727018 -1111180942   441830920
#> [385]   225692337 -1402288451  2064568077  -304219905  -803779644  -682015964
#> [391]  -988310376  1906354306  1737921523   569339731  -599146985  -905293419
#> [397]  1692766722  -443835058  -704292986   653896820  -324164179   819851769
#> [403]  1856353105   611739123  -806247512   703228128    43181812   947019318
#> [409] -2042512809   652616887 -2055916397   216443425   117484206   749085122
#> [415]  -302172294   861152592  1939616137  1030341269    77651045  1631847927
#> [421]  1247230380   -49196372  1262065232 -1618060742   450771675   -35376293
#> [427]   212637807  1665353277 -1183997574 -2006196426   829847518 -1656299860
#> [433]   763821301  1382682593 -1071924231  1882345115   822682656  1168204248
#> [439]  -697987812 -1005451042   -65847793 -1909375761  2138999227   672704537
#> [445] -1320080234  -722412006 -1923446910  -968788040     2095201  1374086893
#> [451]   246471869 -1696533521  -326445804  1944080052   565463432 -1370129666
#> [457] -1621861889 -2121738646  1630688449  1532638238 -1828154623 -1773143332
#> [463]  1082721021 -1953537198  2112402343  -527063146  1854793885 -1718529306
#> [469]  1756225137  1028779640  1929601717  1185503622  -878006161  1362789898
#> [475]  1136563665   935086502  1084552105  -909542756  1591286909   354414122
#> [481] -1780297201  1698837478 -1176842339   883520758   866541809  -150018712
#> [487]    49108701  1765894430 -2131001553 -2014337734   583615969 -1569372370
#> [493]  1527545569  -729382900   196974845 -1323762766  -426515305   152185622
#> [499]   799394013  1443739782  -886094799  2137472680  1491304725  1141630790
#> [505]  -980557105  1143953706  1461920481  -147491082 -2069901879   544745020
#> [511]  -607271827  1171406794  1439660511   328562129 -1755467382   800091316
#> [517]  1572253965  1173136865  1274053924  1965289050  1981679529   578948761
#> [523]  1419012746 -2002532640  1108234681  1589823881 -2138779632   166171842
#> [529] -1578691739   886441161   775707378  1441230012 -1433379243   154873897
#> [535]  1015225636  -595683046  -987438239   437575073   339364546 -1643537248
#> [541]  1315682617 -1646764511 -1420860296  1414716642   508724629  1321472817
#> [547] -2029655542 -1294612364  1104293549   446290577 -1310485612 -1555582326
#> [553]  -946765751 -1483273671   416125466 -2049805920   870967785  -505464119
#> [559] -1707366816   809179154   399078917  -715831175  1197842402  -408207796
#> [565]   310280965  -580739479   -37397740  -567088502 -1589733647   395886449
#> [571] -1151095998     -205472  -928124823   942762353  1146087992   396882898
#> [577]  1628804213 -1332185679  1612944778 -1157808268 -2019048691 -1760274911
#> [583]  -735959356  -338843494   463339337  -476012583 -2036923094  -977437568
#> [589]   218160729  1714035721  1366672880  1278226210 -2112216987  -975186423
#> [595] -1901807918  1313784028  -320267851   466821065  1058075332 -1939276934
#> [601] -1270881567 -1807609407 -1880631390  -763388480  2031634777  -504872735
#> [607]  1982736248  -162966878  -711402667 -1112457391   375859146  -545425164
#> [613]  1016969197   190520369  -915868332  1071877098   611895337  -119955527
#> [619]   287622810  -860544256 -1490228887  -719466231  -404385376  -822132846
#> [625]  2057457283  -358578937
#> attr(,"class")
#> [1] "sim"     "numeric"