Simulation of a Blanchard (1979) bubble process

Simulation of a Blanchard (1979) rational bubble process.

sim_blan(n, pi = 0.7, sigma = 0.03, r = 0.05, b0 = 0.1, seed = NULL)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

pi
A positive value in (0, 1) which governs the probability of the bubble continuing to grow.

sigma
A positive scalar indicating the standard deviation of the innovations.

r
A positive scalar that determines the growth rate of the bubble process.

b0
The initial value of the bubble

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute. Results between seeds in parallell and non-parallel differ.

Value

A numeric vector of length n.

Details

Blanchard's bubble process has two regimes, which occur with probability \(\pi\) and \(1-\pi\). In the first regime, the bubble grows exponentially, whereas in the second regime, the bubble collapses to a white noise.

With probability \(\pi\): $$B_{t+1} = \frac{1+r}{\pi}B_t+\epsilon_{t+1}$$ With probability \(1 - \pi\): $$B_{t+1} = \epsilon_{t+1}$$

where r is a positive constant and \(\epsilon \sim iid(0, \sigma^2)\).

References

Blanchard, O. J. (1979). Speculative bubbles, crashes and rational expectations. Economics letters, 3(4), 387-389.

See also

Examples

sim_blan(n = 100)

#>   [1] 0.1000000000 0.1113835557 0.1361741262 0.0454226795 0.0104071076
#>   [6] 0.0689939076 0.1150897892 0.1450738266 0.1700806452 0.0001070398
#>  [11] 0.0345513522 0.0451913749 0.0305453706 0.0379064771 0.1066159848
#>  [16] 0.0353935274 0.0723214123 0.0895944632 0.1101596456 0.1394247706
#>  [21] 0.1440599951 0.0145797939 0.0195715891 0.0224656491 0.0605952822
#>  [26] 0.1355719361 0.1835757997 0.2914821949 0.4596273843 0.7463305890
#>  [31] 1.0576937618 1.5884769588 2.3747610162 3.5487211883 5.2807607547
#>  [36] 7.9059485655 0.0108647738 0.0409141349 0.0400254783 0.0198653639
#>  [41] 0.0385319525 0.0637366688 0.0206094739 0.0520722211 0.1078515821
#>  [46] 0.0343274691 0.0143238964 0.1011327946 0.0271280736 0.0023894763
#>  [51] 0.0307705532 0.0242324717 0.0306443423 0.0158659408 0.0403052272
#>  [56] 0.0769483710 0.0956362853 0.1451770791 0.1335253034 0.0187892931
#>  [61] 0.0081108400 0.0261905660 0.0217198487 0.0509948812 0.0579562458
#>  [66] 0.0935561158 0.1741719716 0.0544036301 0.0410312000 0.0427237614
#>  [71] 0.0575867676 0.0658687366 0.0854620229 0.0181946942 0.0460175335
#>  [76] 0.0537826821 0.0911433241 0.1444880992 0.1925535919 0.2920003288
#>  [81] 0.4279925028 0.6912441933 1.0175491142 1.5439342882 0.0429309767
#>  [86] 0.0334227429 0.0161365632 0.0147148795 0.0030403192 0.0290389617
#>  [91] 0.0526423138 0.0542125987 0.0544980954 0.0803542968 0.1063960711
#>  [96] 0.1437933278 0.2128359444 0.2443929723 0.3715961350 0.5679089740