Simulation of a Blanchard (1979) bubble process

Simulation of a Blanchard (1979) rational bubble process.

sim_blan(n, pi = 0.7, sigma = 0.03, r = 0.05)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

pi
A positive value in (0, 1) which governs the probability of the bubble continuing to grow.

sigma
A positive scalar indicating the standard deviation of the innovations.

r
A positive scalar that determines the growth rate of the bubble process.

Value

A numeric vector of length n.

Details

Blanchard's bubble process has two regimes, which occur with probability \(\pi\) and \(1-\pi\). In the first regime, the bubble grows exponentially, whereas in the second regime, the bubble collapses to a white noise.

With probability \(\pi\): $$B_{t+1} = \frac{1+r}{\pi}B_t+\epsilon_{t+1}$$ With probability \(1 - \pi\): $$B_{t+1} = \epsilon_{t+1}$$

where r is a positive constant and \(\epsilon \sim iid(0, \sigma^2)\).

References

Blanchard, O. J. (1979). Speculative bubbles, crashes and rational expectations. Economics letters, 3(4), 387-389.

See also

Examples

sim_blan(n = 100)

#>   [1] 1.000000e+00 1.233636e-03 3.285195e-04 2.880365e-02 7.096379e-02
#>   [6] 8.595323e-03 3.572668e-02 6.679685e-02 1.270826e-01 6.378333e-03
#>  [11] 1.706247e-02 1.247242e-02 3.131681e-02 5.024774e-02 9.400409e-02
#>  [16] 1.397903e-01 1.757979e-01 2.799068e-01 4.056755e-02 1.303054e-02
#>  [21] 2.916954e-02 4.361679e-02 4.718931e-02 1.745219e-02 2.938358e-02
#>  [26] 2.538544e-02 5.645753e-02 5.910530e-02 4.826914e-02 4.965121e-02
#>  [31] 8.792417e-02 1.060437e-01 1.625467e-01 2.376726e-01 3.649864e-01
#>  [36] 4.701847e-02 5.241216e-02 1.294804e-01 2.243185e-01 3.217862e-01
#>  [41] 4.501550e-01 6.772718e-01 1.040241e+00 1.584648e+00 2.391877e+00
#>  [46] 2.922528e-02 3.506878e-02 2.795432e-02 3.896927e-02 1.194414e-02
#>  [51] 4.466756e-02 4.513744e-02 3.648963e-02 4.845778e-02 3.331317e-02
#>  [56] 5.830184e-04 1.047069e-02 2.134195e-02 1.910518e-02 3.095918e-02
#>  [61] 5.511641e-03 2.168138e-02 9.059752e-03 2.419259e-02 4.050368e-02
#>  [66] 6.673904e-02 9.893808e-02 1.585558e-01 2.286458e-01 3.154792e-01
#>  [71] 4.655665e-01 6.938547e-01 1.089802e+00 1.635053e-02 8.363830e-03
#>  [76] 5.534854e-02 1.030268e-01 1.470251e-01 2.246529e-01 3.213477e-01
#>  [81] 3.534393e-02 8.845054e-02 2.130473e-02 3.019785e-02 1.582523e-02
#>  [86] 4.626354e-02 7.705294e-03 3.182737e-03 2.321595e-02 2.445848e-02
#>  [91] 3.935487e-02 1.311286e-01 2.044910e-01 5.925895e-02 5.410774e-02
#>  [96] 7.200895e-06 8.739690e-03 5.348998e-02 5.133719e-02 1.124780e-01