Simulation of a Blanchard (1979) bubble process

Simulation of a Blanchard (1979) rational bubble process.

sim_blan(n, pi = 0.7, sigma = 0.03, r = 0.05)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

pi
A positive value in (0, 1) which governs the probability of the bubble continuing to grow.

sigma
A positive scalar indicating the standard deviation of the innovations.

r
A positive scalar that determines the growth rate of the bubble process.

Value

A numeric vector of length n.

Details

Blanchard's bubble process has two regimes, which occur with probability \(\pi\) and \(1-\pi\). In the first regime, the bubble grows exponentially, whereas in the second regime, the bubble collapses to a white noise.

With probability \(\pi\): $$B_{t+1} = \frac{1+r}{\pi}B_t+\epsilon_{t+1}$$ With probability \(1 - \pi\): $$B_{t+1} = \epsilon_{t+1}$$

where r is a positive constant and \(\epsilon \sim iid(0, \sigma^2)\).

References

Blanchard, O. J. (1979). Speculative bubbles, crashes and rational expectations. Economics letters, 3(4), 387-389.

See also

Examples

sim_blan(n = 100)

#>   [1] 1.0000000000 0.0156360047 0.0233334960 0.0273224496 0.0694462930
#>   [6] 0.0106252757 0.0159079180 0.0145326290 0.0616181592 0.0137513747
#>  [11] 0.0192285085 0.0084141246 0.0021980239 0.0371451395 0.0772688494
#>  [16] 0.1320872689 0.1662938836 0.0049491083 0.0428862951 0.0700058289
#>  [21] 0.0074684734 0.0426020125 0.0252239906 0.0112534671 0.0001726489
#>  [26] 0.0093317723 0.0206656524 0.0021334698 0.0015609711 0.0293426681
#>  [31] 0.0321669757 0.0768748434 0.0414182801 0.1123777480 0.0353071990
#>  [36] 0.0484938484 0.0674060717 0.1252027633 0.0363746876 0.0372108010
#>  [41] 0.0205702764 0.0092874153 0.0214270527 0.0501219721 0.0754423919
#>  [46] 0.0027214983 0.0046149838 0.0476536250 0.0364518353 0.0982951405
#>  [51] 0.1448653513 0.1987709644 0.0076319141 0.0203239801 0.0321896640
#>  [56] 0.0452048503 0.1263649663 0.0263569787 0.0003369602 0.0027103282
#>  [61] 0.0270707412 0.0164826397 0.0235475202 0.0197677298 0.0022961061
#>  [66] 0.0079535684 0.0251410161 0.0773049859 0.1233543126 0.0282547980
#>  [71] 0.0321176570 0.0398920812 0.0115097034 0.0606718618 0.1430248015
#>  [76] 0.2282300835 0.3635702299 0.0619136767 0.0935878327 0.0073784838
#>  [81] 0.0192295001 0.0666853549 0.1046931493 0.1309662344 0.1612146682
#>  [86] 0.1890316831 0.2852983722 0.4628831542 0.7044535680 1.0250447904
#>  [91] 1.5573902588 2.3113831181 3.4801856869 0.0111716433 0.0281824558
#>  [96] 0.0835448441 0.1500927280 0.2126556419 0.3484479504 0.5299373475