Simulation of a single-bubble process

The following function generates a time series which switches from a martingale to a mildly explosive process and then back to a martingale.

sim_dgp1(n, te = 0.4 * n, tf = 0.15 * n + te, c = 1, alpha = 0.6,
  sigma = 6.79)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te
A scalar in (0, tf) specifying the observation in which the bubble originates.

tf
A scalar in (te, n) specifying the observation in which the bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

Value

A numeric vector of length n.

Details

The data generating process is described by the following equation: $$X_t = X_{t-1}1\{t < \tau_e\}+ \delta_T X_{t-1}1\{\tau_e \leq t\leq \tau_f\} + \left(\sum_{k=\tau_f+1}^t \epsilon_k + X^*_{\tau_f}\right) 1\{t > \tau_f\} + \epsilon_t 1\{t \leq \tau_f\}$$

where the autoregressive coefficient \(\delta_T\) is given by:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\) and \(X_{\tau_f} = X_{\tau_e} + X^*\) \(\tau\) is the last observation of the sample. During the pre- and post- bubble periods, \(N_0 = [1, \tau_e)\), X is a pure random walk process. During the bubble expansion period \(B = [\tau_e, \tau_f]\) is a mildly explosive process with expansion rate given by the autoregressive coefficient \(\delta_T\), and continues its martingale path of the subsequent period \(N_1 = (\tau_f, \tau]\).

For further details the user can refer to Phillips et al., (2015) p. 1054.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination date 40 and termination date 55
sim_dgp1(n = 100)

#>   [1] 100.00000  94.99734 105.90203 110.47475 103.73450 110.70539 113.51777
#>   [8] 121.14121 121.10282 129.30579 127.87792 120.00670 124.65248 132.68720
#>  [15] 143.65999 134.69017 123.33733 127.03725 131.36030 140.53685 135.90925
#>  [22] 139.87781 138.42548 146.92987 137.88692 133.35262 136.48425 144.27919
#>  [29] 149.60110 141.73285 134.57071 137.67383 131.32177 125.39012 124.77441
#>  [36] 119.74958 114.61436 125.73584 128.20602 117.55094 128.49622 135.84731
#>  [43] 140.93541 147.40915 150.38104 168.39698 183.29393 193.49101 200.32602
#>  [50] 208.93433 214.97083 225.05870 245.75837 276.24528 289.20258 117.55094
#>  [57] 109.31654 111.20034 104.45704 101.87964  93.03370  99.23914  93.82483
#>  [64]  98.65154  99.50475 107.16845 114.00342 104.74326 106.49873 105.76338
#>  [71] 105.01636  99.87741  96.54751  95.71190  88.59965  93.33545  94.44575
#>  [78] 102.84788 106.80982 108.74364 112.01602 120.03355 125.85306 116.62910
#>  [85] 111.83721 113.88779 122.28759 119.99714 124.44791 121.11763 120.60533
#>  [92] 121.21868 138.81256 133.99065 131.07804 130.12160 129.37776 126.08485
#>  [99] 133.74100 136.42905

# 200 periods with bubble origination date 80 and termination date 110
sim_dgp1(n = 200)

#>   [1] 100.00000  94.89322  85.44842  85.44252  88.50301  94.40307  97.49173
#>   [8]  89.41275  94.88582  94.69167  92.77838  88.03184  78.55892  86.53856
#>  [15]  95.13256  80.87426  77.71188  80.93029  83.93801  92.11102  85.22332
#>  [22]  87.31963  78.71667  81.00193  91.58931  88.21429  86.43357  87.98994
#>  [29]  83.57193  85.43562  87.67604  98.25862  98.55241 100.76164 101.90211
#>  [36] 104.42047 103.77973 104.91526  98.06737  98.74061  99.58084 108.09092
#>  [43] 109.43421 112.38850 106.12437 102.82811  99.36153 104.40647  99.91467
#>  [50] 104.30083  95.98026 101.48177 114.15416 119.04550 113.12420 106.44620
#>  [57] 116.62815 106.90142 110.06217 122.48875 121.12085 106.01010 105.10711
#>  [64] 112.69223 120.73936 124.58882 132.90052 122.73579 133.21003 138.75618
#>  [71] 138.26107 133.81552 130.51404 134.11827 124.59398 124.45351 129.84434
#>  [78] 134.96402 130.53250 132.78775 137.25709 132.38204 128.35522 129.57596
#>  [85] 136.95950 150.41159 160.22037 171.07293 187.10448 196.58743 210.60953
#>  [92] 217.02165 236.15047 250.99327 259.14877 275.02773 289.71535 300.85358
#>  [99] 321.18455 323.80506 357.33821 374.86489 400.26468 414.70811 432.95670
#> [106] 449.18618 471.16375 480.50539 502.35395 518.84616 132.78775 129.05236
#> [113] 139.45986 123.40215 121.55558 116.55277 126.45208 112.60973 106.75205
#> [120]  99.74012  98.50004  90.09284  82.40652  68.05942  74.51407  73.11299
#> [127]  74.48147  74.69178  84.67925  84.09683  82.41986  84.36169  94.21682
#> [134]  85.40148  72.66573  77.76175  76.91555  80.15592  75.64273  65.95385
#> [141]  69.58912  52.35206  58.77638  68.82407  59.89373  59.72469  69.43726
#> [148]  68.48634  50.53230  51.49131  46.18834  43.38619  36.98579  41.21018
#> [155]  36.97336  35.57358  39.15124  38.67816  39.12782  36.85520  29.23500
#> [162]  30.88441  28.13687  45.51073  39.81604  46.54872  56.77654  51.29811
#> [169]  55.97530  63.34461  64.33545  59.96999  67.04509  50.74524  39.89617
#> [176]  31.52259  32.39122  35.75068  35.92021  58.14879  62.24245  58.66207
#> [183]  57.40029  61.19866  71.62353  65.20066  63.90114  64.96199  68.75729
#> [190]  80.42467  78.78158  78.20577  80.21106  83.79696  80.37915  76.48866
#> [197]  84.40159  92.98445  78.94942  79.62335

# 200 periods with bubble origination date 100 and termination date 150
sim_dgp1(n = 200, te = 100, tf = 150)

#>   [1] 100.00000 101.78669 101.32827 111.39870 109.19541 102.01888 104.86286
#>   [8] 107.39672 115.21758 119.68592 113.78058 119.03571 128.43713 133.51662
#>  [15] 142.00045 145.75720 150.81348 148.16409 145.78642 135.71796 130.21718
#>  [22] 119.31353 122.92429 118.17236 104.84894 105.97328 101.22830  95.89838
#>  [29]  97.54062  91.34214 100.20546  96.67637 106.49608 101.94967  97.39541
#>  [36]  91.02839  87.07122  86.23330  87.99424  93.87531 106.63722 121.19992
#>  [43] 122.58864 122.81297 132.82124 132.75797 120.95411 127.15932 127.36655
#>  [50] 127.65005 117.67496 117.03142 110.10338 116.49579 114.32568 118.88601
#>  [57] 119.41797 118.90615 112.74838 110.74610 118.08287 113.36723 100.71127
#>  [64] 102.68368  85.60214  79.61144  82.26145  87.62334  89.55138  84.64692
#>  [71]  85.23469  78.43138  75.14229  82.00486  86.31191  85.62930  85.73508
#>  [78]  84.96116  90.67265  88.05029  86.87349  92.05738  94.29914  94.07094
#>  [85]  90.65391  94.48996  97.05920  98.44727  93.88023  93.48673  86.50550
#>  [92]  92.82618  92.52369  92.29096  94.83673  99.87093  98.87730  94.57768
#>  [99]  86.39403  98.86266  98.11070 102.83475 111.16985 112.87891 115.15850
#> [106] 114.92285 124.52141 132.78460 138.70481 140.56593 154.52316 177.26000
#> [113] 173.28366 178.32678 189.06563 204.03899 219.90361 217.37478 230.40286
#> [120] 235.59387 250.47273 254.85351 281.53726 307.64912 322.45235 338.92731
#> [127] 353.50797 365.49106 374.02231 380.40549 399.46988 417.14194 429.28723
#> [134] 452.22459 469.87006 482.88341 512.39086 528.60127 555.08265 575.75781
#> [141] 601.42674 633.39508 668.07723 710.21955 747.08306 777.68476 830.68615
#> [148] 867.15004 901.60828 942.35605  98.86266  94.91945  83.46405  77.87589
#> [155]  69.04265  73.80421  62.81781  68.08009  66.63007  80.25794  68.12718
#> [162]  63.38087  58.36410  50.57075  42.55330  37.36998  43.66983  34.64557
#> [169]  37.55553  30.74118  32.65967  29.45783  35.04337  31.90596  35.29779
#> [176]  31.19570  37.36769  43.64078  36.50741  37.70436  41.95929  40.00751
#> [183]  46.26797  44.68433  50.12147  50.96683  43.79541  37.19343  37.83579
#> [190]  48.12207  49.81265  48.52586  43.43775  50.88124  53.36823  48.59354
#> [197]  48.44252  50.73886  49.42596  45.94219