Simulation of a single-bubble process

The following function generates a time series which switches from a martingale to a mildly explosive process and then back to a martingale.

sim_dgp1(n, te = 0.4 * n, tf = 0.15 * n + te, c = 1, alpha = 0.6,
  sigma = 6.79)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te
A scalar in (0, tf) specifying the observation in which the bubble originates.

tf
A scalar in (te, n) specifying the observation in which the bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

Value

A numeric vector of length n.

Details

The data generating process is described by the following equation: $$X_t = X_{t-1}1\{t < \tau_e\}+ \delta_T X_{t-1}1\{\tau_e \leq t\leq \tau_f\} + \left(\sum_{k=\tau_f+1}^t \epsilon_k + X^*_{\tau_f}\right) 1\{t > \tau_f\} + \epsilon_t 1\{t \leq \tau_f\}$$

where the autoregressive coefficient \(\delta_T\) is given by:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\) and \(X_{\tau_f} = X_{\tau_e} + X^*\) \(\tau\) is the last observation of the sample. During the pre- and post- bubble periods, \(N_0 = [1, \tau_e)\), X is a pure random walk process. During the bubble expansion period \(B = [\tau_e, \tau_f]\) is a mildly explosive process with expansion rate given by the autoregressive coefficient \(\delta_T\), and continues its martingale path of the subsequent period \(N_1 = (\tau_f, \tau]\).

For further details the user can refer to Phillips et al., (2015) p. 1054.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination date 40 and termination date 55
sim_dgp1(n = 100)

#>   [1] 100.00000  98.34487  93.32586  91.87160  93.96129 101.79869  99.46256
#>   [8]  90.60657  90.65291  80.73509  75.66380  82.18663  83.33216  81.96362
#>  [15]  81.26315  72.27132  71.15263  70.18445  61.39151  76.18125  72.24149
#>  [22]  56.28239  59.73763  68.98917  62.39672  50.94769  50.73080  57.98699
#>  [29]  56.53239  56.91474  65.27046  69.09428  70.61422  68.70567  55.56158
#>  [36]  57.92914  58.73565  64.09230  69.11701  74.78858  93.47150  99.95380
#>  [43] 102.54767 100.35381 118.40546 127.93822 130.29604 143.04430 141.89756
#>  [50] 159.13570 173.49460 189.43652 202.25789 206.58399 213.08068  74.78858
#>  [57]  73.89378  70.15985  75.49786  89.65499  98.04751  98.78120  95.82525
#>  [64]  80.40414  86.91823  81.25308  72.07097  63.50121  59.28261  59.36191
#>  [71]  62.87818  55.23253  66.83517  64.95335  59.67234  62.42951  61.14759
#>  [78]  70.55609  74.31178  73.38280  70.11455  64.85583  59.56540  59.66038
#>  [85]  65.82641  66.98535  76.70843  84.56562  70.30509  70.43376  67.82623
#>  [92]  74.03639  78.45105  80.09946  85.88450  84.23247  86.88849  80.01908
#>  [99]  89.50563  85.30049

# 200 periods with bubble origination date 80 and termination date 110
sim_dgp1(n = 200)

#>   [1] 100.00000000  95.82184276  83.90103658  84.57668595  82.92547877
#>   [6]  88.19582163  88.76670077  85.89196577  77.76051375  90.15416686
#>  [11]  94.82113415  91.50630746  95.24153607  87.53380529  77.41690053
#>  [16]  72.52357705  63.26781516  58.86722409  59.68894671  68.40609962
#>  [21]  61.40872650  66.63310555  52.29138052  53.66441810  55.18847624
#>  [26]  56.61259434  54.48038339  50.63238079  43.69494789  27.56761393
#>  [31]  34.54867681  35.24751180  34.03760071  34.66815289  38.19416227
#>  [36]  33.53469556  51.78887802  41.07142315  44.39370982  47.79854819
#>  [41]  51.92165504  49.15177597  48.13011320  49.38718973  52.61663862
#>  [46]  53.19282677  56.97528481  46.67481836  51.78398038  56.08306386
#>  [51]  53.96343620  45.83749158  41.39612799  39.90120658  33.48963403
#>  [56]  28.47600009  26.80297922  32.93986406  32.61655457  26.43248019
#>  [61]  37.96800081  40.37214799  41.12389605  37.66635242  42.96483445
#>  [66]  45.22022254  53.85979684  58.42446579  51.27632712  47.16914635
#>  [71]  41.69700369  38.26427289  28.65563884  29.20866814  29.08035185
#>  [76]  29.92213497  24.76578122  22.78753346  37.71305686  26.81498356
#>  [81]  31.28794090  35.25754186  36.37384120  38.78456461  34.89228602
#>  [86]  41.95081688  54.39137929  59.82933739  56.43532600  53.66401664
#>  [91]  60.05615953  64.78489330  63.01512370  63.43921379  76.57319042
#>  [96]  78.02491557  83.18058395  77.05322242  78.22386808  83.46572586
#> [101]  83.34627961  79.64030061  83.60941394  95.57421136  95.11799919
#> [106]  92.17174120 100.19960561 107.48343401 107.71262275 115.92550903
#> [111]  26.81498356  15.19015378   3.45283975   1.84558903  -1.45997037
#> [116]  -2.67077094   0.33129105  -4.88223250   6.83059911  -0.49216328
#> [121]   2.82191538  -2.60427371  -3.83871659  -3.27189711  -0.48687091
#> [126]  -8.59435715   8.64896073  12.87222235  22.96054566  27.31895275
#> [131]  27.42676060  15.13287840  -1.59783582  -2.44488085  -7.63918900
#> [136]  -5.76522615  -8.92340119  -5.69374610  -3.70588859 -17.66669760
#> [141] -14.00706922 -13.95219418 -18.46770581  -7.61054169  -7.37005301
#> [146]  -2.52870970   6.07026822   7.71330027  12.33693722  12.74384690
#> [151]  -0.25627606   5.91647497   0.07032775   0.17780362   2.45843841
#> [156]   3.48622188  -0.07638166  11.08874215  19.56566211  22.91519124
#> [161]  22.76187343  26.50318409  27.74924438  37.85433568  40.76748776
#> [166]  37.05935324  39.42443645  41.40988762  45.31990224  51.10171060
#> [171]  51.90201876  41.34207298  41.18347736  47.73994055  49.76719356
#> [176]  60.55008542  58.89653358  64.43575783  61.51964480  58.52307730
#> [181]  65.57555628  73.01033520  81.09105652  79.05368442  65.44044513
#> [186]  73.27006877  76.55616938  91.37185466  88.78758877 107.10240104
#> [191] 109.32215490 105.48820430 111.76970835 114.81757158  95.25294236
#> [196] 101.47655314 101.90093179  95.41795648  89.07857233  95.29598040

# 200 periods with bubble origination date 100 and termination date 150
sim_dgp1(n = 200, te = 100, tf = 150)

#>   [1]  100.00000   96.97234   82.30136   85.31686   98.62428  102.87080
#>   [7]  104.04728  101.77596  100.52895   97.55815   91.83557   94.70697
#>  [13]  103.55736  104.99687  117.64103  128.49272  129.53694  125.31637
#>  [19]  120.14304  119.34314  108.93900  110.47477  111.02363  102.91680
#>  [25]  103.01034  102.71239   97.57186   95.12906   92.91823  102.25571
#>  [31]  107.35241  103.45552  104.95740   94.60767   93.11397   96.47611
#>  [37]   90.17034   87.02412   88.13365   79.83463   78.93305   79.55673
#>  [43]   77.09751   75.81511   70.52007   69.77369   62.55121   57.52477
#>  [49]   58.35411   68.20562   57.92742   75.78845   77.00200   90.57218
#>  [55]   95.89388   95.77761  104.52178  106.23951   93.86445   91.21182
#>  [61]   92.20252   88.00736   82.46449   87.96059   85.72636   94.63914
#>  [67]   99.80723  101.74096  106.21261  103.56069  100.02613  100.49098
#>  [73]  102.90271  106.29593  113.93620  115.29515  106.03950  120.65117
#>  [79]  128.93862  136.53410  136.45099  131.99191  128.30065  124.51397
#>  [85]  125.93155  136.00970  132.36419  131.07013  132.47487  139.53204
#>  [91]  142.69592  161.38275  151.21723  148.97119  153.76972  157.77406
#>  [97]  155.84729  166.58122  168.76573  172.25668  188.28446  201.97977
#> [103]  203.92294  206.19128  219.29926  222.39775  239.71735  259.94951
#> [109]  269.98868  279.71256  299.07190  306.84889  313.72738  323.96080
#> [115]  333.43995  342.32673  349.87396  352.55073  361.18163  370.60291
#> [121]  389.46427  398.89650  409.40309  414.95937  428.99578  442.01568
#> [127]  468.49282  491.33938  504.92316  530.58269  548.17260  574.97329
#> [133]  604.67565  635.44639  675.14044  714.34230  757.72289  793.38519
#> [139]  825.41575  857.53279  903.02411  953.05847  989.54769 1025.62748
#> [145] 1072.78578 1113.37112 1166.47028 1203.49528 1250.00947 1309.91966
#> [151]  172.25668  164.34893  175.79341  167.23419  164.34508  169.49601
#> [157]  159.87081  165.04876  163.87562  161.24232  159.21570  163.20587
#> [163]  163.13163  169.41304  184.01002  209.41996  204.84420  207.03574
#> [169]  216.81203  202.10032  195.01714  201.26300  203.18894  196.15559
#> [175]  197.52361  199.72098  197.41387  188.72956  195.69483  209.56258
#> [181]  211.71971  212.52802  210.29679  216.76718  214.12004  216.45192
#> [187]  209.73988  218.45781  224.97224  230.64335  228.50797  216.03613
#> [193]  218.57740  218.69009  213.01918  212.93709  208.68268  205.34874
#> [199]  210.58306  194.95557