Simulation of a single-bubble process

The following function generates a time series which switches from a martingale to a mildly explosive process and then back to a martingale.

sim_dgp1(n, te = 0.4 * n, tf = 0.15 * n + te, c = 1, alpha = 0.6,
  sigma = 6.79)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te
A scalar in (0, tf) specifying the observation in which the bubble originates.

tf
A scalar in (te, n) specifying the observation in which the bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

Value

A numeric vector of length n.

Details

The data generating process is described by the following equation: $$X_t = X_{t-1}1\{t < \tau_e\}+ \delta_T X_{t-1}1\{\tau_e \leq t\leq \tau_f\} + \left(\sum_{k=\tau_f+1}^t \epsilon_k + X^*_{\tau_f}\right) 1\{t > \tau_f\} + \epsilon_t 1\{t \leq \tau_f\}$$

where the autoregressive coefficient \(\delta_T\) is given by:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\) and \(X_{\tau_f} = X_{\tau_e} + X^*\). During the pre- and post- bubble periods, \(N_0 = [1, \tau_e)\), X is a pure random walk process. During the bubble expansion period \(B = [\tau_e, \tau_f]\) is a mildly explosive process with expansion rate given by the autoregressive coefficient \(\delta_T\), and continues its martingale path for the subsequent period \(N_1 = (\tau_f, \tau]\).

For further details the user can refer to Phillips et al. (2015) p. 1054.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination date 40 and termination date 55
sim_dgp1(n = 100)

#>   [1] 100.00000  93.63148 103.85626  96.61913  97.72094  99.48954  89.38465
#>   [8]  98.99374 102.81317 107.17273 101.12601 100.33472  93.93871 101.51519
#>  [15] 104.70740 110.55096 110.07297 105.90948 103.62318 102.15763 106.37512
#>  [22]  97.65658  88.82895  86.27068  86.97514  82.19795  92.36539  90.30919
#>  [29]  80.96051  86.95724  91.51618  99.86803 111.04076 109.79869 100.43353
#>  [36] 108.92511  96.18546  93.68139  79.22346  86.51474  81.28209  85.07763
#>  [43]  96.65893 110.08164 126.89185 123.67241 112.72253 121.62990 136.26107
#>  [50] 140.38270 146.61724 154.22159 165.45015 165.83274 181.21673  70.37142
#>  [57]  69.92729  66.73799  64.83329  68.73643  67.33988  68.93472  69.68589
#>  [64]  71.54772  82.24142  86.09657  85.69180  84.67610  86.48339  90.07124
#>  [71]  86.13720  92.53473  87.73854  79.07856  84.05739  81.20419  79.47019
#>  [78]  76.96306  85.19559  82.17157  89.90313  91.19678  81.91897  73.10485
#>  [85]  58.59395  57.94072  62.65643  59.43396  71.40562  85.16111  88.06464
#>  [92]  84.21825  78.55651  81.26069  71.24857  68.56353  67.67606  54.09949
#>  [99]  47.33712  50.18970

# 200 periods with bubble origination date 80 and termination date 110
sim_dgp1(n = 200)

#>   [1]  1.000000e+02  9.957317e+01  9.902907e+01  9.880987e+01  9.392799e+01
#>   [6]  8.634654e+01  8.104851e+01  6.898296e+01  6.607796e+01  5.228728e+01
#>  [11]  7.096498e+01  8.124279e+01  8.147358e+01  8.650356e+01  7.787478e+01
#>  [16]  7.676417e+01  7.261585e+01  8.111615e+01  8.281582e+01  7.123492e+01
#>  [21]  6.542666e+01  6.444278e+01  6.223979e+01  6.106807e+01  5.267520e+01
#>  [26]  3.975856e+01  3.911687e+01  3.933793e+01  4.247009e+01  5.184980e+01
#>  [31]  4.902192e+01  5.364552e+01  5.083193e+01  4.731236e+01  4.266787e+01
#>  [36]  3.665432e+01  3.698864e+01  3.824860e+01  3.411581e+01  2.915162e+01
#>  [41]  4.758745e+01  3.849301e+01  3.410657e+01  2.777520e+01  2.255310e+01
#>  [46]  2.507613e+01  2.748952e+01  2.080817e+01  2.226618e+01  2.172240e+01
#>  [51]  1.206451e+01  1.958885e+01  2.691171e+01  2.392241e+01  1.863864e+01
#>  [56]  6.290416e+00 -1.342093e+00  5.858864e+00 -4.181511e+00 -1.471660e+01
#>  [61] -9.437334e+00 -2.182625e+00 -3.429186e+00  7.151284e+00  5.703368e+00
#>  [66]  1.202170e+01  1.481111e+01  6.120972e+00  8.087024e-01 -1.464590e+01
#>  [71] -1.550413e+01 -5.671216e+00 -1.546754e+01 -5.505338e+00 -1.055234e+01
#>  [76] -1.261802e+01 -1.032533e+01 -1.445027e+01 -1.645712e+01 -1.805570e+01
#>  [81] -1.327493e+01 -1.568067e+01 -1.670512e+00  5.672846e+00  1.095658e+01
#>  [86]  3.211308e+00  5.570872e+00  2.175793e+00  1.098436e+01  1.163806e+01
#>  [91]  9.435880e+00  5.110518e+00 -4.807971e+00  4.798391e+00  2.641888e+00
#>  [96]  2.478504e+00  1.104631e+01  2.364755e+00 -1.433595e+00 -6.286736e+00
#> [101] -1.916021e+01 -2.163474e+01 -2.481380e+01 -2.754001e+01 -2.556628e+01
#> [106] -2.975446e+01 -3.059291e+01 -3.690862e+01 -5.313570e+01 -5.692898e+01
#> [111] -1.211508e+01 -1.872510e+01 -2.220666e+01 -2.847714e+01 -3.862429e+01
#> [116] -3.975096e+01 -2.900756e+01 -2.211841e+01 -2.990300e+01 -3.608087e+01
#> [121] -1.496952e+01 -2.223179e+01 -2.698553e+01 -2.833297e+01 -2.328379e+01
#> [126] -2.362437e+01 -1.470813e+01 -1.412410e+01 -1.840079e+01 -1.395286e+01
#> [131] -1.382118e+01 -9.093504e+00  4.546723e+00  4.159841e+00  5.792262e+00
#> [136]  5.643628e+00  2.440188e-03  9.525241e+00  4.610170e+00  1.364481e+01
#> [141]  8.135859e+00  2.032415e+01  2.498635e+01  2.559322e+01  2.779690e+01
#> [146]  2.828115e+01  2.986355e+01  4.335105e+01  3.688920e+01  3.309602e+01
#> [151]  3.483300e+01  3.348807e+01  3.639954e+01  3.688007e+01  2.644619e+01
#> [156]  2.952875e+01  3.546591e+01  2.665428e+01  1.932810e+01  1.162247e+01
#> [161]  1.177602e+01  1.163030e+01  9.737435e+00  9.573074e+00  1.517496e+01
#> [166]  1.008122e+01  1.372209e+01  3.043338e+00 -3.654504e+00  9.813544e+00
#> [171] -2.757315e+00 -8.934777e+00 -2.218505e+01 -2.761893e+01 -4.031190e+01
#> [176] -4.540828e+01 -4.942075e+01 -5.445960e+01 -4.975066e+01 -5.015442e+01
#> [181] -6.281026e+01 -7.146418e+01 -8.356240e+01 -8.701575e+01 -9.880916e+01
#> [186] -9.853429e+01 -9.937747e+01 -1.035381e+02 -1.024472e+02 -1.069446e+02
#> [191] -1.092182e+02 -1.049880e+02 -9.800654e+01 -1.057103e+02 -9.948450e+01
#> [196] -9.124229e+01 -9.589977e+01 -1.033319e+02 -1.008544e+02 -1.076988e+02

# 200 periods with bubble origination date 100 and termination date 150
sim_dgp1(n = 200, te = 100, tf = 150)

#>   [1]  100.00000  103.76843  106.72931  107.44480  105.72192   99.09154
#>   [7]  101.76828  108.65947  112.94698  109.66470   99.30089  103.48925
#>  [13]  111.83611  112.90522  122.45428  125.14856  127.36497  117.45603
#>  [19]  112.02881  115.22020  107.41047  105.56633  108.67224  108.55710
#>  [25]  104.87967  110.82645  115.85674  114.73766  109.63861  101.12826
#>  [31]   94.09221   93.89232   92.02546   95.32167  101.54798   94.85749
#>  [37]   84.73232   90.29735   97.17836  104.33234  103.84734  110.31444
#>  [43]  107.70683  104.87024  101.67712   90.20079   83.78086   81.45181
#>  [49]   83.88577   84.21297   91.88684   88.30422   91.56054  101.18556
#>  [55]  100.64175   97.96522  101.53171   96.12732   97.79152   92.49994
#>  [61]   96.53947  104.06884   97.66608  102.42142   99.53787   91.77597
#>  [67]   93.36103   92.07611   80.78155   93.78909   88.24238   90.84736
#>  [73]   87.73325   82.87362   85.79026   77.84696   86.82394   96.09884
#>  [79]   94.40321   91.51935   98.97776  103.53332  103.74351  119.82570
#>  [85]  137.38320  145.61459  137.11896  139.83677  146.18344  142.97401
#>  [91]  153.72267  157.78142  160.94613  155.89188  162.40386  163.44027
#>  [97]  158.21626  167.67527  173.91303  188.05186  199.51019  201.10934
#> [103]  217.54616  223.08798  230.61220  237.98449  248.38983  257.04673
#> [109]  258.41846  268.90119  282.95716  296.10485  296.90707  313.62775
#> [115]  319.92566  339.78771  342.68779  364.22423  378.64574  407.30335
#> [121]  416.53800  431.97820  443.88931  463.29559  477.49467  500.89022
#> [127]  520.43514  542.29526  567.30457  590.72326  623.10210  651.57705
#> [133]  680.89615  703.60579  742.36366  771.96594  805.94984  832.15518
#> [139]  850.97923  882.67154  918.92228  964.18410 1005.78162 1044.29338
#> [145] 1098.03229 1139.80625 1198.62554 1241.72486 1301.71712 1363.21866
#> [151]  183.89675  187.33838  190.46224  200.54126  206.52979  202.88558
#> [157]  211.61449  215.60598  206.72139  208.87196  216.98030  223.17996
#> [163]  217.83790  215.04769  218.28227  219.54464  210.54981  218.51021
#> [169]  220.08205  223.36211  219.72729  229.08483  220.23969  224.55037
#> [175]  231.33803  229.04471  228.46055  216.78977  210.48104  216.00590
#> [181]  219.58025  226.45144  232.09651  234.91527  230.42533  234.15176
#> [187]  230.48945  229.32620  224.12007  220.00143  231.18668  236.89754
#> [193]  239.19375  236.44460  242.55566  250.64029  250.31750  255.66072
#> [199]  263.58250  261.73682