Simulation of a two-bubble process

The following data generating process is similar to sim_dgp1, with the difference that there are two episodes of midly explosive dynamics.

sim_dgp2(n, te1 = 0.2 * n, tf1 = 0.2 * n + te1, te2 = 0.6 * n,
  tf2 = 0.1 * n + te2, c = 1, alpha = 0.6, sigma = 6.79)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te1
A scalar in (0, n) specifying the observation in which the first bubble originates.

tf1
A scalar in (te1, n) specifying the observation in which the first bubble collapses.

te2
A scalar in (tf1, n) specifying the observation in which the second bubble originates.

tf2
A scalar in (te2, n) specifying the observation in which the second bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

Value

A numeric vector of length n.

Details

The data generating process is described by:

$$X_t = X_{t-1}1\{t \in N_0\}+ \delta_T X_{t-1}1\{t \in B_1 \cup B_2\} + \left(\sum_{k=\tau_{1f}+1}^t \epsilon_k + X^*_{\tau_{1f}}\right) 1\{t \in N_1\} $$

$$ + \left(\sum_{l=\tau_{2f}+1}^t \epsilon_l + X^*_{\tau_{2f}}\right) 1\{t \in N_2\} + \epsilon_t 1\{t \in N_0 \cup B_1 \cup B_2\}$$

where the autoregressive coefficient \(\delta_T\) is given:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\), \(X_{\tau_{1f}} = X_{\tau_{1e}} + X^*\) and \(X_{\tau_{2f}} = X_{\tau_{2e}} + X^*\). We use the notation \(N_0 = [1, \tau_{1e})\), \(B_1 = [\tau_{1e}, \tau_{1f}]\), \(N_1 = (\tau_{1f}, \tau_{2e})\), \(B_2 = [\tau_{2e}, \tau_{2f}]\), \(N_2 = (\tau_{2f}, \tau]\), where \(\tau\) is the last observation of the sample. After the collapse of the first bubble \(X_t\) resumes a martingale path until the time \(\tau_{2e}-1\), and a second episode of exuberance begins at \(\tau_{2e}\). The expansion process lasts until \(\tau_{2f}\) and collapses to a value of \(X^*_{\tau_{2f}}\). The process then continues on a martingale path until the end of the sample period \(\tau\). The expansion duration of the first bubble is assumed to be longer than that of the second bubble, i.e. \(\tau_{1f}-\tau_{1e}>\tau_{2f}-\tau_{2e}\).

For further details the user can refer to Phillips et al., (2015) p. 1055.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination dates 20/60 and termination dates 40/70 respectively
sim_dgp2(n = 100)

#>   [1] 100.00000  98.75017 101.18212  98.41221 102.32244 105.44443 115.89018
#>   [8] 110.34708  99.98846 102.69378  88.60604  89.27519  95.17854 107.97552
#>  [15] 103.90130 112.59371 105.33042 108.66455 114.05606 113.43710 122.75092
#>  [22] 120.01632 140.77649 145.97185 152.50475 160.08246 171.43279 181.60914
#>  [29] 186.14951 188.31780 194.52655 213.79382 225.36985 236.07428 243.49136
#>  [36] 251.72185 251.94973 272.34122 287.56575 301.37128 113.43710 111.26329
#>  [43] 108.93004 111.08634 108.89573  95.72134  97.12106  88.11914  94.21353
#>  [50]  96.37458  94.44275  91.86625  99.85397 104.75222 103.40482 101.91499
#>  [57] 108.17906 104.72701 105.49821  92.51974  93.33772  88.49283  95.17549
#>  [64]  91.48515 102.31231 106.59041 116.56927 125.34324 138.44784 149.19976
#>  [71]  92.51974 101.61295 102.07771 110.96414 108.22728 101.27384  98.91253
#>  [78]  98.16684 106.54238 110.65843 107.74366 110.99823 108.65990 108.83020
#>  [85] 119.13083 126.25946 129.72987 133.11322 138.44440 148.27644 148.07098
#>  [92] 153.48493 152.65304 144.03029 137.26692 148.56088 147.89171 136.32354
#>  [99] 129.69787 138.47941

# 200 periods with bubble origination dates 40/120 and termination dates 80/140 respectively
sim_dgp2(n = 200)

#>   [1] 100.00000 102.31283 104.43880 106.87575 106.09545 109.55281 105.89113
#>   [8] 114.44487 109.92524 109.28904 115.35822 106.20372 104.02803 115.12757
#>  [15] 120.36223 116.86557 115.92166 111.79044 109.84655 114.43743 127.63391
#>  [22] 127.68428 131.28488 137.85385 137.02538 126.64224 136.81330 138.87600
#>  [29] 149.81777 146.65076 148.35688 152.48253 142.84391 129.67736 137.99963
#>  [36] 132.54366 120.95770 118.52017 113.30609 108.44282 102.82132 116.92509
#>  [43] 119.57538 125.29430 134.66602 136.58927 144.94024 155.70920 166.60430
#>  [50] 190.45212 206.87430 216.68015 223.01863 225.40294 248.95673 252.19682
#>  [57] 265.95331 268.24020 278.07310 285.15063 294.30565 309.59434 325.50344
#>  [64] 336.65976 347.40803 366.65161 392.57853 392.18004 409.07812 436.87902
#>  [71] 446.95624 468.48966 485.73618 502.67666 516.74521 535.42321 556.36571
#>  [78] 576.38850 610.77514 638.95046 108.44282  97.99786 105.86022 110.52438
#>  [85] 107.95405 108.56750 115.28249 115.03108 121.66713 114.98437 116.97504
#>  [92] 127.57991 117.08135 122.17334 125.28645 116.43267 124.69352 126.22864
#>  [99] 120.87107 120.31530 109.66896 115.02986 127.56204 129.37688 134.84527
#> [106] 133.13926 131.52069 133.75108 128.45495 138.09869 133.43696 130.58983
#> [113] 137.08745 134.55331 134.34858 134.57154 128.34012 129.07435 136.25713
#> [120] 133.32114 138.05536 140.40348 140.08544 142.06562 149.70205 164.56587
#> [127] 171.43835 183.74445 201.78900 209.62498 228.22502 231.52111 251.03099
#> [134] 261.21460 258.35878 274.80627 287.16846 308.46627 329.32075 337.05790
#> [141] 133.32114 134.52287 135.65912 133.16026 137.79048 137.69715 127.41545
#> [148] 126.43122 127.49278 124.18735 132.27290 129.31601 113.37814 110.87841
#> [155] 115.05622 123.42495 131.63186 133.03501 124.99366 125.27983 135.59669
#> [162] 123.47309 125.81676 129.56149 137.78202 147.87801 133.19805 130.86545
#> [169] 132.13859 122.18638 117.55266 121.51437 111.31007 116.01410 123.20072
#> [176] 132.86109 121.98610 110.46131 120.26792 125.89852 120.78836 111.92792
#> [183] 103.32577  97.04478 100.44584  78.43935  72.61123  77.65693  71.84845
#> [190]  67.71015  72.11576  66.00604  65.10498  73.27928  74.26771  56.55609
#> [197]  53.01011  43.08856  54.74897  53.92645