Simulation of a two-bubble process

The following data generating process is similar to sim_dgp1, with the difference that there are two episodes of mildly explosive dynamics.

sim_dgp2(n, te1 = 0.2 * n, tf1 = 0.2 * n + te1, te2 = 0.6 * n,
  tf2 = 0.1 * n + te2, c = 1, alpha = 0.6, sigma = 6.79)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te1
A scalar in (0, n) specifying the observation in which the first bubble originates.

tf1
A scalar in (te1, n) specifying the observation in which the first bubble collapses.

te2
A scalar in (tf1, n) specifying the observation in which the second bubble originates.

tf2
A scalar in (te2, n) specifying the observation in which the second bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

Value

A numeric vector of length n.

Details

The data generating process is described by:

$$X_t = X_{t-1}1\{t \in N_0\}+ \delta_T X_{t-1}1\{t \in B_1 \cup B_2\} + \left(\sum_{k=\tau_{1f}+1}^t \epsilon_k + X^*_{\tau_{1f}}\right) 1\{t \in N_1\} $$

$$ + \left(\sum_{l=\tau_{2f}+1}^t \epsilon_l + X^*_{\tau_{2f}}\right) 1\{t \in N_2\} + \epsilon_t 1\{t \in N_0 \cup B_1 \cup B_2\}$$

where the autoregressive coefficient \(\delta_T\) is given:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\), \(X_{\tau_{1f}} = X_{\tau_{1e}} + X^*\) and \(X_{\tau_{2f}} = X_{\tau_{2e}} + X^*\). We use the notation \(N_0 = [1, \tau_{1e})\), \(B_1 = [\tau_{1e}, \tau_{1f}]\), \(N_1 = (\tau_{1f}, \tau_{2e})\), \(B_2 = [\tau_{2e}, \tau_{2f}]\), \(N_2 = (\tau_{2f}, \tau]\), where \(\tau\) is the last observation of the sample. After the collapse of the first bubble, \(X_t\) resumes a martingale path until time \(\tau_{2e}-1\), and a second episode of exuberance begins at \(\tau_{2e}\). The expansion process lasts until \(\tau_{2f}\) and collapses to a value of \(X^*_{\tau_{2f}}\). The process then continues on a martingale path until the end of the sample period \(\tau\). The expansion duration of the first bubble is assumed to be longer than that of the second bubble, i.e. \(\tau_{1f}-\tau_{1e}>\tau_{2f}-\tau_{2e}\).

For further details the user can refer to Phillips et al., (2015) p. 1055.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination dates 20/60 and termination dates 40/70 respectively
sim_dgp2(n = 100)

#>   [1] 100.00000 110.52836 125.09015 134.05696 128.66394 118.03424 117.66766
#>   [8] 125.67813 109.95265 107.40987 102.03002 105.90338 106.98828  99.32445
#>  [15] 102.21531  99.03171 103.86297  97.93696 104.40201 119.67750 121.37870
#>  [22] 120.90522 128.85674 136.79571 139.48543 141.96109 144.86699 156.45372
#>  [29] 164.85326 172.82334 177.86016 183.66994 195.71281 199.00327 215.89695
#>  [36] 229.77891 253.22695 273.37227 296.32712 324.91998 119.70695 118.25305
#>  [43] 117.86709 107.95484 114.38140 127.54494 125.45517 111.20320 107.31958
#>  [50] 105.64443 107.13817 109.35501 104.73403 104.93703  95.36941 103.46112
#>  [57] 108.67038 105.33562 113.25570 129.44761 136.22975 139.14525 142.90844
#>  [64] 144.32916 160.35008 169.40591 183.44726 189.80655 196.11203 214.59387
#>  [71] 133.93368 144.28120 144.05650 136.65842 138.38751 137.98978 135.79839
#>  [78] 138.68838 136.72432 153.44117 161.03480 158.25853 175.97876 181.18453
#>  [85] 184.32414 190.71918 187.50558 170.80183 169.12981 183.24518 175.33713
#>  [92] 177.67533 179.27699 168.21728 159.48169 167.08563 171.71494 185.24503
#>  [99] 191.69806 187.18858

# 200 periods with bubble origination dates 40/120 and termination dates 80/140 respectively
sim_dgp2(n = 200)

#>   [1] 100.00000  98.32932  95.91879  95.27656  92.13119  98.63851 102.29054
#>   [8]  96.87921  93.76523  98.50230 107.26975 101.09534  94.59831  93.07896
#>  [15]  79.56086  78.74756  77.37058  80.16355  90.34747  89.65741  89.96256
#>  [22]  95.75805 103.38330 103.30378 104.28949 107.00126 115.95646 120.71679
#>  [29] 126.75086 135.66755 135.67168 129.84287 121.73722 114.97189 114.42731
#>  [36] 113.90225 109.44571 100.96412  96.96728 102.26928 105.41833 118.62356
#>  [43] 124.32619 129.92623 136.38406 136.05837 142.69662 145.34012 139.88036
#>  [50] 160.11522 147.40376 160.76564 175.51014 172.84237 176.59046 189.10601
#>  [57] 196.72414 223.68752 232.04697 238.22813 250.97170 262.37510 277.44918
#>  [64] 290.73018 303.62489 318.74916 334.51844 339.16464 362.25638 379.72987
#>  [71] 391.79359 411.13238 416.23077 433.81952 447.99994 471.45155 500.02812
#>  [78] 512.11772 541.33009 564.09936  93.72427 102.25327 112.02230 124.48701
#>  [85] 119.11209 132.66018 134.36448 149.41165 150.41184 156.63839 156.72279
#>  [92] 151.66050 158.92743 138.98612 134.02071 118.53274 106.09184 111.48883
#>  [99] 119.33984 122.69968 114.34876 108.84215 113.74812  94.19457 100.74975
#> [106] 100.81306 107.33371 114.99191 119.50715 133.91484 127.00085 114.74710
#> [113] 115.43494 114.13884 118.11001 110.49913 110.50314 122.51213 125.69639
#> [120] 138.30509 133.41573 153.30958 168.18771 173.36541 172.02514 177.06810
#> [127] 179.41593 197.72778 210.45770 228.69139 233.50701 248.37818 265.44918
#> [134] 273.11707 294.53487 298.36370 311.77776 328.14081 348.64587 351.90946
#> [141] 142.25687 153.38162 156.90840 148.72814 143.87192 147.29128 146.51200
#> [148] 140.74043 149.88188 154.12025 165.00498 155.10476 155.48467 164.34546
#> [155] 173.72655 169.41403 171.39371 177.86375 184.14388 180.55773 191.19672
#> [162] 195.39026 202.02629 197.27226 195.72450 187.08980 188.76465 175.44362
#> [169] 173.15400 178.46611 184.55212 191.22736 194.30061 190.64918 188.54360
#> [176] 183.81420 180.37382 180.82952 191.06424 182.54587 182.95757 184.03141
#> [183] 171.88296 175.10623 168.30942 174.73742 175.69096 179.94165 184.70557
#> [190] 189.87836 181.51990 180.92358 175.56888 170.86587 160.41756 155.99360
#> [197] 151.35924 143.97968 142.23637 142.09411