Simulation of a two-bubble process

The following data generating process is similar to sim_dgp1, with the difference that there are two episodes of mildly explosive dynamics.

sim_dgp2(n, te1 = 0.2 * n, tf1 = 0.2 * n + te1, te2 = 0.6 * n,
  tf2 = 0.1 * n + te2, c = 1, alpha = 0.6, sigma = 6.79)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te1
A scalar in (0, n) specifying the observation in which the first bubble originates.

tf1
A scalar in (te1, n) specifying the observation in which the first bubble collapses.

te2
A scalar in (tf1, n) specifying the observation in which the second bubble originates.

tf2
A scalar in (te2, n) specifying the observation in which the second bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

Value

A numeric vector of length n.

Details

The data generating process is described by:

$$X_t = X_{t-1}1\{t \in N_0\}+ \delta_T X_{t-1}1\{t \in B_1 \cup B_2\} + \left(\sum_{k=\tau_{1f}+1}^t \epsilon_k + X^*_{\tau_{1f}}\right) 1\{t \in N_1\} $$

$$ + \left(\sum_{l=\tau_{2f}+1}^t \epsilon_l + X^*_{\tau_{2f}}\right) 1\{t \in N_2\} + \epsilon_t 1\{t \in N_0 \cup B_1 \cup B_2\}$$

where the autoregressive coefficient \(\delta_T\) is given:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\), \(X_{\tau_{1f}} = X_{\tau_{1e}} + X^*\) and \(X_{\tau_{2f}} = X_{\tau_{2e}} + X^*\). We use the notation \(N_0 = [1, \tau_{1e})\), \(B_1 = [\tau_{1e}, \tau_{1f}]\), \(N_1 = (\tau_{1f}, \tau_{2e})\), \(B_2 = [\tau_{2e}, \tau_{2f}]\), \(N_2 = (\tau_{2f}, \tau]\), where \(\tau\) is the last observation of the sample. After the collapse of the first bubble, \(X_t\) resumes a martingale path until time \(\tau_{2e}-1\), and a second episode of exuberance begins at \(\tau_{2e}\). The expansion process lasts until \(\tau_{2f}\) and collapses to a value of \(X^*_{\tau_{2f}}\). The process then continues on a martingale path until the end of the sample period \(\tau\). The expansion duration of the first bubble is assumed to be longer than that of the second bubble, i.e. \(\tau_{1f}-\tau_{1e}>\tau_{2f}-\tau_{2e}\).

For further details the user can refer to Phillips et al., (2015) p. 1055.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination dates 20/60 and termination dates 40/70 respectively
sim_dgp2(n = 100)

#>   [1] 100.00000  97.86995  95.92597  90.47061  89.82858  89.32580  94.17719
#>   [8]  95.02403 103.82791  96.26050  90.54292  80.32876  78.40022  78.69131
#>  [15]  78.63110  58.60682  58.74506  58.07660  62.07665  63.10257  71.39341
#>  [22]  71.96984  75.36258  67.08605  60.90123  57.17461  71.63122  71.80657
#>  [29]  86.97509  82.62307  82.46193  84.23950  92.28305 104.70150 107.37181
#>  [36] 125.03889 134.46267 139.45081 142.65604 153.92771  62.46777  64.53235
#>  [43]  61.29686  59.65836  65.25437  54.69155  55.32643  52.83484  51.95615
#>  [50]  54.72509  50.75994  49.44641  47.61619  48.11633  50.54196  54.27937
#>  [57]  54.54012  43.61110  36.48357  52.71884  57.24318  68.51618  75.79293
#>  [64]  84.30171  94.01676 105.91359 114.97820 122.57152 135.98000 142.64684
#>  [71]  47.34334  47.35457  39.29124  41.75205  38.02138  42.72651  42.31338
#>  [78]  34.20924  33.39506  28.58701  17.60986  20.97667  29.80198  18.82944
#>  [85]  10.33944  21.08212  29.58679  28.16499  23.82595  26.53451  35.64651
#>  [92]  35.39115  39.08267  50.89875  48.69570  45.65710  48.48230  39.68337
#>  [99]  44.31845  47.69249

# 200 periods with bubble origination dates 40/120 and termination dates 80/140 respectively
sim_dgp2(n = 200)

#>   [1] 100.0000 109.7753 117.9733 119.5928 117.5637 126.9652 122.2118 134.8475
#>   [9] 147.0961 137.8886 135.7515 128.8306 124.6735 122.7107 132.9431 135.4399
#>  [17] 135.3039 128.6502 119.1380 128.9497 126.0743 121.0921 116.2815 126.2974
#>  [25] 132.0349 140.8212 142.8457 140.0911 139.1486 137.6372 149.5003 148.9456
#>  [33] 149.7885 152.4518 155.0212 156.1496 163.9811 164.0812 164.2268 165.4522
#>  [41] 172.6615 177.8588 193.4624 201.7331 211.3557 220.6778 233.6557 242.2493
#>  [49] 250.7855 262.8728 267.5112 288.9367 295.4537 316.6561 335.3755 350.1514
#>  [57] 370.7147 373.9333 383.0008 395.7254 399.5660 417.9216 433.8629 435.8300
#>  [65] 453.9947 477.4411 498.8559 515.6033 534.3139 566.1588 591.3673 606.2335
#>  [73] 629.3337 652.2685 685.0991 713.2627 749.4180 790.7623 821.3426 859.9180
#>  [81] 155.0330 159.1048 155.6802 152.9172 149.7875 164.8267 174.9232 172.5121
#>  [89] 178.4615 177.0037 190.7250 178.2405 179.4147 174.2323 166.6377 174.7720
#>  [97] 181.4343 189.5262 193.8232 197.4411 191.8724 190.7906 190.6716 188.0190
#> [105] 185.2841 193.9530 199.5281 202.3784 214.0006 198.9564 197.7011 202.2584
#> [113] 196.8594 186.2245 194.4657 204.6766 196.7362 200.9010 210.9769 216.7462
#> [121] 227.5925 228.7271 241.4977 262.8890 273.5849 281.9796 298.7044 307.2325
#> [129] 320.5304 334.2760 353.9508 386.1719 402.8874 416.1282 438.7090 465.3115
#> [137] 497.4506 517.3079 533.6880 552.4460 215.6867 216.5869 209.9729 201.4789
#> [145] 201.1643 208.8871 217.4728 213.7468 218.3802 212.0671 205.4458 205.2617
#> [153] 208.2946 206.0599 206.8275 213.9654 216.1856 221.5295 224.5277 217.6067
#> [161] 216.5291 205.4091 224.6877 231.2185 234.6707 236.4329 231.1815 218.9219
#> [169] 220.1332 216.5631 219.2804 218.4480 217.9636 232.1936 241.1706 249.3224
#> [177] 244.0134 241.0158 239.3453 241.8290 241.5972 243.6086 242.1076 243.3822
#> [185] 240.4077 239.8052 244.1118 249.9888 247.0651 247.1535 247.5015 251.1589
#> [193] 254.6874 238.2359 232.7612 229.4702 225.0358 223.0411 221.1604 222.0870