Simulation of a single-bubble process

The following function generates a time series which switches from a martingale to a mildly explosive process and then back to a martingale.

sim_psy1(n, te = 0.4 * n, tf = 0.15 * n + te, c = 1, alpha = 0.6,
  sigma = 6.79, seed = NULL)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te
A scalar in (0, tf) specifying the observation in which the bubble originates.

tf
A scalar in (te, n) specifying the observation in which the bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to `set.seed` before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute.

Value

A numeric vector of length n.

Details

The data generating process is described by the following equation: $$X_t = X_{t-1}1\{t < \tau_e\}+ \delta_T X_{t-1}1\{\tau_e \leq t\leq \tau_f\} + \left(\sum_{k=\tau_f+1}^t \epsilon_k + X^*_{\tau_f}\right) 1\{t > \tau_f\} + \epsilon_t 1\{t \leq \tau_f\}$$

where the autoregressive coefficient \(\delta_T\) is given by:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\) and \(X_{\tau_f} = X_{\tau_e} + X^*\). During the pre- and post- bubble periods, \(N_0 = [1, \tau_e)\), X is a pure random walk process. During the bubble expansion period \(B = [\tau_e, \tau_f]\) is a mildly explosive process with expansion rate given by the autoregressive coefficient \(\delta_T\), and continues its martingale path for the subsequent period \(N_1 = (\tau_f, \tau]\).

For further details the user can refer to Phillips et al. (2015) p. 1054.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination date 40 and termination date 55
sim_psy1(n = 100)

#>   [1] 100.00000 104.92065  88.77733  88.33320  85.14390  83.23920  87.14234
#>   [8]  85.74579  87.34063  88.09180  89.95363 100.64734 104.50248 104.09771
#>  [15] 103.08201 104.88930 108.47716 104.54311 110.94064 106.14445  97.48447
#>  [22] 102.46331  99.61010  97.87610  95.36897 103.60150 100.57748 108.30904
#>  [29] 109.60269 100.32488  91.51076  76.99986  76.34663  81.06235  77.83987
#>  [36]  89.81153 103.56702 106.47055 102.62416 103.43756 112.66821 109.76498
#>  [43] 114.00564 120.31144 114.32601 114.77713 124.87165 132.32369 140.12865
#>  [50] 148.75097 153.25464 155.34290 159.84634 157.86642 164.92212  89.64689
#>  [57] 108.32458 118.60240 118.83319 123.86317 115.23438 114.12378 109.97546
#>  [64] 118.47575 120.17542 108.59452 102.78627 101.80239  99.59939  98.42768
#>  [71]  90.03481  77.11816  76.47648  76.69753  79.82969  89.20940  86.38153
#>  [78]  91.00513  88.19154  84.67197  80.02748  74.01392  74.34824  75.60820
#>  [85]  71.47542  66.51123  84.94706  75.85262  71.46618  65.13481  59.91271
#>  [92]  62.43573  64.84912  58.16777  59.62579  59.08201  49.42412  56.94845
#>  [99]  64.27132  61.28202
#> attr(,"seed")
#>   [1]       10403         200  1901576442   305780618  1771555321   140805362
#>   [7] -1701321808  1010115073  1617144363  1953263465  1802686505  1454118368
#>  [13] -2086616346 -1842807025 -2117162845  1745683843   934523388  -734262448
#>  [19] -2049594901  2015105553  1611510008  -324706675  -606067377  1407800146
#>  [25]  1515101296 -1950581902  1518616120  -629054251   599773999 -2047790264
#>  [31]   304401920 -1054243178  1069407915   883488859 -2008579564  1371823036
#>  [37]  1439620823  -946554104  1854423094  1759924975 -1216782695 -1139259265
#>  [43]   327340523   894318778 -2117641404  1965290645   225956641   494447817
#>  [49] -2010179086  -649663946  -231758035  -898707605  2139143054  1628405715
#>  [55]   956945101 -1568134152 -1446113978  1892913848  2017957546   199415119
#>  [61]  1790039181  -468342618   501539198  -562077348 -1658683111   705664625
#>  [67]   980877302 -1957052298 -1628468859  1848635206  1813330132  1705692725
#>  [73] -1003087097 -1275258027  1172389957   370743324  -264583774  -498197861
#>  [79]  2016058631 -1726134153   409950272   -25621124  2083601871   614263141
#>  [85]   207536284 -1206256263  -283341557   217860654  2091420596   394827334
#>  [91] -1614207588 -1248309767 -1430542813 -2109435812 -1173681380 -1909852638
#>  [97]  1621007303  1273511183  -418519056  2138334824  -367745741  -938199940
#> [103] -1805861126  1314589251  2085500197  1130560139   691940695  -408974170
#> [109] -1902962144 -1954449663  2039328229  1644769277 -1182385738  2038219554
#> [115]   337823601   472194355 -1939439954 -1061723946 -1886400977   996732383
#> [121]  1422785625   211357998  -346633083 -1409636555   -58880725  1052223605
#> [127]   282245830  1822954861   875710060   915512429  1323664643  -263286750
#> [133]  -612798999   286342823 -1419243366  1468642094 -2069778668  1345010447
#> [139]   -73411786   838819202   289537784 -1311577548 -1014895613  -918621414
#> [145]  1826951849   -96151748   980322846  -803763027 -1699569644  -809391132
#> [151]  -865448823  -947733159 -1855606333  1216206720 -1472762525 -1763951869
#> [157]  1837205813  -549894825  1486649036 -1379691261 -1762168990  1906633971
#> [163]  1656679941  -121901876  -584939041   243089765 -1115695948 -1712702392
#> [169]  1210660686 -1181537391  2058198156  1859124400  -592619022  -948353695
#> [175]  -429954978  -540629814    60801539  1301891067  2043239177  -877833093
#> [181]  1773495724  1851170785 -1099624840  -373710907 -1987896811 -1003750521
#> [187]   618586020   634950401   666813757  -387662482  1510350629 -1555200469
#> [193]  -142694436    13290928   262282462  1641692642  -555419877  1448114120
#> [199]  -225916771   -77952692 -2098902360   143570574   176619519  1926182646
#> [205] -1628636458 -1794799193  -268363776  1200745916   337526889   316819017
#> [211] -1953654301   451074333  -795557722  -345014109  -113704090  -278430133
#> [217]  -132544289  2020000577   833514186   489884175 -1839850229   213250684
#> [223]   932982895   371670189  1177485754  -743849546  -880968624  1649975908
#> [229] -1206746827  1560320122  1478136347 -1723815310    52828146 -1585102888
#> [235]  -305403131  1003191412  -416690548  1738677189   390955978  1253648174
#> [241]  -681695625 -1807050561  -337901675  1414519455   238094000 -1391251219
#> [247]  1872146220 -1390313447  -968888871  -247588349   -92702416  1826821133
#> [253]   234363713   438448226   -10581103 -1191496401 -1144758464   476145364
#> [259]   209683178   827894950 -1566112561  -783678748  1641006273 -1083388544
#> [265] -1182531652  2110046026  -555500373  -590189870  1422566522  1859071243
#> [271]  1743164236   667836192  1705173213     7235437   -83886353 -2081055231
#> [277]   951567850 -1268458257  -149264358 -1409382833  1650185827  1296031373
#> [283]  1219595846  -302308117  -208533415  1007814606  -574109056  -582308574
#> [289]  1179677980 -1748538852   352247261   680866386 -1579452542  1195456651
#> [295]   758675167   962398808  -221812549 -2091453857  -217958427   603594135
#> [301]    19858768  1711139405  -207319969   398107947  2064946991  -741884697
#> [307]   609021670 -1409051369  1618891393  1794216642 -1022057890  -553780759
#> [313]  1794137386  1789485030 -1509134442  -505559000 -1301788737 -1849607148
#> [319]  2091306890   684395500   307513562  1090852530   808509091   -75541348
#> [325]  -700603596 -1846293687 -1199714551   979506766  2139974729  1055041689
#> [331] -1183363849 -1944245683  -487858678   148288759   507856461  1888084705
#> [337]   657025921 -1377064551 -1217760928  1121179749 -1761045329   672851084
#> [343]   351591716  -570288429   408784692   730097636   486338084   598407178
#> [349] -1235158915  -956114974 -1163157548 -1756640994 -1331824712  -316097824
#> [355]  1166711105  2027142942  -142031970  1561237751 -1488038021 -1866054596
#> [361] -1324464369   700214163    41689569  -954701517  -530929780 -1661164239
#> [367]  1689609459  1145400807  1044335987   674262091  -662644382  1641922203
#> [373]    65727125  1015764214   -82969342  -741282355   352086030   956123610
#> [379] -1014457526  -860158196 -1980139653  -645399440  -710315498  2131544416
#> [385]  1436831966 -2047646402 -2134823265 -1616458112 -1257142792  1698418253
#> [391]  -675233587  -459505910 -1444455259  2031951285 -1127648501  1683329857
#> [397]  1531399422   613626531   848994329 -1669724523   544762618   470767874
#> [403] -1792486666  1918590221 -1000388702  1347966225  -877522972  1300507977
#> [409]  -234489037  2046991115 -1542463586 -1311421890   634663188  1641067723
#> [415]   294700472  1045111144   410775387 -1675373237 -1968709025 -2142821890
#> [421] -1086498917    27069170    52876731  -630594594 -1575195400  1210655492
#> [427] -1367816053 -1041735841 -1877146523  1580707424   747238751 -1959903927
#> [433]  2055285744   130597896  1641299916   826286291 -1498839504  -529169917
#> [439]  1162385618  -950360425  1510986833  1037445429  -288190576 -2142258264
#> [445] -1079335950   671599329   731175426 -2039647530 -1366551011  -203469483
#> [451]  1971435297  1977493696   626266925 -1371721248  1890587437 -2003908072
#> [457]  -594035334  -572003758  -119994999  1844935237  1060719551  -404862454
#> [463]  1402855085 -1253984685  -983743810 -1983843962   738680266  -579042159
#> [469]   142776510 -1599085763   702367112   964303373    11985967 -1392791385
#> [475]  -441949022 -1093205974  2044365536 -1884940281  1404203164  1810122580
#> [481] -1048681105   782469431 -1160076381  -667776494 -1491897593    40987662
#> [487]   856119567  -866410038   897137628 -2142127448  -159307297 -1833624821
#> [493]  -638857543 -1113006884 -2108031021  1196716645 -1914221716  -104470972
#> [499]   109077240   184171327 -1899766828  -713890361 -1782040274   826641571
#> [505]  1919649573   246096393  -886333364   152139356 -1166652570  1257510101
#> [511] -1338406050 -1581635716 -1533523433   537841250  -397373314   551709038
#> [517]  1327022135  1514088981   355067971  -148636313  -161786485 -1989070386
#> [523]  1004598619 -1407994476   741608656  -556033062  1505965934  1262777047
#> [529] -1754864090    56673075  1863465607   361512481   923425358   386555630
#> [535] -1129674712   225497764 -1607000160   589205069 -1438071198  1793262145
#> [541] -2050010091 -1595912759  -290405791  -776635846  1191243713 -1725087772
#> [547]   256099640  1366167112  -160174087  1072736863  -990506139 -1216276687
#> [553]   406205353  2113993216  1398945785    99553918  -221096094   -77759336
#> [559]  -332365804 -2084880347 -1041237392   -25869515  1065374289  1230321707
#> [565]   117242656  -863165056    71660730 -1995045810 -1007510146   601500267
#> [571]   734128644   993514042  -560188255 -1484152332  -654020108  -718019342
#> [577]   173447480 -1522145048  1209570648  -822192819  1708798839  -816590956
#> [583] -2024156914   -67808469 -1236732778 -1054748926  1840624335  -585674475
#> [589]   -13278746  -577959879   296472509 -1261427139    20860509    40693173
#> [595] -1611182043  1118487926 -1941581646  1395424539  -731512159   181976060
#> [601] -1063268567  -460821539   829971186    37180268 -1310603945  1125669166
#> [607]   831483814 -1298304796  1420170318 -2130483906  -685110146   140733163
#> [613]   180767641   456378150   174437848 -1172413523 -1610649732  -342449140
#> [619]  1634032437  -290005609  -997805860   861750131  2047212673  -272674113
#> [625] -2056673838   170999908
#> attr(,"class")
#> [1] "sim"     "numeric"

# 200 periods with bubble origination date 80 and termination date 110
sim_psy1(n = 200)

#>   [1] 100.000000  94.716232  82.368006  74.735497  81.936454  71.896080
#>   [7]  61.360986  66.640256  73.894965  72.648405  83.228874  81.780958
#>  [13]  88.099293  90.888705  82.198562  76.886293  61.431686  60.573458
#>  [19]  70.406374  60.610051  70.572252  65.525246  63.459566  65.752264
#>  [25]  61.627325  59.620471  58.706964  64.239350  62.386211  77.049120
#>  [31]  84.462018  89.509601  81.308236  83.534120  79.907139  88.625137
#>  [37]  88.821581  86.134934  81.416778  71.285550  81.092057  78.735809
#>  [43]  78.462449  86.927081  77.785694  73.888904  69.095441  56.483664
#>  [49]  54.806734  52.528275  50.835007  53.955165  50.831252  51.231406
#>  [55]  46.189212  31.498545  29.917178  35.857795  29.247771  25.766220
#>  [61]  19.495736   9.348585   8.221918  18.965316  25.854466  18.069878
#>  [67]  11.892002  33.003353  25.741090  20.987346  19.639908  24.689089
#>  [73]  24.348504  33.264748  33.848780  29.572089  34.020017  34.151699
#>  [79]  38.879372  54.138056  56.004814  59.968585  62.316302  59.269197
#>  [85]  71.259235  69.310519  81.230395  79.102875  94.584038 103.183543
#>  [91] 108.085709 114.788743 120.051380 126.631237 145.390096 144.980501
#>  [97] 147.222518 155.088028 160.199042 169.779232 177.327273 174.275114
#> [103] 184.612333 198.234478 197.674878 198.577447 199.138130 207.581336
#> [109] 216.076739 223.178643  53.973695  59.575582  54.481841  58.122710
#> [115]  47.443959  40.746116  54.214165  41.643306  35.465844  22.215566
#> [121]  16.781689   4.088724  -1.007657  -5.020128 -10.058982  -5.350035
#> [127]  -5.753796 -18.409643 -27.063559 -39.161783 -42.615132 -54.408536
#> [133] -54.133667 -54.976846 -59.137442 -58.046544 -62.543965 -64.817608
#> [139] -60.587360 -53.605920 -61.309707 -55.083881 -46.841671 -51.499152
#> [145] -58.931270 -56.453775 -63.298211 -59.529785 -56.568898 -55.853413
#> [151] -57.576293 -64.206668 -61.529930 -54.638738 -50.351235 -53.633509
#> [157] -63.997316 -59.808957 -51.462100 -50.392994 -40.843931 -38.149653
#> [163] -35.933241 -45.842182 -51.269399 -48.078012 -55.887737 -57.731879
#> [169] -54.625969 -54.741113 -58.418539 -52.471766 -47.441475 -48.560551
#> [175] -53.659598 -62.169955 -69.205999 -69.405887 -71.272756 -67.976540
#> [181] -61.750232 -68.440717 -78.565896 -73.000863 -66.119846 -58.965867
#> [187] -59.450867 -52.983768 -55.591383 -58.427973 -61.621094 -73.097418
#> [193] -79.517347 -81.846399 -79.412439 -79.085236 -71.411367 -74.993990
#> [199] -71.737674 -62.112650
#> attr(,"seed")
#>   [1]       10403         398  1901576442   305780618  1771555321   140805362
#>   [7] -1701321808  1010115073  1617144363  1953263465  1802686505  1454118368
#>  [13] -2086616346 -1842807025 -2117162845  1745683843   934523388  -734262448
#>  [19] -2049594901  2015105553  1611510008  -324706675  -606067377  1407800146
#>  [25]  1515101296 -1950581902  1518616120  -629054251   599773999 -2047790264
#>  [31]   304401920 -1054243178  1069407915   883488859 -2008579564  1371823036
#>  [37]  1439620823  -946554104  1854423094  1759924975 -1216782695 -1139259265
#>  [43]   327340523   894318778 -2117641404  1965290645   225956641   494447817
#>  [49] -2010179086  -649663946  -231758035  -898707605  2139143054  1628405715
#>  [55]   956945101 -1568134152 -1446113978  1892913848  2017957546   199415119
#>  [61]  1790039181  -468342618   501539198  -562077348 -1658683111   705664625
#>  [67]   980877302 -1957052298 -1628468859  1848635206  1813330132  1705692725
#>  [73] -1003087097 -1275258027  1172389957   370743324  -264583774  -498197861
#>  [79]  2016058631 -1726134153   409950272   -25621124  2083601871   614263141
#>  [85]   207536284 -1206256263  -283341557   217860654  2091420596   394827334
#>  [91] -1614207588 -1248309767 -1430542813 -2109435812 -1173681380 -1909852638
#>  [97]  1621007303  1273511183  -418519056  2138334824  -367745741  -938199940
#> [103] -1805861126  1314589251  2085500197  1130560139   691940695  -408974170
#> [109] -1902962144 -1954449663  2039328229  1644769277 -1182385738  2038219554
#> [115]   337823601   472194355 -1939439954 -1061723946 -1886400977   996732383
#> [121]  1422785625   211357998  -346633083 -1409636555   -58880725  1052223605
#> [127]   282245830  1822954861   875710060   915512429  1323664643  -263286750
#> [133]  -612798999   286342823 -1419243366  1468642094 -2069778668  1345010447
#> [139]   -73411786   838819202   289537784 -1311577548 -1014895613  -918621414
#> [145]  1826951849   -96151748   980322846  -803763027 -1699569644  -809391132
#> [151]  -865448823  -947733159 -1855606333  1216206720 -1472762525 -1763951869
#> [157]  1837205813  -549894825  1486649036 -1379691261 -1762168990  1906633971
#> [163]  1656679941  -121901876  -584939041   243089765 -1115695948 -1712702392
#> [169]  1210660686 -1181537391  2058198156  1859124400  -592619022  -948353695
#> [175]  -429954978  -540629814    60801539  1301891067  2043239177  -877833093
#> [181]  1773495724  1851170785 -1099624840  -373710907 -1987896811 -1003750521
#> [187]   618586020   634950401   666813757  -387662482  1510350629 -1555200469
#> [193]  -142694436    13290928   262282462  1641692642  -555419877  1448114120
#> [199]  -225916771   -77952692 -2098902360   143570574   176619519  1926182646
#> [205] -1628636458 -1794799193  -268363776  1200745916   337526889   316819017
#> [211] -1953654301   451074333  -795557722  -345014109  -113704090  -278430133
#> [217]  -132544289  2020000577   833514186   489884175 -1839850229   213250684
#> [223]   932982895   371670189  1177485754  -743849546  -880968624  1649975908
#> [229] -1206746827  1560320122  1478136347 -1723815310    52828146 -1585102888
#> [235]  -305403131  1003191412  -416690548  1738677189   390955978  1253648174
#> [241]  -681695625 -1807050561  -337901675  1414519455   238094000 -1391251219
#> [247]  1872146220 -1390313447  -968888871  -247588349   -92702416  1826821133
#> [253]   234363713   438448226   -10581103 -1191496401 -1144758464   476145364
#> [259]   209683178   827894950 -1566112561  -783678748  1641006273 -1083388544
#> [265] -1182531652  2110046026  -555500373  -590189870  1422566522  1859071243
#> [271]  1743164236   667836192  1705173213     7235437   -83886353 -2081055231
#> [277]   951567850 -1268458257  -149264358 -1409382833  1650185827  1296031373
#> [283]  1219595846  -302308117  -208533415  1007814606  -574109056  -582308574
#> [289]  1179677980 -1748538852   352247261   680866386 -1579452542  1195456651
#> [295]   758675167   962398808  -221812549 -2091453857  -217958427   603594135
#> [301]    19858768  1711139405  -207319969   398107947  2064946991  -741884697
#> [307]   609021670 -1409051369  1618891393  1794216642 -1022057890  -553780759
#> [313]  1794137386  1789485030 -1509134442  -505559000 -1301788737 -1849607148
#> [319]  2091306890   684395500   307513562  1090852530   808509091   -75541348
#> [325]  -700603596 -1846293687 -1199714551   979506766  2139974729  1055041689
#> [331] -1183363849 -1944245683  -487858678   148288759   507856461  1888084705
#> [337]   657025921 -1377064551 -1217760928  1121179749 -1761045329   672851084
#> [343]   351591716  -570288429   408784692   730097636   486338084   598407178
#> [349] -1235158915  -956114974 -1163157548 -1756640994 -1331824712  -316097824
#> [355]  1166711105  2027142942  -142031970  1561237751 -1488038021 -1866054596
#> [361] -1324464369   700214163    41689569  -954701517  -530929780 -1661164239
#> [367]  1689609459  1145400807  1044335987   674262091  -662644382  1641922203
#> [373]    65727125  1015764214   -82969342  -741282355   352086030   956123610
#> [379] -1014457526  -860158196 -1980139653  -645399440  -710315498  2131544416
#> [385]  1436831966 -2047646402 -2134823265 -1616458112 -1257142792  1698418253
#> [391]  -675233587  -459505910 -1444455259  2031951285 -1127648501  1683329857
#> [397]  1531399422   613626531   848994329 -1669724523   544762618   470767874
#> [403] -1792486666  1918590221 -1000388702  1347966225  -877522972  1300507977
#> [409]  -234489037  2046991115 -1542463586 -1311421890   634663188  1641067723
#> [415]   294700472  1045111144   410775387 -1675373237 -1968709025 -2142821890
#> [421] -1086498917    27069170    52876731  -630594594 -1575195400  1210655492
#> [427] -1367816053 -1041735841 -1877146523  1580707424   747238751 -1959903927
#> [433]  2055285744   130597896  1641299916   826286291 -1498839504  -529169917
#> [439]  1162385618  -950360425  1510986833  1037445429  -288190576 -2142258264
#> [445] -1079335950   671599329   731175426 -2039647530 -1366551011  -203469483
#> [451]  1971435297  1977493696   626266925 -1371721248  1890587437 -2003908072
#> [457]  -594035334  -572003758  -119994999  1844935237  1060719551  -404862454
#> [463]  1402855085 -1253984685  -983743810 -1983843962   738680266  -579042159
#> [469]   142776510 -1599085763   702367112   964303373    11985967 -1392791385
#> [475]  -441949022 -1093205974  2044365536 -1884940281  1404203164  1810122580
#> [481] -1048681105   782469431 -1160076381  -667776494 -1491897593    40987662
#> [487]   856119567  -866410038   897137628 -2142127448  -159307297 -1833624821
#> [493]  -638857543 -1113006884 -2108031021  1196716645 -1914221716  -104470972
#> [499]   109077240   184171327 -1899766828  -713890361 -1782040274   826641571
#> [505]  1919649573   246096393  -886333364   152139356 -1166652570  1257510101
#> [511] -1338406050 -1581635716 -1533523433   537841250  -397373314   551709038
#> [517]  1327022135  1514088981   355067971  -148636313  -161786485 -1989070386
#> [523]  1004598619 -1407994476   741608656  -556033062  1505965934  1262777047
#> [529] -1754864090    56673075  1863465607   361512481   923425358   386555630
#> [535] -1129674712   225497764 -1607000160   589205069 -1438071198  1793262145
#> [541] -2050010091 -1595912759  -290405791  -776635846  1191243713 -1725087772
#> [547]   256099640  1366167112  -160174087  1072736863  -990506139 -1216276687
#> [553]   406205353  2113993216  1398945785    99553918  -221096094   -77759336
#> [559]  -332365804 -2084880347 -1041237392   -25869515  1065374289  1230321707
#> [565]   117242656  -863165056    71660730 -1995045810 -1007510146   601500267
#> [571]   734128644   993514042  -560188255 -1484152332  -654020108  -718019342
#> [577]   173447480 -1522145048  1209570648  -822192819  1708798839  -816590956
#> [583] -2024156914   -67808469 -1236732778 -1054748926  1840624335  -585674475
#> [589]   -13278746  -577959879   296472509 -1261427139    20860509    40693173
#> [595] -1611182043  1118487926 -1941581646  1395424539  -731512159   181976060
#> [601] -1063268567  -460821539   829971186    37180268 -1310603945  1125669166
#> [607]   831483814 -1298304796  1420170318 -2130483906  -685110146   140733163
#> [613]   180767641   456378150   174437848 -1172413523 -1610649732  -342449140
#> [619]  1634032437  -290005609  -997805860   861750131  2047212673  -272674113
#> [625] -2056673838   170999908
#> attr(,"class")
#> [1] "sim"     "numeric"

# 200 periods with bubble origination date 100 and termination date 150
sim_psy1(n = 200, te = 100, tf = 150)

#>   [1]  100.00000   99.45618   96.77965  100.34615   94.94175   96.60596
#>   [7]   91.31438   95.35391  102.88328   96.48052  101.23586   98.35231
#>  [13]   90.59041   92.17547   90.89055   79.59599   92.60353   87.05682
#>  [19]   89.66180   86.54769   81.68806   84.60469   76.66139   85.63837
#>  [25]   94.91327   93.21765   90.33379   97.79220  102.34776  102.55794
#>  [31]  118.64014  136.19764  144.42903  135.93340  138.65121  144.99787
#>  [37]  141.78845  152.53711  156.59586  159.76057  154.70632  161.21830
#>  [43]  162.25471  157.03069  166.48971  172.72747  179.62670  183.25688
#>  [49]  176.55088  184.61599  181.10188  179.33946  177.11191  177.61051
#>  [55]  175.92752  166.59900  166.32437  169.18661  170.55546  159.03153
#>  [61]  163.39266  156.63498  163.17928  151.93479  159.20594  158.46565
#>  [67]  171.36112  163.64068  161.74139  155.67025  156.59846  151.51162
#>  [73]  155.03019  153.72422  153.91984  156.35467  156.15780  163.94621
#>  [79]  166.48288  168.67836  163.04388  172.51228  171.21170  173.06046
#>  [85]  165.71600  149.89938  146.16742  145.67460  152.68384  154.14463
#>  [91]  150.78806  161.05548  157.12092  168.49274  161.69609  169.99824
#>  [97]  177.31235  173.15724  176.59886  187.07413  204.94061  219.46033
#> [103]  224.95175  243.04487  257.15375  258.97387  271.90492  291.33202
#> [109]  309.65914  317.20748  327.62187  344.49458  360.09745  366.09264
#> [115]  389.29262  407.06980  427.29522  441.44770  469.18168  479.86747
#> [121]  504.15391  531.92832  551.77794  574.16299  586.39328  604.49473
#> [127]  635.18329  665.19883  699.76069  734.53516  767.93091  795.40814
#> [133]  832.24555  863.22767  897.99857  930.17402  964.77635 1016.12298
#> [139] 1064.13266 1110.72623 1154.21400 1208.37229 1266.75863 1319.16804
#> [145] 1379.42514 1444.76917 1503.06585 1563.50491 1626.64598 1688.90409
#> [151]  186.43210  185.92932  190.78071  191.62755  200.43142  192.86402
#> [157]  187.14644  176.93228  175.00374  175.29482  175.23462  155.21034
#> [163]  155.34857  154.68012  158.68017  155.78931  160.09865  156.17046
#> [169]  155.02221  141.99063  131.57296  124.00373  134.85287  130.50859
#> [175]  141.14642  131.30665  125.93234  122.50692  125.23532  131.83110
#> [181]  127.89519  138.78757  140.32193  136.82605  131.23253  133.50321
#> [187]  132.86841  134.93299  131.69750  130.05900  135.65501  125.09219
#> [193]  125.72706  123.23548  122.35679  125.12572  121.16057  119.84705
#> [199]  118.01683  118.51697
#> attr(,"seed")
#>   [1]       10403         172 -1788416688 -1917108519   406657915 -1738333394
#>   [7] -1253475118  1835410024  -211365254  1728304943  1725048761  -204105152
#>  [13] -1440852909  2111847184  1557887008  1041099242  1271837027  -898457198
#>  [19]  -615671617   679294759   863619794  2120171559  -374110121 -1843783261
#>  [25]    71513931  1852099239  1847714155  1791702448  1255258784  -412438645
#>  [31]  -511927276  1456167919  -580164974   682243157  -473165673  -915475388
#>  [37]   604150156   378544087 -1673807493   605919651   661722851 -1786691336
#>  [43]  -570574134  1525447155  -565222496  1902237407   402349331 -1145229557
#>  [49]  1151647738  -880035509  2052236732  -783528998  1515209315  -267522408
#>  [55]  1687829820  1907561870  -699074372  1284757624   343347296  1892149219
#>  [61]  -269664767 -1238665930  1120455403  -392119316   735130093  1319543126
#>  [67] -1326623128   321009315   -19462619   437109152  1990284116 -1282333214
#>  [73]  1616502344  -769305227   846632451   950765310  1154747701    11901694
#>  [79]   726734030   904734400 -1332044103 -1264427356  -526220487  -950991583
#>  [85] -1415067820  1415337465 -1639141627 -1723386147   162866989  1021561793
#>  [91]   262092777  -114534382 -1049365882 -1417368495  -710811622   816591749
#>  [97]    22311556 -1310782421   955138385   568699178 -1647313798  1286487429
#> [103]  -186708287   703361433  1842030685   408919898    42617584   891423285
#> [109]  -765253290   786810465 -1591914115 -1498064963   910849348   589322297
#> [115]   916461882 -1565821632  1079344649   916441014  -463556770   -59339068
#> [121]  1551082114 -1182122274  1003888162   292365249  1327650091   864659512
#> [127]    55513085   908931046  1553720627  -407288528  1935882950  1674764301
#> [133]  -757011265  2057712330  2118707126   895126980  -434670666 -2124062285
#> [139]  1435466853 -1466441508  1006107479  1384153276   243039180  1790274974
#> [145] -1648220585 -1282963602  1622543539   173653963   -20059626  -260156509
#> [151] -1415061445   667518919  1540243871  1239749899  1709685359  -134330388
#> [157]  -929765260  1059851423 -1979291616 -1180049197   445102654  1147619273
#> [163] -1075399101   913706560   541283160   557824011    94381583  1659396231
#> [169]  1229717655  2031336188 -1099826666  -308006521  -647088124  1492016683
#> [175]  1423212639   106653055   407745238  2095606959  1777204752  2123115502
#> [181]   191240135  1461850724   589965936 -1206710622   877560456  -717832484
#> [187]  1880205940  1017063127  -179072587 -1715597950  1438196447 -2071318008
#> [193] -1645711647   377143762 -2076362548 -1220246257  1766118993  1069253812
#> [199]  1057456592  -852003354   392898140  1210961281   868163719  -806768062
#> [205]   922886097  1232656114  1644144050   585520572  -748565803 -1388244088
#> [211]  1726262453  2095871005  1841042864 -1414099507  -422025967  -272482519
#> [217]   -33024423   313688253   764368501   266129702 -1379115726  -475794467
#> [223]  2091774750  -409915391   445227944   532365993 -1634532467     6837655
#> [229]  1689324633   543364226 -2128909344  1504579714 -2002666302  1520712079
#> [235]   813931602 -1057868484 -1915065838  1836377692  -695662121   414840759
#> [241] -1362245744 -1338127563  -853363590  1289105979  1173588283 -1400451287
#> [247] -1478146060  -915501855 -1614217735 -1802554161  2103649146 -1682470348
#> [253]  1668666006  -930737028 -1346494472   389256448 -1453712159  -415015583
#> [259]  1315376572  1429229610  1074738471   170082856  -690449788  2026071634
#> [265]  1752318066   709023729  1249160394  1293173982  1005479586  -301010068
#> [271]  1225479779  1972806417  -396470858 -1282665797  1666142420   419388687
#> [277]  1219586019    31429041   498841634 -1300948595  1335537921  1086519839
#> [283]  -565823708   915754063   312138655   982840608   509784434 -1933245297
#> [289] -1110917064 -1849748262   -51764539   993823276 -1274114612  1064951256
#> [295]   260879247  1589379751 -1028782000 -1775699847  2070191094  1624136928
#> [301]  2043875468 -1834196749  -360815436  -531171715 -1335954350  2037867453
#> [307] -1159439212  1504511782 -2116380731 -1054289386   824658186  -555750975
#> [313]  1627978250 -1924491058  -374746423  -913722323  1946375627  1909215276
#> [319]  -313926172  -933636373 -1962531064  1420902292   219388363  1783276318
#> [325]  1613650256   -36170230  2094309901   989985409  -891823850  1902909541
#> [331]  -152415136   482264920  -985893378 -1801430775 -1804711014  1493118025
#> [337] -1592221070   -89626163  1590954832 -1530490932  1999448703  1020950440
#> [343]   346214134   209435027   593359940  -366946228  -308022591  -565481885
#> [349]  -636791505  1719914952  1523107662   392998647  1159591240  -125635710
#> [355]   172290153  1734313212   272021908  -104029788   733069843   326074343
#> [361] -1217349668  -427931751 -1881273982 -1774575504  1444119540 -1092780349
#> [367]   331212528 -1109332687  -544209338   574363517 -2091711988    88073154
#> [373] -1379783403  1563488818   -88507918 -1263280623  -335481010 -1973178978
#> [379]  1100219457 -2043000067   623307571   255001444 -2005392548  -498679781
#> [385]  1037417984  1399738288  -159742061  1074755778 -1343090640   219047674
#> [391]  1144533445  -306688555  -994474482 -2100547755 -1983243784  1693427432
#> [397]  -229469326  1043602117  2056356370 -1991111303  1458172586  1581312292
#> [403]  -427918554   979505415 -1915947272  1278540258 -2126082667 -1418962303
#> [409] -1288966338   825083327  -520508483  1722063362  2094421094  2027538600
#> [415]   599438947  1617352391   245848056 -1981652177  -995515657  1535014387
#> [421]  1454654635   480699086 -1703906080   944431149  1538200886 -1322035894
#> [427]   885779094  -965929807   657857201 -1124898057  1400249401  1245874729
#> [433]  1251783926  1360378452 -1555106550 -1070265907  1148412502    71360732
#> [439]  -968255607 -1716302009   743778696  -776920359 -1343888643 -1579333728
#> [445]  -541469422   944061300   216688717  1669523171    54846888 -1620731823
#> [451]  -578935711 -1859446047   145087065  2143533764    70201499   184210148
#> [457]  1318117803  1686067707 -1235411585   238397890  1766530186 -1069169701
#> [463]  1129995978  -277558451   697605791 -2138396878 -1357970144  -359887921
#> [469]   962759284 -1714678082  -916221982  -593474573   603170725  -718178907
#> [475] -1778053900  -484067191   173312019   347740980 -1371166306  -624799490
#> [481]   115302456   735837942  2066403337  1675302333 -1501828250  -823389724
#> [487]  1931313359   447981769   168533884   881899188 -1466868984 -1651566607
#> [493]  1956034975 -1840192516 -1212756685  1025360660 -1830027186   172713759
#> [499]  -370995631  -281036708  1077792903   699092035  -648729727  -393472850
#> [505] -1618825110   942565380  -151149917  1386321074 -1129299798  -967118197
#> [511]  1716019185  -639742389  1788056081   709556557  -592170952    55316732
#> [517]   631779599  1870030395 -1717189568  1445697494  2076263743  -881559043
#> [523]  1219876973 -1800139208 -1186646636   394592321  -597371564  1918919583
#> [529]  1242870197  1157613864  1939364274 -1411655307  1928481994 -1531897728
#> [535]   521498228 -1849616363   162473455   604187195   849795134   -85466913
#> [541]  1144131317   133759782 -1588393680 -1903455500  2103103454 -1427619976
#> [547]  -521219633   691494955 -1841178012 -1235214170 -1467836327 -1254231973
#> [553] -1026004214  -865269458   972452798  -205589849   214083113 -2131618710
#> [559]  -999565419 -1513387834  1922189388  1386664557  -446889853 -1570316318
#> [565]  -588317939   485929229 -1605530133   383018176  -197143998    21117940
#> [571]   299913358 -1793368373 -1176237898 -2128491973 -1261265894 -1625286221
#> [577]  1948329404  1053068770 -1457834866 -1359455188  -550824888  1239947502
#> [583] -2085773898  1259520991 -1722005979 -1000100949 -1528010830  -924186833
#> [589]   296776538  1577279107   358253489 -1887169531   652458182 -2058420675
#> [595]  -592081014  -642585057  -761095889   812168067  1079230012   344291934
#> [601] -1233357053 -1576708981 -1247250521  1845011770  -336625399  1498657682
#> [607] -1619725681  1333125396   263369211  -470997661  1065338289 -1581372653
#> [613]  1593289763  -213361281 -1006044725  -535976050  2026190076 -1383977856
#> [619]  1139512257 -1493508512  1114258995   767580007  -902250648 -1327087205
#> [625]  2066529015  1881339066
#> attr(,"class")
#> [1] "sim"     "numeric"