Simulation of a single-bubble process

The following function generates a time series which switches from a martingale to a mildly explosive process and then back to a martingale.

sim_psy1(
  n,
  te = 0.4 * n,
  tf = 0.15 * n + te,
  c = 1,
  alpha = 0.6,
  sigma = 6.79,
  seed = NULL
)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te
A scalar in (0, tf) specifying the observation in which the bubble originates.

tf
A scalar in (te, n) specifying the observation in which the bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute.

Value

A numeric vector of length n.

Details

The data generating process is described by the following equation: $$X_t = X_{t-1}1\{t < \tau_e\}+ \delta_T X_{t-1}1\{\tau_e \leq t\leq \tau_f\} + \left(\sum_{k=\tau_f+1}^t \epsilon_k + X^*_{\tau_f}\right) 1\{t > \tau_f\} + \epsilon_t 1\{t \leq \tau_f\}$$

where the autoregressive coefficient \(\delta_T\) is given by:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\) and \(X_{\tau_f} = X_{\tau_e} + X^*\). During the pre- and post- bubble periods, \(N_0 = [1, \tau_e)\), X is a pure random walk process. During the bubble expansion period \(B = [\tau_e, \tau_f]\) is a mildly explosive process with expansion rate given by the autoregressive coefficient \(\delta_T\), and continues its martingale path for the subsequent period \(N_1 = (\tau_f, \tau]\).

For further details the user can refer to Phillips et al. (2015) p. 1054.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination date 40 and termination date 55
sim_psy1(n = 100)

#>   [1] 100.00000  98.58853 101.46729  98.02681  93.87457 110.24973 109.12967
#>   [8] 106.13931 109.68353  96.65812  83.19593  86.73488  92.01603  86.50174
#>  [15]  92.68572  99.12776  89.97450  92.37936  95.97985  93.86833  92.21051
#>  [22]  90.22866  82.56896  74.94459  88.89083  82.71520  85.82760  85.51106
#>  [29]  80.88738  70.77822  68.11306  56.23272  55.97017  50.55750  44.33840
#>  [36]  44.83588  53.24307  58.12081  61.78378  58.47630  58.49429  57.32879
#>  [43]  59.73113  64.62006  63.75821  64.12916  64.10371  76.17476  82.26582
#>  [50]  89.14681 101.87830  99.55203 108.38037 111.43724 120.52193  59.98549
#>  [57]  53.45237  53.80567  59.91695  57.23557  63.71422  59.72025  49.69424
#>  [64]  40.33914  31.20203  26.20025  23.00820  32.38253  43.75586  51.74705
#>  [71]  50.73604  49.52861  54.98181  52.15856  60.39137  68.81341  73.46915
#>  [78]  73.28719  75.38924  77.08582  67.87545  71.94524  72.00396  72.61992
#>  [85]  68.15850  64.88720  65.00778  59.17153  68.39034  76.64061  86.51268
#>  [92]  85.92934  81.73604  80.25406  71.25048  55.21108  45.63739  47.36475
#>  [99]  49.37371  49.75931

# 200 periods with bubble origination date 80 and termination date 110
sim_psy1(n = 200)

#>   [1] 100.00000  99.30297 112.55653 118.52199 109.65006 110.26350 116.39051
#>   [8] 120.12108 114.21488 113.94861 110.42832 104.23686 105.25749 108.24751
#>  [15] 117.20883 118.88298 125.27798 122.95477 121.07973 123.68476 133.50928
#>  [22] 145.28247 148.38162 153.18557 167.19870 167.36105 169.03105 170.87833
#>  [29] 164.13954 163.95228 179.04540 180.10127 174.19997 166.22519 154.27698
#>  [36] 154.67326 162.58035 164.87284 157.71266 162.19928 156.60833 159.57579
#>  [43] 162.10431 150.92288 137.86196 140.44782 149.78886 155.39637 152.57095
#>  [50] 159.23974 160.88415 154.51563 164.74041 157.50328 158.60509 160.37369
#>  [57] 150.26880 159.87789 163.69732 168.05688 162.01016 161.21887 154.82286
#>  [64] 162.39934 165.59155 171.43511 170.95712 166.79363 164.50733 163.04178
#>  [71] 167.25927 158.54073 149.71310 147.15483 147.85929 143.08210 153.24954
#>  [78] 151.19335 141.84466 153.74605 164.70508 179.91322 198.57531 205.59947
#>  [85] 204.79293 221.80956 218.30333 224.88672 219.79030 231.23227 230.16656
#>  [92] 238.41485 254.55276 272.47314 293.68009 294.67953 288.19326 301.98514
#>  [99] 321.51292 330.42088 341.55252 354.12399 370.36317 375.72392 396.28508
#> [106] 396.63817 412.70516 426.69582 442.55346 464.87907 152.34950 153.94434
#> [113] 154.69551 156.55734 167.25104 171.10619 170.70142 169.68572 171.49301
#> [120] 175.08086 171.14682 177.54435 172.74816 164.08818 169.06701 166.21381
#> [127] 164.47981 161.97268 170.20521 167.18119 174.91274 176.20640 166.92859
#> [134] 158.11447 143.60357 142.95034 147.66605 144.44358 156.41524 170.17073
#> [141] 173.07426 169.22787 163.56613 166.27031 156.25819 153.57315 152.68568
#> [148] 139.10911 132.34674 135.19932 134.77249 134.22839 134.00919 129.12731
#> [155] 121.54585 116.24782 104.18227 101.27727  87.48660 106.16429 116.44211
#> [162] 116.67290 121.70288 113.07409 111.96349 107.81517 116.31546 118.01513
#> [169] 106.43423 100.62598  99.64210  97.43910  96.26739  87.87452  74.95787
#> [176]  74.31619  74.53725  77.66940  87.04911  84.22124  88.84484  86.03125
#> [183]  82.51168  77.86719  71.85363  72.18795  73.44791  69.31513  64.35094
#> [190]  82.78677  73.69233  69.30589  62.97452  57.75242  60.27545  62.68883
#> [197]  56.00748  57.46550  56.92172  47.26383

# 200 periods with bubble origination date 100 and termination date 150
sim_psy1(n = 200, te = 100, tf = 150)

#>   [1] 100.00000 107.52433 114.84719 111.85789 106.57413  94.22590  86.59339
#>   [8]  93.79435  83.75397  73.21888  78.49815  85.75286  84.50630  95.08677
#>  [15]  93.63885  99.95719 102.74660  94.05646  88.74419  73.28958  72.43135
#>  [22]  82.26427  72.46795  82.43015  77.38314  75.31746  77.61016  73.48522
#>  [29]  71.47837  70.56486  76.09724  74.24411  88.90701  96.31991 101.36750
#>  [36]  93.16613  95.39202  91.76503 100.48303 100.67948  97.99283  93.27467
#>  [43]  83.14344  92.94995  90.59370  90.32034  98.78498  89.64359  85.74680
#>  [50]  80.95334  68.34156  66.66463  64.38617  62.69290  65.81306  62.68915
#>  [57]  63.08930  58.04711  43.35644  41.77507  47.71569  41.10567  37.62411
#>  [64]  31.35363  21.20648  20.07981  30.82321  37.71236  29.92777  23.74990
#>  [71]  44.86125  37.59898  32.84524  31.49780  36.54698  36.20640  45.12264
#>  [78]  45.70668  41.42998  45.87791  46.00959  50.73727  64.37749  63.99061
#>  [85]  65.62303  65.47440  59.83321  69.35601  64.44094  73.47558  67.96663
#>  [92]  80.15492  84.81712  85.42399  87.62767  88.11192  89.69432 103.18182
#>  [99]  96.71997  96.95302 102.72593 105.65723 112.96697 118.15005 112.63448
#> [106] 120.40575 131.35512 128.01149 126.01413 123.55418 128.85100 134.06904
#> [113] 137.75716 143.32731 154.89557 156.24977 166.39496 162.64284 162.71543
#> [120] 182.95695 178.00216 179.23451 173.44534 175.23159 169.83311 171.80648
#> [127] 174.94591 177.18965 189.27458 196.74988 192.28427 191.63470 187.51378
#> [134] 191.86619 188.05973 196.16308 203.48571 207.79575 217.53670 222.09482
#> [141] 229.06647 242.83222 259.92219 263.03836 280.21386 300.12072 307.95656
#> [148] 313.34395 328.86522 335.71067 100.72144 103.68233 104.39782 102.67494
#> [155]  96.04456  98.72130 105.61249 109.90000 106.61772  96.25391 100.44227
#> [162] 108.78913 109.85824 119.40730 122.10158 124.31799 114.40905 108.98183
#> [169] 112.17322 104.36349 102.51935 105.62526 105.51012 101.83269 107.77946
#> [176] 112.80976 111.69068 106.59163  98.08127  91.04523  90.84534  88.97847
#> [183]  92.27469  98.50100  91.81051  81.68533  87.25037  94.13138 101.28536
#> [190] 100.80036 107.26746 104.65985 101.82326  98.63014  87.15381  80.73388
#> [197]  78.40483  80.83879  81.16599  88.83986