Simulation of a single-bubble process

The following function generates a time series which switches from a martingale to a mildly explosive process and then back to a martingale.

sim_psy1(
  n,
  te = 0.4 * n,
  tf = 0.15 * n + te,
  c = 1,
  alpha = 0.6,
  sigma = 6.79,
  seed = NULL
)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te
A scalar in (0, tf) specifying the observation in which the bubble originates.

tf
A scalar in (te, n) specifying the observation in which the bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute. Results between seeds in parallel and non-parallel differ.

Value

A numeric vector of length n.

Details

The data generating process is described by the following equation: $$X_t = X_{t-1}1\{t < \tau_e\}+ \delta_T X_{t-1}1\{\tau_e \leq t\leq \tau_f\} + \left(\sum_{k=\tau_f+1}^t \epsilon_k + X^*_{\tau_f}\right) 1\{t > \tau_f\} + \epsilon_t 1\{t \leq \tau_f\}$$

where the autoregressive coefficient \(\delta_T\) is given by:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\) and \(X_{\tau_f} = X_{\tau_e} + X^*\). During the pre- and post- bubble periods, \(N_0 = [1, \tau_e)\), X is a pure random walk process. During the bubble expansion period \(B = [\tau_e, \tau_f]\) is a mildly explosive process with expansion rate given by the autoregressive coefficient \(\delta_T\), and continues its martingale path for the subsequent period \(N_1 = (\tau_f, \tau]\).

For further details the user can refer to Phillips et al. (2015) p. 1054.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination date 40 and termination date 55
sim_psy1(n = 100)

#>   [1] 100.00000 101.82002 104.78678 111.98502 115.05539 119.55851 111.84253
#>   [8] 109.32685 119.35548 111.04517 112.79746 115.54743 122.17313 119.80426
#>  [15] 120.88132 108.90882 111.20789 106.68191 105.06150  96.99658  99.61029
#>  [22] 104.13636 102.06803 114.45986 119.01296 125.45417 139.36962 134.94856
#>  [29] 140.43908 147.13796 147.09606 149.26243 142.39216 145.58460 140.82501
#>  [36] 146.34991 140.84030 143.00901 137.26112 144.25298 142.81081 152.69363
#>  [43] 169.01911 180.92775 180.35082 187.51676 210.59280 236.16889 243.09164
#>  [50] 255.93977 274.48959 293.97579 308.58650 321.58423 340.65650 151.10958
#>  [57] 151.27000 146.86310 143.43840 154.40012 151.36526 156.54723 166.54020
#>  [64] 169.55268 166.68919 166.41758 163.07500 171.41120 170.39573 180.92012
#>  [71] 177.10677 172.71284 173.68471 173.84895 170.42373 159.68604 159.89019
#>  [78] 155.02464 162.37557 155.90683 163.55565 159.14865 161.13452 167.23671
#>  [85] 163.71445 167.47909 166.88175 159.17365 157.33981 168.33818 166.88423
#>  [92] 161.33149 160.96467 163.20633 169.69299 177.46048 178.14302 186.05049
#>  [99] 180.86116 164.94192

# 200 periods with bubble origination date 80 and termination date 110
sim_psy1(n = 200)

#>   [1] 100.00000  96.79727  93.29461  77.56873  81.38791  76.06608  74.53117
#>   [8]  63.75474  67.47243  80.31386  74.35172  73.58744  86.81920  93.15982
#>  [15] 106.14949 106.11395 105.08010 101.61971 111.36046 102.62962 104.71628
#>  [22] 104.40178 119.69178 115.56324 105.31517 106.89474 106.62552 100.92786
#>  [29] 101.82612  99.95531  95.34654  98.74721  96.49523  84.03571  66.03039
#>  [36]  62.08823  71.96216  77.65306  85.90327  92.57449  94.71852  84.48557
#>  [43]  85.88139  96.72657  73.66730  68.36191  75.84765  79.43783  84.79781
#>  [50]  87.90152  91.56020  91.65962  85.43666  77.10659  77.35207  74.49081
#>  [57]  68.38412  71.21855  72.26044  82.19613  74.58113  71.06535  70.55646
#>  [64]  60.99756  59.06441  54.25593  39.67346  37.74620  34.11985  41.81299
#>  [71]  37.71137  41.49687  42.46532  34.06704  36.60748  35.88715  37.01472
#>  [78]  41.48277  42.25658  54.77885  57.84378  51.73702  58.92798  53.07751
#>  [85]  42.25771  38.51733  35.06436  32.39904  44.65764  57.62235  64.95119
#>  [92]  59.00562  61.61905  67.49166  69.37076  73.42466  70.06600  64.23974
#>  [99]  76.47262  81.47146  82.08991  87.71571 105.37136 112.22023 126.47513
#> [106] 141.02562 144.13146 155.30897 157.35247 153.89656  46.61954  45.60973
#> [113]  57.81178  58.52344  53.07217  54.63641  59.37566  50.58124  43.41996
#> [120]  30.23645  21.61613  15.59042  13.60832  23.18285  29.21673  31.07023
#> [127]  27.22273  26.94972  30.66801  27.24563  34.43671  41.88842  33.96786
#> [134]  28.89633  37.09008  25.63386  28.48369  30.06915  51.76932  33.23479
#> [141]  27.52578  32.10244  43.47111  49.17259  39.47397  54.67941  42.74758
#> [148]  35.19020  34.89727  50.00901  53.44690  58.40629  70.13993  76.82659
#> [155]  68.51062  73.32966  72.58806  84.69197  83.03898  72.66992  76.00947
#> [162]  78.41655  78.29687  71.12290  65.42666  65.34094  72.38970  69.91504
#> [169]  63.99789  45.63636  48.99231  51.89395  60.78336  70.74241  69.77105
#> [176]  78.58796  78.59688  74.00406  68.24480  65.88907  67.43366  67.24760
#> [183]  68.61156  69.96238  74.81451  72.95961  84.06362  87.03556  78.93145
#> [190]  77.05300  80.55584  77.76478  80.10730  83.05684  91.90746  87.48244
#> [197]  87.36782  83.51319  75.85996  63.30545

# 200 periods with bubble origination date 100 and termination date 150
sim_psy1(n = 200, te = 100, tf = 150)

#>   [1]  100.000000  101.327642  105.412270  109.752154  100.314103   92.643590
#>   [7]   84.895153   80.219074   81.374717   85.411259   79.877188   60.428273
#>  [13]   66.904734   71.082518   66.323606   59.033445   56.996505   57.730548
#>  [19]   62.641082   72.079827   66.315909   68.289917   77.348972   71.228859
#>  [25]   69.716296   57.688028   59.639625   55.285933   49.005606   43.709559
#>  [31]   46.727737   32.044573   29.155354   22.272211   15.739422   18.962192
#>  [37]   21.350603   20.623935   24.273677   20.185385    6.447194   14.285714
#>  [43]   12.719309   10.240356   11.309903   16.510583   19.403713   25.699845
#>  [49]   26.013099   22.588999   20.548948   26.136548   20.370012   13.895000
#>  [55]   16.023520   20.182214    8.699759   14.027863   14.108834   12.878111
#>  [61]   20.394331   30.439422   22.648580   29.517713   25.225964   26.124697
#>  [67]   29.390185   30.294461   20.361611   16.923143   27.377457   28.550725
#>  [73]   22.875481   18.421725   25.602599   27.952506   21.860609   21.903150
#>  [79]   25.954543   37.650669   31.663590   29.083757   35.847344   32.729913
#>  [85]   42.645118   49.728759   38.463871   41.342544   31.706428   21.761166
#>  [91]   22.239325   22.624306   22.402217   30.986687   32.989655   28.419395
#>  [97]   29.231399   25.738647   25.276170   18.533357   18.522347   19.992327
#> [103]   16.059722   17.008207   13.311082    9.492378    8.499383   12.186003
#> [109]   11.299039   13.064186   20.194482   19.098208   19.189647   23.459885
#> [115]   15.893542   20.247732   17.458195   10.457606   -1.111814   -6.012798
#> [121]  -14.208770  -22.801911  -12.701793  -21.461015  -15.581294  -12.527689
#> [127]  -19.056906  -24.000457  -27.784794  -27.648530  -31.164600  -26.847615
#> [133]  -29.952562  -45.902118  -48.466228  -63.058551  -59.315793  -66.861762
#> [139]  -77.217550  -84.448264  -83.989932 -100.513246 -102.973962 -113.862048
#> [145] -118.659665 -118.988495 -120.073911 -114.867197 -118.772360 -128.477607
#> [151]   33.157029   31.097987   36.581454   42.902886   45.991671   55.670278
#> [157]   54.251228   45.890003   43.441505   39.725926   41.953019   35.615913
#> [163]   40.238619   32.335857   44.255359   54.098397   51.051079   52.002136
#> [169]   49.465051   41.273363   23.202824   23.032432   22.799038   33.950656
#> [175]   20.401468   15.905328   21.880786   24.205597   25.032305   20.817338
#> [181]   22.181756   20.579959   21.351718   17.427579   25.497934   32.694902
#> [187]   41.771490   39.278411   37.211965   34.833566   44.164009   42.128907
#> [193]   42.791850   45.380117   49.985180   48.167678   64.096785   57.799096
#> [199]   48.112773   48.400008