Simulation of a single-bubble process

The following function generates a time series which switches from a martingale to a mildly explosive process and then back to a martingale.

sim_psy1(n, te = 0.4 * n, tf = 0.15 * n + te, c = 1, alpha = 0.6,
  sigma = 6.79, seed = NULL)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te
A scalar in (0, tf) specifying the observation in which the bubble originates.

tf
A scalar in (te, n) specifying the observation in which the bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute.

Value

A numeric vector of length n.

Details

The data generating process is described by the following equation: $$X_t = X_{t-1}1\{t < \tau_e\}+ \delta_T X_{t-1}1\{\tau_e \leq t\leq \tau_f\} + \left(\sum_{k=\tau_f+1}^t \epsilon_k + X^*_{\tau_f}\right) 1\{t > \tau_f\} + \epsilon_t 1\{t \leq \tau_f\}$$

where the autoregressive coefficient \(\delta_T\) is given by:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\) and \(X_{\tau_f} = X_{\tau_e} + X^*\). During the pre- and post- bubble periods, \(N_0 = [1, \tau_e)\), X is a pure random walk process. During the bubble expansion period \(B = [\tau_e, \tau_f]\) is a mildly explosive process with expansion rate given by the autoregressive coefficient \(\delta_T\), and continues its martingale path for the subsequent period \(N_1 = (\tau_f, \tau]\).

For further details the user can refer to Phillips et al. (2015) p. 1054.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination date 40 and termination date 55
sim_psy1(n = 100)

#>   [1] 100.00000 104.92065  88.77733  88.33320  85.14390  83.23920  87.14234
#>   [8]  85.74579  87.34063  88.09180  89.95363 100.64734 104.50248 104.09771
#>  [15] 103.08201 104.88930 108.47716 104.54311 110.94064 106.14445  97.48447
#>  [22] 102.46331  99.61010  97.87610  95.36897 103.60150 100.57748 108.30904
#>  [29] 109.60269 100.32488  91.51076  76.99986  76.34663  81.06235  77.83987
#>  [36]  89.81153 103.56702 106.47055 102.62416 103.43756 112.66821 109.76498
#>  [43] 114.00564 120.31144 114.32601 114.77713 124.87165 132.32369 140.12865
#>  [50] 148.75097 153.25464 155.34290 159.84634 157.86642 164.92212  89.64689
#>  [57] 108.32458 118.60240 118.83319 123.86317 115.23438 114.12378 109.97546
#>  [64] 118.47575 120.17542 108.59452 102.78627 101.80239  99.59939  98.42768
#>  [71]  90.03481  77.11816  76.47648  76.69753  79.82969  89.20940  86.38153
#>  [78]  91.00513  88.19154  84.67197  80.02748  74.01392  74.34824  75.60820
#>  [85]  71.47542  66.51123  84.94706  75.85262  71.46618  65.13481  59.91271
#>  [92]  62.43573  64.84912  58.16777  59.62579  59.08201  49.42412  56.94845
#>  [99]  64.27132  61.28202

# 200 periods with bubble origination date 80 and termination date 110
sim_psy1(n = 200)

#>   [1] 100.000000  94.716232  82.368006  74.735497  81.936454  71.896080
#>   [7]  61.360986  66.640256  73.894965  72.648405  83.228874  81.780958
#>  [13]  88.099293  90.888705  82.198562  76.886293  61.431686  60.573458
#>  [19]  70.406374  60.610051  70.572252  65.525246  63.459566  65.752264
#>  [25]  61.627325  59.620471  58.706964  64.239350  62.386211  77.049120
#>  [31]  84.462018  89.509601  81.308236  83.534120  79.907139  88.625137
#>  [37]  88.821581  86.134934  81.416778  71.285550  81.092057  78.735809
#>  [43]  78.462449  86.927081  77.785694  73.888904  69.095441  56.483664
#>  [49]  54.806734  52.528275  50.835007  53.955165  50.831252  51.231406
#>  [55]  46.189212  31.498545  29.917178  35.857795  29.247771  25.766220
#>  [61]  19.495736   9.348585   8.221918  18.965316  25.854466  18.069878
#>  [67]  11.892002  33.003353  25.741090  20.987346  19.639908  24.689089
#>  [73]  24.348504  33.264748  33.848780  29.572089  34.020017  34.151699
#>  [79]  38.879372  54.138056  56.004814  59.968585  62.316302  59.269197
#>  [85]  71.259235  69.310519  81.230395  79.102875  94.584038 103.183543
#>  [91] 108.085709 114.788743 120.051380 126.631237 145.390096 144.980501
#>  [97] 147.222518 155.088028 160.199042 169.779232 177.327273 174.275114
#> [103] 184.612333 198.234478 197.674878 198.577447 199.138130 207.581336
#> [109] 216.076739 223.178643  53.973695  59.575582  54.481841  58.122710
#> [115]  47.443959  40.746116  54.214165  41.643306  35.465844  22.215566
#> [121]  16.781689   4.088724  -1.007657  -5.020128 -10.058982  -5.350035
#> [127]  -5.753796 -18.409643 -27.063559 -39.161783 -42.615132 -54.408536
#> [133] -54.133667 -54.976846 -59.137442 -58.046544 -62.543965 -64.817608
#> [139] -60.587360 -53.605920 -61.309707 -55.083881 -46.841671 -51.499152
#> [145] -58.931270 -56.453775 -63.298211 -59.529785 -56.568898 -55.853413
#> [151] -57.576293 -64.206668 -61.529930 -54.638738 -50.351235 -53.633509
#> [157] -63.997316 -59.808957 -51.462100 -50.392994 -40.843931 -38.149653
#> [163] -35.933241 -45.842182 -51.269399 -48.078012 -55.887737 -57.731879
#> [169] -54.625969 -54.741113 -58.418539 -52.471766 -47.441475 -48.560551
#> [175] -53.659598 -62.169955 -69.205999 -69.405887 -71.272756 -67.976540
#> [181] -61.750232 -68.440717 -78.565896 -73.000863 -66.119846 -58.965867
#> [187] -59.450867 -52.983768 -55.591383 -58.427973 -61.621094 -73.097418
#> [193] -79.517347 -81.846399 -79.412439 -79.085236 -71.411367 -74.993990
#> [199] -71.737674 -62.112650

# 200 periods with bubble origination date 100 and termination date 150
sim_psy1(n = 200, te = 100, tf = 150)

#>   [1]  100.00000   99.45618   96.77965  100.34615   94.94175   96.60596
#>   [7]   91.31438   95.35391  102.88328   96.48052  101.23586   98.35231
#>  [13]   90.59041   92.17547   90.89055   79.59599   92.60353   87.05682
#>  [19]   89.66180   86.54769   81.68806   84.60469   76.66139   85.63837
#>  [25]   94.91327   93.21765   90.33379   97.79220  102.34776  102.55794
#>  [31]  118.64014  136.19764  144.42903  135.93340  138.65121  144.99787
#>  [37]  141.78845  152.53711  156.59586  159.76057  154.70632  161.21830
#>  [43]  162.25471  157.03069  166.48971  172.72747  179.62670  183.25688
#>  [49]  176.55088  184.61599  181.10188  179.33946  177.11191  177.61051
#>  [55]  175.92752  166.59900  166.32437  169.18661  170.55546  159.03153
#>  [61]  163.39266  156.63498  163.17928  151.93479  159.20594  158.46565
#>  [67]  171.36112  163.64068  161.74139  155.67025  156.59846  151.51162
#>  [73]  155.03019  153.72422  153.91984  156.35467  156.15780  163.94621
#>  [79]  166.48288  168.67836  163.04388  172.51228  171.21170  173.06046
#>  [85]  165.71600  149.89938  146.16742  145.67460  152.68384  154.14463
#>  [91]  150.78806  161.05548  157.12092  168.49274  161.69609  169.99824
#>  [97]  177.31235  173.15724  176.59886  187.07413  204.94061  219.46033
#> [103]  224.95175  243.04487  257.15375  258.97387  271.90492  291.33202
#> [109]  309.65914  317.20748  327.62187  344.49458  360.09745  366.09264
#> [115]  389.29262  407.06980  427.29522  441.44770  469.18168  479.86747
#> [121]  504.15391  531.92832  551.77794  574.16299  586.39328  604.49473
#> [127]  635.18329  665.19883  699.76069  734.53516  767.93091  795.40814
#> [133]  832.24555  863.22767  897.99857  930.17402  964.77635 1016.12298
#> [139] 1064.13266 1110.72623 1154.21400 1208.37229 1266.75863 1319.16804
#> [145] 1379.42514 1444.76917 1503.06585 1563.50491 1626.64598 1688.90409
#> [151]  186.43210  185.92932  190.78071  191.62755  200.43142  192.86402
#> [157]  187.14644  176.93228  175.00374  175.29482  175.23462  155.21034
#> [163]  155.34857  154.68012  158.68017  155.78931  160.09865  156.17046
#> [169]  155.02221  141.99063  131.57296  124.00373  134.85287  130.50859
#> [175]  141.14642  131.30665  125.93234  122.50692  125.23532  131.83110
#> [181]  127.89519  138.78757  140.32193  136.82605  131.23253  133.50321
#> [187]  132.86841  134.93299  131.69750  130.05900  135.65501  125.09219
#> [193]  125.72706  123.23548  122.35679  125.12572  121.16057  119.84705
#> [199]  118.01683  118.51697