Simulation of a single-bubble process

The following function generates a time series which switches from a martingale to a mildly explosive process and then back to a martingale.

sim_psy1(n, te = 0.4 * n, tf = 0.15 * n + te, c = 1, alpha = 0.6,
  sigma = 6.79)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te
A scalar in (0, tf) specifying the observation in which the bubble originates.

tf
A scalar in (te, n) specifying the observation in which the bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

Value

A numeric vector of length n.

Details

The data generating process is described by the following equation: $$X_t = X_{t-1}1\{t < \tau_e\}+ \delta_T X_{t-1}1\{\tau_e \leq t\leq \tau_f\} + \left(\sum_{k=\tau_f+1}^t \epsilon_k + X^*_{\tau_f}\right) 1\{t > \tau_f\} + \epsilon_t 1\{t \leq \tau_f\}$$

where the autoregressive coefficient \(\delta_T\) is given by:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\) and \(X_{\tau_f} = X_{\tau_e} + X^*\). During the pre- and post- bubble periods, \(N_0 = [1, \tau_e)\), X is a pure random walk process. During the bubble expansion period \(B = [\tau_e, \tau_f]\) is a mildly explosive process with expansion rate given by the autoregressive coefficient \(\delta_T\), and continues its martingale path for the subsequent period \(N_1 = (\tau_f, \tau]\).

For further details the user can refer to Phillips et al. (2015) p. 1054.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination date 40 and termination date 55
sim_psy1(n = 100)

#>   [1]  100.000000   98.655063  101.566541  102.047070   91.613191   94.695745
#>   [7]  100.632911   91.821273   84.495099   76.789468   76.943019   76.797297
#>  [13]   74.904432   74.740071   80.341959   75.248218   78.889087   68.210335
#>  [19]   61.512493   74.980541   62.409682   56.232220   42.981942   37.548065
#>  [25]   24.855100   19.758720   15.746248   10.707394   15.416341   15.012581
#>  [31]    2.356733   -6.297183  -18.395407  -21.848756  -33.642159  -33.367291
#>  [37]  -34.210470  -38.371065  -37.280167  -44.129809  -49.187854  -48.061149
#>  [43]  -44.112164  -54.599239  -51.818392  -46.845702  -54.458947  -65.327193
#>  [49]  -66.971565  -78.041621  -79.197288  -81.233412  -85.643409  -92.770023
#>  [55] -105.253791  -41.453070  -34.561878  -30.274375  -33.556649  -43.920457
#>  [61]  -39.732098  -31.385240  -30.316134  -20.767071  -18.072793  -15.856381
#>  [67]  -25.765322  -31.192539  -28.001152  -35.810878  -37.655019  -34.549109
#>  [73]  -34.664253  -38.341679  -32.394906  -27.364615  -28.483691  -33.582738
#>  [79]  -42.093096  -49.129140  -49.329028  -51.195896  -47.899680  -41.673372
#>  [85]  -48.363858  -58.489036  -52.924003  -46.042986  -38.889007  -39.374007
#>  [91]  -32.906908  -35.514523  -38.351113  -41.544234  -53.020559  -59.440488
#>  [97]  -61.769539  -59.335579  -59.008377  -51.334507

# 200 periods with bubble origination date 80 and termination date 110
sim_psy1(n = 200)

#>   [1] 100.00000  96.41738  99.67369 109.29872 108.75490 106.07837 109.64487
#>   [8] 104.24047 105.90468 100.61309 104.65263 112.18200 105.77924 110.53457
#>  [15] 107.65103  99.88913 101.47419 100.18927  88.89471 101.90225  96.35553
#>  [22]  98.96052  95.84641  90.98678  93.90341  85.96011  94.93709 104.21199
#>  [29] 102.51637  99.63251 107.09091 111.64648 111.85666 127.93886 145.49636
#>  [36] 153.72775 145.23212 147.94993 154.29659 151.08716 161.83582 165.89457
#>  [43] 169.05929 164.00503 170.51702 171.55342 166.32941 175.78843 182.02619
#>  [50] 188.92542 192.55560 185.84960 193.91471 190.40060 188.63818 186.41062
#>  [57] 186.90923 185.22624 175.89771 175.62309 178.48533 179.85418 168.33025
#>  [64] 172.69138 165.93370 172.47800 161.23351 168.50465 167.76436 180.65984
#>  [71] 172.93940 171.04010 164.96897 165.89717 160.81034 164.32891 163.02294
#>  [78] 163.21856 165.65339 172.35228 187.31532 197.64948 208.07264 211.09975
#>  [85] 229.35574 237.60270 249.34230 252.37738 247.06663 253.61948 263.68425
#>  [92] 281.67004 294.85610 303.77370 326.68651 336.35114 361.72448 369.98557
#>  [99] 393.68935 417.39183 430.61176 451.97875 473.91743 503.72452 530.68193
#> [106] 549.12877 580.71662 608.88199 625.34373 653.52590 180.46061 186.66027
#> [113] 181.31822 178.52801 181.76258 183.02495 174.03013 181.99053 183.56236
#> [120] 186.84242 183.20760 192.56514 183.72000 188.03069 194.81835 192.52503
#> [127] 191.94086 180.27009 173.96135 179.48622 183.06056 189.93175 195.57682
#> [134] 198.39558 193.90565 197.63208 193.96977 192.80652 187.60039 183.48175
#> [141] 194.66700 200.37786 202.67407 199.92491 206.03598 214.12060 213.79781
#> [148] 219.14103 227.06282 225.21714 223.08708 221.14311 215.68775 215.04571
#> [155] 214.54294 219.39433 220.24117 229.04504 221.47764 215.76006 205.54590
#> [162] 203.61736 203.90844 203.84824 183.82396 183.96219 183.29374 187.29379
#> [169] 184.40293 188.71227 184.78408 183.63583 170.60425 160.18658 152.61735
#> [176] 163.46649 159.12221 169.76004 159.92027 154.54596 151.12054 153.84894
#> [183] 160.44472 156.50881 167.40119 168.93555 165.43967 159.84615 162.11683
#> [190] 161.48203 163.54661 160.31112 158.67262 164.26863 153.70581 154.34068
#> [197] 151.84910 150.97041 153.73934 149.77419

# 200 periods with bubble origination date 100 and termination date 150
sim_psy1(n = 200, te = 100, tf = 150)

#>   [1]  100.00000   98.68648   96.85625   97.35639   99.78203  103.51944
#>   [7]  103.78018   92.85117   85.72364   99.65695  100.85495  108.51616
#>  [13]  111.46982  115.19640  119.59237  125.55713  127.93905  128.27774
#>  [19]  133.95247  132.03956  126.66406  126.67529  118.61196  121.07277
#>  [25]  117.34210  122.04723  121.63410  113.52996  112.71579  107.90773
#>  [31]   96.93058  100.29739  109.12270   98.15016   89.66016  100.40284
#>  [37]  108.90751  107.48571  103.14668  105.85523  114.96723  114.71187
#>  [43]  118.40339  130.21947  128.01642  124.97782  127.80302  119.00409
#>  [49]  123.63917  127.01321  136.78851  144.98647  146.60599  144.57688
#>  [55]  153.97845  149.22499  161.86071  174.10936  164.90183  162.76467
#>  [61]  155.84386  151.68668  149.72390  159.95632  162.45316  162.31712
#>  [67]  155.66345  146.15126  155.96291  153.08747  148.10531  143.29470
#>  [73]  153.31057  159.04815  167.83437  169.85889  167.10430  166.16180
#>  [79]  164.65040  176.51355  175.95877  176.80169  179.46498  182.03445
#>  [85]  183.16278  190.99428  191.09442  191.24000  185.62907  185.95091
#>  [91]  183.96071  192.16054  192.37783  193.60273  194.12665  197.91815
#>  [97]  196.78521  195.23722  196.88492  198.77638  217.34059  220.87721
#> [103]  238.97519  254.46086  265.86849  282.92333  282.48730  287.74821
#> [109]  296.50764  296.21802  310.27144  321.73153  319.03093  332.33356
#> [115]  350.71540  366.85498  378.10751  391.09447  416.97741  435.97584
#> [121]  444.37346  460.73587  476.65234  502.17243  522.72124  550.94470
#> [127]  584.02698  606.00144  635.61267  651.65248  682.85103  707.85192
#> [133]  734.55523  762.00329  808.76293  852.52635  885.60391  928.41895
#> [139]  965.60907 1019.52634 1049.48232 1094.34406 1134.71661 1174.35764
#> [145] 1231.37767 1289.29934 1351.06184 1411.60032 1473.97983 1529.76953
#> [151]  197.69456  197.57552  194.92291  192.18807  200.85691  206.43201
#> [157]  209.28235  220.90453  205.86038  204.60506  209.16229  203.76329
#> [163]  193.12840  201.36964  211.58052  203.64012  207.80495  217.88084
#> [169]  214.86762  216.69134  208.35175  211.60096  222.93931  222.69178
#> [175]  219.69782  224.68446  220.77810  221.28666  221.68932  227.44904
#> [181]  244.93598  245.57600  242.04555  247.30392  255.64400  268.41332
#> [187]  267.56293  262.40868  258.95050  257.89100  258.79123  252.17720
#> [193]  243.68325  243.36863  251.09149  259.67711  255.95115  260.58459
#> [199]  254.27143  247.65017