Simulation of a single-bubble process

The following function generates a time series which switches from a martingale to a mildly explosive process and then back to a martingale.

sim_psy1(
  n,
  te = 0.4 * n,
  tf = 0.15 * n + te,
  c = 1,
  alpha = 0.6,
  sigma = 6.79,
  seed = NULL
)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te
A scalar in (0, tf) specifying the observation in which the bubble originates.

tf
A scalar in (te, n) specifying the observation in which the bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute. Results between seeds in parallel and non-parallel differ.

Value

A numeric vector of length n.

Details

The data generating process is described by the following equation: $$X_t = X_{t-1}1\{t < \tau_e\}+ \delta_T X_{t-1}1\{\tau_e \leq t\leq \tau_f\} + \left(\sum_{k=\tau_f+1}^t \epsilon_k + X^*_{\tau_f}\right) 1\{t > \tau_f\} + \epsilon_t 1\{t \leq \tau_f\}$$

where the autoregressive coefficient \(\delta_T\) is given by:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\) and \(X_{\tau_f} = X_{\tau_e} + X^*\). During the pre- and post- bubble periods, \(N_0 = [1, \tau_e)\), X is a pure random walk process. During the bubble expansion period \(B = [\tau_e, \tau_f]\) is a mildly explosive process with expansion rate given by the autoregressive coefficient \(\delta_T\), and continues its martingale path for the subsequent period \(N_1 = (\tau_f, \tau]\).

For further details the user can refer to Phillips et al. (2015) p. 1054.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination date 40 and termination date 55
sim_psy1(n = 100)

#>   [1] 100.00000  98.35234 103.24342 104.45529  99.98029 111.24607 113.55288
#>   [8] 106.43048 108.52497 107.78248 106.26776 107.31782 109.30632 109.36323
#>  [15] 111.24364  96.30337  96.99560  88.69684  90.52777  89.81514 108.20044
#>  [22] 110.42097 116.14204 105.66474 110.96702 108.76453  97.98062  90.89666
#>  [29]  98.26049 121.42593 119.15921 118.20943 110.49401 114.56020 126.80924
#>  [36] 134.28770 125.60516 121.98823 121.31205 115.53253 130.02311 125.14117
#>  [43] 125.84637 122.30973 133.45534 143.45405 139.75780 149.42145 149.66111
#>  [50] 149.80292 155.49959 162.24700 160.01494 174.64509 180.34641 112.16399
#>  [57] 126.53872 122.62274 108.82420 114.57817 118.38564 114.57163 108.37093
#>  [64]  95.96705 108.23781 111.58154 114.49967 103.63652 107.15543 117.49912
#>  [71] 106.05644 115.63261 104.23219  96.18788  93.75470  92.40615  94.71333
#>  [78]  83.31352  73.74011  76.60487  71.71051  71.45690  65.86230  62.22831
#>  [85]  55.24713  46.74825  49.62288  59.82497  62.70346  62.84662  67.90720
#>  [92]  57.50980  53.02088  54.71445  51.83809  57.39358  73.27259  80.72583
#>  [99]  80.99939  91.05165

# 200 periods with bubble origination date 80 and termination date 110
sim_psy1(n = 200)

#>   [1] 100.00000  93.58305 101.04414  89.65078  84.48234  95.88880  99.84286
#>   [8] 100.14104 105.93285 103.50671 103.02262 105.78687 104.96686 107.02758
#>  [15] 111.07763 119.89831 132.50188 128.42687 118.97110 119.91188 128.51295
#>  [22] 136.68639 137.32941 140.96470 131.47673 130.36763 129.33879 125.73527
#>  [29] 119.65083 121.72016 126.03151 133.62512 128.85372 137.82133 143.49231
#>  [36] 145.10123 145.96389 131.00558 131.81569 126.89854 117.80367 115.46779
#>  [43] 117.95210 122.65616 124.76334 125.12886 116.29327 116.35893 119.97959
#>  [50] 120.66322 124.19009 128.79648 126.81595 128.36620 119.94768 124.11948
#>  [57] 123.83400 118.21135 105.37775 100.05186  97.33024 103.47682 105.44290
#>  [64] 102.96764  99.96835 100.78409 106.74071  95.88925  97.16405  87.83903
#>  [71]  95.95182  92.00069 100.49748 101.59793  95.39386  98.04202 110.53772
#>  [78] 108.43140 112.18881 107.12173 122.50476 134.90652 140.70262 157.15573
#>  [85] 156.49446 161.58116 168.48437 184.96143 199.68691 207.58400 212.70250
#>  [92] 225.00051 226.43310 222.26349 226.61616 248.87622 256.10375 262.76959
#>  [99] 278.07564 282.73229 290.10839 307.60412 328.41175 335.45098 343.04638
#> [106] 347.73806 358.48228 372.60643 388.46034 421.07744 110.75739 110.16538
#> [113]  97.08655  99.07194 112.19093 113.10447 126.85328 133.62073 132.41190
#> [120] 145.51136 145.69346 154.24536 148.15901 148.67885 139.87125 145.65598
#> [127] 133.43001 130.88000 132.64932 126.97682 127.49618 124.77654 123.10878
#> [134] 123.56577 131.28549 128.23507 130.36600 138.12843 141.50152 138.85305
#> [141] 137.58602 139.65090 141.64925 137.88848 131.51012 123.63125 129.06650
#> [148] 129.69774 124.10964 116.71880 120.49823 131.24990 131.14913 129.69222
#> [155] 135.30755 121.78328 121.44470 113.68964 118.94511 113.77165 106.10047
#> [162] 106.43381 105.70386 111.24937 120.39747 135.19624 139.01792 145.35598
#> [169] 133.88640 130.77127 129.35344 133.81710 118.95404 106.47205  91.34912
#> [176]  86.05474  82.27062  86.89461  77.85329  80.63548  88.91439  96.72223
#> [183]  93.12379  87.88836  92.15285  98.12499 111.60009 116.55811 122.13523
#> [190] 119.14598 111.96979 110.30117 125.01655 128.34662 119.44282 116.45322
#> [197] 119.99705 132.32879 115.66780 117.54548

# 200 periods with bubble origination date 100 and termination date 150
sim_psy1(n = 200, te = 100, tf = 150)

#>   [1]  100.0000  112.0929  113.9836  105.8560  103.5160  104.4421  101.4614
#>   [8]  107.4552  111.2377  106.6700  110.4554  119.2272  122.8115  133.7864
#>  [15]  133.6460  137.1079  135.9737  138.6932  143.7370  144.0395  140.6136
#>  [22]  143.8829  148.0477  151.7543  156.2889  172.0066  182.4293  190.4450
#>  [29]  199.5308  205.7022  215.0517  219.3679  211.0849  213.9203  209.4428
#>  [36]  206.9492  204.3942  195.1881  192.5766  196.4599  189.8971  187.2517
#>  [43]  195.1267  192.0732  193.3219  195.9017  189.3226  190.6345  201.1105
#>  [50]  198.9675  192.1900  194.3499  187.2988  185.9839  189.7594  193.8655
#>  [57]  197.2022  203.8355  209.4606  199.0074  198.5809  194.8907  197.6174
#>  [64]  194.7709  198.1236  212.1785  215.3988  210.2300  219.2869  224.1457
#>  [71]  221.1412  210.0564  216.8328  220.9467  216.8018  213.7885  217.2589
#>  [78]  212.8377  214.0833  210.2179  209.3889  219.5448  218.4109  215.8136
#>  [85]  217.3390  208.9976  212.4182  215.1801  215.2567  216.9399  210.6122
#>  [92]  219.3909  213.5753  217.9755  220.8442  229.8256  214.7849  205.8878
#>  [99]  221.7479  234.2410  235.7574  252.6826  269.4179  280.5315  297.1455
#> [106]  304.3233  317.6172  327.7455  336.9884  355.4367  365.2678  386.5008
#> [113]  388.3937  410.0219  427.9355  464.3171  491.4409  504.6710  530.9497
#> [120]  552.9986  577.6788  610.8542  638.3499  666.1446  688.4481  724.1546
#> [127]  753.3008  783.5138  812.6233  838.7701  867.7824  906.9652  940.3848
#> [134]  986.0261 1032.0867 1080.8022 1125.9837 1191.7604 1239.4374 1285.5658
#> [141] 1346.0863 1412.7403 1472.6860 1520.0212 1594.2502 1659.3930 1732.1404
#> [148] 1804.7981 1861.0461 1946.8539  242.4850  251.2600  250.3653  254.9050
#> [155]  259.3712  264.0903  266.4513  266.9799  264.1009  269.3760  260.5674
#> [162]  260.9817  261.8056  277.5861  280.2782  278.2036  278.7271  284.7781
#> [169]  286.4030  288.7704  287.5584  286.0381  270.0858  269.7870  266.7911
#> [176]  264.5035  270.5802  271.9736  271.4551  260.3638  264.8608  268.4597
#> [183]  278.8366  267.9060  277.3486  273.2184  272.9659  279.8527  288.7576
#> [190]  280.8834  285.5503  290.4302  293.2804  277.9358  270.1375  276.9339
#> [197]  271.3657  268.8871  272.4040  262.1007