Simulation of a single-bubble process

The following function generates a time series which switches from a martingale to a mildly explosive process and then back to a martingale.

sim_psy1(
  n,
  te = 0.4 * n,
  tf = 0.15 * n + te,
  c = 1,
  alpha = 0.6,
  sigma = 6.79,
  seed = NULL
)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te
A scalar in (0, tf) specifying the observation in which the bubble originates.

tf
A scalar in (te, n) specifying the observation in which the bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute. Results between seeds in parallel and non-parallel differ.

Value

A numeric vector of length n.

Details

The data generating process is described by the following equation: $$X_t = X_{t-1}1\{t < \tau_e\}+ \delta_T X_{t-1}1\{\tau_e \leq t\leq \tau_f\} + \left(\sum_{k=\tau_f+1}^t \epsilon_k + X^*_{\tau_f}\right) 1\{t > \tau_f\} + \epsilon_t 1\{t \leq \tau_f\}$$

where the autoregressive coefficient \(\delta_T\) is given by:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\) and \(X_{\tau_f} = X_{\tau_e} + X^*\). During the pre- and post- bubble periods, \(N_0 = [1, \tau_e)\), X is a pure random walk process. During the bubble expansion period \(B = [\tau_e, \tau_f]\) is a mildly explosive process with expansion rate given by the autoregressive coefficient \(\delta_T\), and continues its martingale path for the subsequent period \(N_1 = (\tau_f, \tau]\).

For further details the user can refer to Phillips et al. (2015) p. 1054.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination date 40 and termination date 55
sim_psy1(n = 100)

#>   [1] 100.00000  98.79950 108.45852  99.51807  99.10159  99.75069  96.06412
#>   [8]  99.44862  99.16427 102.53124 108.64856 119.64415 122.21734 121.07039
#>  [15] 117.38724 107.51562 108.74020 112.12087 103.34032 108.46509 106.64991
#>  [22] 111.00963 108.17599  99.91212 114.39088 106.99329 107.78425 116.34767
#>  [29] 123.41349 116.83096 119.48399 123.61025 116.83634 124.96475 112.92825
#>  [36] 118.25734 115.28779 112.89947 121.75340 122.08058 121.46711 129.70787
#>  [43] 141.83975 153.84872 166.93054 182.75588 204.52365 205.04112 223.33021
#>  [50] 230.81492 251.44116 272.48766 294.10385 309.02208 327.94937 108.42001
#>  [57] 104.93398 105.62330 105.24163  88.72760  84.47988  78.53209  80.47077
#>  [64]  82.80550  76.45617  78.43701  78.76155  81.77475  86.30866  84.18701
#>  [71] 101.32336  98.47352 107.55307 106.03737 110.06467 109.62151 104.86046
#>  [78] 101.26741 102.42018 103.19500  97.78902  96.18625 104.23808 105.51430
#>  [85] 112.56782 109.51372 104.70205 104.90865 101.85401 106.40686 118.41986
#>  [92] 122.72814 125.23005 134.60126 128.71365 145.25609 145.55379 143.39274
#>  [99] 161.64868 165.52195

# 200 periods with bubble origination date 80 and termination date 110
sim_psy1(n = 200)

#>   [1] 100.00000  96.30127  94.71418  84.76096  87.42280  93.57406 101.73316
#>   [8] 109.74602 102.49982 104.69862 103.39205 108.38756 117.03417 115.30175
#>  [15] 115.23351 106.82334  99.60294  92.52065  84.15446  85.71073  89.93290
#>  [22]  90.71283 100.90651  97.00856  89.03180  89.39840  97.65023  94.65160
#>  [29]  99.71683 101.16261  92.15812 103.48981 102.63274  88.53071  99.88854
#>  [36]  99.95683 109.61163 111.43130 114.17377 108.43878 111.26860 101.32569
#>  [43]  98.06235 104.15189 100.62885  94.49895 101.68238 112.55457 119.07592
#>  [50] 118.11507 111.42379 113.37531 111.58712 118.29429 120.29839 130.81027
#>  [57] 128.18345 128.67673 126.07499 125.63286 131.31688 140.41408 136.64786
#>  [64] 146.14394 149.16188 150.30991 151.46859 156.25563 154.99759 158.75323
#>  [71] 163.83869 150.13648 139.52963 143.91569 147.21870 146.55566 125.48650
#>  [78] 120.24706 124.44027 120.91235 127.96966 138.78587 150.08631 158.16460
#>  [85] 158.92679 176.14937 176.05230 183.20542 180.54782 185.97873 180.94465
#>  [92] 191.64275 202.22824 212.36067 225.40366 241.33964 255.20248 257.92984
#>  [99] 265.38021 283.40108 294.69146 303.64925 313.58410 322.55862 334.56119
#> [106] 351.78478 376.13146 388.53003 413.49803 431.83184 123.12248 119.27443
#> [113] 112.14105 105.71562 105.20515 106.48117 114.96261 111.23676  99.62361
#> [120]  94.86391 107.75016 115.31604 121.06691 123.72094 130.87526 127.10352
#> [127] 122.82981 124.88965 124.63141 136.60035 140.77368 129.49830 122.37927
#> [134] 116.06888 126.53232 127.74967 118.61753 113.53247 106.81427 101.05657
#> [141]  94.87566  97.65930  78.51115  80.71933  75.12935  74.51142  77.86606
#> [148]  84.93863  85.57383  82.02982  78.56732  82.49333  68.45731  72.06202
#> [155]  59.85736  57.65403  57.51646  60.88723  58.73859  70.07008  69.09882
#> [162]  79.93394  79.17728  71.81692  70.19631  56.89586  61.09950  63.36822
#> [169]  56.29263  63.34917  58.82967  64.28526  53.98266  48.29053  50.97629
#> [176]  52.63674  70.40819  65.39800  69.36492  60.95297  54.90919  44.65612
#> [183]  40.68273  30.13327  26.22159  34.55275  37.68460  30.38561  28.69077
#> [190]  25.37417  28.54009  30.05667  26.21513  27.08165  29.58058  33.71525
#> [197]  39.92543  43.29169  45.85180  38.05564

# 200 periods with bubble origination date 100 and termination date 150
sim_psy1(n = 200, te = 100, tf = 150)

#>   [1]  100.00000  106.39414  113.17598  118.17681  107.80775   98.63149
#>   [7]   93.57575   94.63220   89.56930   89.44739   89.30038   76.00494
#>  [13]   87.54940   91.31231   97.94452   95.89399   93.18542   95.96956
#>  [19]   95.45667  104.25692  110.13708  105.72762   98.99227   98.45647
#>  [25]  102.51726  104.57252  109.53847  103.69138   99.08657   97.49335
#>  [31]  102.08404  101.63011   92.68876   90.71554   91.04820   97.52431
#>  [37]  107.51470  112.82967  110.36598  114.38600  122.05705  119.52860
#>  [43]  116.96332  125.99032  120.75280  124.27553  124.99919  128.94911
#>  [49]  117.63421  123.21102  117.71304  131.69668  124.70793  123.82141
#>  [55]  136.53304  131.57622  122.88887  116.99270  125.09193  115.28483
#>  [61]  117.01939  116.47700  121.79610  129.31909  126.24557  116.72996
#>  [67]  113.26712  123.87925  121.52085  126.16327  131.76828  133.88478
#>  [73]  145.85690  147.63476  148.57646  146.23373  151.24343  153.49025
#>  [79]  154.21045  148.24425  158.18224  150.56509  158.75322  162.75947
#>  [85]  159.12179  167.18694  170.93285  156.40967  151.02755  146.30404
#>  [91]  153.73741  153.05595  162.39001  164.83361  158.64360  161.17052
#>  [97]  166.76633  164.87965  166.85354  163.87069  164.33521  171.13948
#> [103]  186.04869  191.75465  188.38135  198.55014  206.84533  206.43086
#> [109]  209.89287  215.90380  223.92416  229.07116  229.81497  230.85221
#> [115]  243.72700  252.14452  257.89986  258.61956  263.82923  288.95676
#> [121]  302.80197  320.28339  338.43615  336.32144  359.85668  373.41341
#> [127]  382.23395  411.46720  432.62919  448.72582  463.54317  482.39739
#> [133]  495.95501  524.54272  546.50490  581.86131  598.52432  621.72319
#> [139]  638.12474  660.48434  689.76666  708.94677  744.05184  774.44772
#> [145]  811.09546  851.72722  898.24988  941.44931  978.58991 1016.63912
#> [151]  165.58846  165.01236  169.39770  159.96447  170.03901  169.60011
#> [157]  164.13234  150.99447  151.57056  155.08015  164.67952  167.37610
#> [163]  169.88911  181.69852  177.07087  169.51712  166.10872  163.01617
#> [169]  156.53149  154.16960  155.69887  151.65800  160.92883  172.78538
#> [175]  171.60157  178.41463  175.11741  171.54667  170.61968  177.99877
#> [181]  175.67758  187.72204  186.55628  180.61597  174.57252  166.99790
#> [187]  164.27421  165.90356  170.37508  176.15786  171.88688  166.49188
#> [193]  163.75063  172.58721  180.80129  196.33402  201.10986  197.00961
#> [199]  203.97784  207.42321