Simulation of a single-bubble process

The following function generates a time series which switches from a martingale to a mildly explosive process and then back to a martingale.

sim_psy1(
  n,
  te = 0.4 * n,
  tf = 0.15 * n + te,
  c = 1,
  alpha = 0.6,
  sigma = 6.79,
  seed = NULL
)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te
A scalar in (0, tf) specifying the observation in which the bubble originates.

tf
A scalar in (te, n) specifying the observation in which the bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute. Results between seeds in parallel and non-parallel differ.

Value

A numeric vector of length n.

Details

The data generating process is described by the following equation: $$X_t = X_{t-1}1\{t < \tau_e\}+ \delta_T X_{t-1}1\{\tau_e \leq t\leq \tau_f\} + \left(\sum_{k=\tau_f+1}^t \epsilon_k + X^*_{\tau_f}\right) 1\{t > \tau_f\} + \epsilon_t 1\{t \leq \tau_f\}$$

where the autoregressive coefficient \(\delta_T\) is given by:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\) and \(X_{\tau_f} = X_{\tau_e} + X^*\). During the pre- and post- bubble periods, \(N_0 = [1, \tau_e)\), X is a pure random walk process. During the bubble expansion period \(B = [\tau_e, \tau_f]\) is a mildly explosive process with expansion rate given by the autoregressive coefficient \(\delta_T\), and continues its martingale path for the subsequent period \(N_1 = (\tau_f, \tau]\).

For further details the user can refer to Phillips et al. (2015) p. 1054.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination date 40 and termination date 55
sim_psy1(n = 100)

#>   [1] 100.00000 106.43713 105.25400  97.74266  91.31942  93.28234  99.23658
#>   [8]  91.43552  83.71113  73.95223  70.59699  76.30604  81.68056  80.53407
#>  [15]  84.69445  79.45829  85.49208  85.58180  87.11186  82.15573  73.85505
#>  [22]  76.61725  71.51788  70.41711  72.80726  70.84455  71.55521  76.44528
#>  [29]  72.29627  64.78033  68.41164  73.40954  65.10876  72.04417  75.20264
#>  [36]  80.56995  79.68545  73.36881  70.89147  76.40922  84.03563 106.18271
#>  [43]  98.08048 107.12668 111.45830 114.09866 120.95733 131.42027 135.41705
#>  [50] 151.77079 159.74800 158.67072 158.47420 154.54977 159.18699  75.49839
#>  [57]  82.32766  97.04286 112.81302 105.88434 106.21579 100.97467  95.64292
#>  [64]  90.70929  95.33924  93.77860  83.52162  79.55811  65.82249  68.56229
#>  [71]  72.29689  72.48944  78.55403  75.99722  80.11117  80.07807  76.54187
#>  [78]  72.20293  67.88521  68.60893  76.60018  79.63797  95.07133  95.99517
#>  [85]  82.42169  79.56649  76.99710  85.28616  74.82257  72.71556  72.57903
#>  [92]  56.34957  62.39175  52.32345  55.35010  64.65087  64.51433  63.77274
#>  [99]  65.56979  67.63292

# 200 periods with bubble origination date 80 and termination date 110
sim_psy1(n = 200)

#>   [1] 100.00000  98.75479 102.55481 101.28811  95.77276  84.63321  88.08201
#>   [8]  99.99396 104.01635 110.91984 111.74564 104.42556  96.66075  93.06447
#>  [15]  88.43863  87.06401  98.50173  91.48241  90.42333  90.10827  83.63313
#>  [22]  86.45954  87.23380  87.66781  81.42555  87.54561  82.12904  86.66626
#>  [29]  87.72016  88.59396  78.18444  79.55847  74.68638  77.14229  86.64153
#>  [36]  89.17216  78.54143  78.19043  81.68104  85.41486  91.30734  95.95414
#>  [43]  94.84955  82.73371  75.69149  81.32819  85.45208  84.62243  90.95835
#>  [50]  84.43129  86.16329  82.46001  88.77716  85.12652  82.05454  79.07176
#>  [57]  74.88760  77.88630  81.16314  84.84430  69.28987  71.39716  80.93086
#>  [64]  90.28148  97.48157 101.77402 109.14069 118.35105 120.81190 135.54175
#>  [71] 136.48645 145.83171 142.49476 152.91684 150.09085 146.55636 152.33182
#>  [78] 154.60305 148.97176 153.68816 149.59472 157.40562 164.16901 173.43289
#>  [85] 191.57506 209.25859 211.53031 227.23246 236.94784 234.96803 238.16240
#>  [92] 246.52145 250.16436 278.85933 299.92194 321.64191 332.54993 340.49313
#>  [99] 351.22754 374.09618 393.11118 395.25062 411.93538 434.87331 458.02259
#> [106] 480.87207 492.41177 511.49825 535.66948 554.52767 149.53591 165.91108
#> [113] 164.79101 161.80066 165.34488 152.31947 138.85727 142.39622 147.67737
#> [120] 142.16308 148.34707 154.78910 145.63585 148.04070 151.64119 149.52968
#> [127] 147.87186 145.89001 138.23030 130.60594 144.55218 138.37655 141.48894
#> [134] 141.17240 136.54872 126.43957 123.77441 111.89407 111.63151 106.21884
#> [141]  99.99974 100.49723 108.90441 113.78215 117.44513 110.23935 106.56774
#> [148] 101.71150 100.49663 101.61678  96.67768  93.02577  88.95404  96.98042
#> [155]  98.26517  99.95554 107.06225  98.30789 100.85493  97.07345  99.12693
#> [162] 100.63612  94.10300  94.45630 100.56758  97.88620 104.36485 100.37088
#> [169]  90.34487  80.98977  71.85266  66.85088  63.65883  73.03316  84.40649
#> [176]  92.39768  91.38666  90.17924  95.63244  92.80919 101.04200 109.46404
#> [183] 114.11978 113.93782 116.03987 117.73645 108.52608 112.59587 112.65458
#> [190] 113.27055 108.80913 105.53783 105.65840  99.82216 109.04097 117.29124
#> [197] 127.16331 126.57997 122.38667 120.90469

# 200 periods with bubble origination date 100 and termination date 150
sim_psy1(n = 200, te = 100, tf = 150)

#>   [1] 100.000000  90.996425  74.957026  65.383333  67.110689  69.119649
#>   [7]  69.505252  68.808225  82.061782  88.027245  79.155316  79.768754
#>  [13]  85.895765  89.626336  83.720128  83.453860  79.933573  73.742114
#>  [19]  74.762744  77.752757  86.714078  88.388228  94.783230  92.460023
#>  [25]  90.584986  93.190015 103.014536 114.787719 117.886874 122.690823
#>  [31] 136.703952 136.866305 138.536302 140.383583 133.644789 133.457535
#>  [37] 148.550654 149.606525 143.705225 135.730442 123.782233 124.178508
#>  [43] 132.085598 134.378092 127.217910 131.704532 126.113585 129.081043
#>  [49] 131.609558 120.428132 107.367212 109.953068 119.294108 124.901620
#>  [55] 122.076199 128.744995 130.389402 124.020879 134.245658 127.008529
#>  [61] 128.110346 129.878940 119.774048 129.383146 133.202573 137.562131
#>  [67] 131.515414 130.724125 124.328116 131.904588 135.096805 140.940365
#>  [73] 140.462372 136.298879 134.012584 132.547029 136.764517 128.045977
#>  [79] 119.218350 116.660085 117.364538 112.587351 122.754791 120.698597
#>  [85] 111.349912 117.346637 121.905581 130.257433 141.430160 140.188092
#>  [91] 130.822932 139.314510 126.574862 124.070794 109.612866 111.905482
#>  [97] 101.214119  99.881110 106.094368 117.834771 132.604488 126.898724
#> [103] 113.428140 119.944955 131.894824 132.909427 135.819129 139.826394
#> [109] 147.144867 143.213563 154.095861 144.367187 149.932726 152.984779
#> [115] 157.448472 167.905824 173.498805 182.315993 190.656547 200.454967
#> [121] 219.493142 232.485272 241.758322 250.806455 263.054227 277.592417
#> [127] 285.213895 303.484216 311.321360 315.620960 333.738357 344.777897
#> [133] 357.396198 369.766638 393.391682 406.743643 431.406981 450.659099
#> [139] 460.141170 470.481654 475.555807 494.698847 520.007719 538.431949
#> [145] 572.817269 610.417806 638.731603 661.474110 683.347988 714.498346
#> [151] 107.822658 105.137612 104.250143  90.673579  83.911208  86.763782
#> [157]  86.336952  85.792856  85.573652  80.691772  73.110320  67.812288
#> [163]  55.746739  52.841740  39.051067  57.728760  68.006577  68.237362
#> [169]  73.267345  64.638557  63.527956  59.379636  67.879930  69.579599
#> [175]  57.998700  52.190444  51.206562  49.003568  47.831852  39.438985
#> [181]  26.522339  25.880657  26.101711  29.233870  38.613579  35.785705
#> [187]  40.409305  37.595712  34.076146  29.431652  23.418098  23.752418
#> [193]  25.012379  20.879592  15.915404  34.351233  25.256795  20.870352
#> [199]  14.538985   9.316885