Simulation of a two-bubble process

The following data generating process is similar to sim_psy1, with the difference that there are two episodes of mildly explosive dynamics.

sim_psy2(
  n,
  te1 = 0.2 * n,
  tf1 = 0.2 * n + te1,
  te2 = 0.6 * n,
  tf2 = 0.1 * n + te2,
  c = 1,
  alpha = 0.6,
  sigma = 6.79,
  seed = NULL
)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te1
A scalar in (0, n) specifying the observation in which the first bubble originates.

tf1
A scalar in (te1, n) specifying the observation in which the first bubble collapses.

te2
A scalar in (tf1, n) specifying the observation in which the second bubble originates.

tf2
A scalar in (te2, n) specifying the observation in which the second bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute. Results between seeds in parallel and non-parallel differ.

Value

A numeric vector of length n.

Details

The data generating process is described by:

$$X_t = X_{t-1}1\{t \in N_0\}+ \delta_T X_{t-1}1\{t \in B_1 \cup B_2\} + \left(\sum_{k=\tau_{1f}+1}^t \epsilon_k + X^*_{\tau_{1f}}\right) 1\{t \in N_1\} $$

$$ + \left(\sum_{l=\tau_{2f}+1}^t \epsilon_l + X^*_{\tau_{2f}}\right) 1\{t \in N_2\} + \epsilon_t 1\{t \in N_0 \cup B_1 \cup B_2\}$$

where the autoregressive coefficient \(\delta_T\) is given:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\), \(X_{\tau_{1f}} = X_{\tau_{1e}} + X^*\) and \(X_{\tau_{2f}} = X_{\tau_{2e}} + X^*\). We use the notation \(N_0 = [1, \tau_{1e})\), \(B_1 = [\tau_{1e}, \tau_{1f}]\), \(N_1 = (\tau_{1f}, \tau_{2e})\), \(B_2 = [\tau_{2e}, \tau_{2f}]\), \(N_2 = (\tau_{2f}, \tau]\), where \(\tau\) is the last observation of the sample. After the collapse of the first bubble, \(X_t\) resumes a martingale path until time \(\tau_{2e}-1\), and a second episode of exuberance begins at \(\tau_{2e}\). The expansion process lasts until \(\tau_{2f}\) and collapses to a value of \(X^*_{\tau_{2f}}\). The process then continues on a martingale path until the end of the sample period \(\tau\). The expansion duration of the first bubble is assumed to be longer than that of the second bubble, i.e. \(\tau_{1f}-\tau_{1e}>\tau_{2f}-\tau_{2e}\).

For further details the user can refer to Phillips et al., (2015) p. 1055.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination dates 20/60 and termination dates 40/70 respectively
sim_psy2(n = 100)

#>   [1] 100.00000 112.00094 112.35912 113.10416 118.96530 118.29327 118.99737
#>   [8] 127.20299 127.83952 122.73211 125.41364 118.00222 108.10041 107.26720
#>  [15]  99.79189 103.73635 102.73858  97.53517 108.55464 114.65748 131.54526
#>  [22] 139.07507 145.61430 157.33774 174.25802 203.62517 209.44577 221.41541
#>  [29] 242.70385 260.33703 275.49112 284.02702 300.05583 317.73359 337.22595
#>  [36] 369.45059 389.62478 403.50706 432.33742 458.85590 113.26997 117.58548
#>  [43]  97.84916  94.08982  93.04788  97.15781  95.38557  88.56068  91.09041
#>  [50]  91.22245  91.97294  99.82605  96.14437  87.66230  78.97101  72.01851
#>  [57]  62.59636  62.26215  74.55857  78.58723  88.82318  86.94092  90.92155
#>  [64] 100.50210 104.43318 116.35440 128.40845 132.21424 135.31479 159.86216
#>  [71]  77.27982  77.32476  76.95439  73.09627  66.80210  59.09046  65.29626
#>  [78]  58.60090  63.13312  72.68391  80.54523  79.60309  75.94110  68.61541
#>  [85]  68.48937  66.90991  67.97115  63.86992  68.14107  72.99028  80.29555
#>  [92]  95.57139  96.91506  93.00710  92.48209  81.42852  82.31070  79.42038
#>  [99]  82.17720  80.35198

# 200 periods with bubble origination dates 40/120 and termination dates 80/140 respectively
sim_psy2(n = 200)

#>   [1] 100.0000000  90.6231150  90.8715158  81.6829854  87.7736574  75.3523050
#>   [7]  74.3889605  77.4986568  67.1804539  76.6800395  70.6685864  67.2306037
#>  [13]  68.3396287  66.0504219  59.0163110  61.7957155  58.2362383  61.6753758
#>  [19]  58.2483076  51.7485778  51.1060105  55.5157035  57.1610685  56.7609061
#>  [25]  43.5353843  53.3906866  53.1194962  47.2529518  45.8834751  47.6574004
#>  [31]  37.3477945  42.2183201  43.8943143  52.0611026  58.7563092  55.1036925
#>  [37]  51.7746727  45.0151179  42.5591550  50.7429623  65.0243587  74.0350704
#>  [43]  64.6917004  72.9652168  83.0388530  90.6878489  88.6579948  99.5766461
#>  [49] 108.6749680 117.6982908 120.0080404 130.7472995 136.9477126 142.1883654
#>  [55] 158.2047305 169.9319700 157.4910918 161.1508542 157.7596985 168.3247609
#>  [61] 172.3264762 185.6198839 196.5233598 204.8538212 212.8035243 219.2090333
#>  [67] 229.6888416 231.7252043 245.9396541 253.5015150 259.0026500 267.5700139
#>  [73] 291.4373289 291.6323688 301.4035764 314.6740430 324.1224841 334.8224274
#>  [79] 351.3564469 367.4934644  46.1681691  50.0210386  51.2397326  55.8999931
#>  [85]  46.8878729  41.6737202  40.3652702  46.8772278  43.2851049  36.2718022
#>  [91]  41.5119541  44.6717259  36.3494193  50.2463578  60.4223573  50.9913930
#>  [97]  65.6128192  53.8721165  54.5649850  46.8156997  44.2558080  41.8875776
#> [103]  44.0799673  36.4457013  17.7253336  17.0747763  12.0046846  13.4358904
#> [109]   7.6745088  13.0539866  13.2521693  17.3491739  10.5935305   9.1585832
#> [115]  11.4944863  10.8876061  22.1185205   8.7903704  13.9244626  14.6020081
#> [121]  21.6312522  21.8608010   4.1678837   7.1952596   6.0119863  -0.4554715
#> [127]  -4.9318555  -4.7592016  11.6921947  13.4566120   7.3646410  20.5967288
#> [133]  18.4765731  13.6118694   5.7097991   1.3723204  14.5692326  21.3312678
#> [139]  27.2202554  29.3489567  16.5243983  20.5255872  21.9912052  23.0440166
#> [145]  29.5204995  28.3943596  30.1258351  24.0319977  34.7871548  44.2439858
#> [151]  31.9387525  30.4699418  34.2034645  37.4812435  42.6474064  39.6213884
#> [157]  31.4371745  33.4841337  23.0333376  27.3475048  32.1205474  19.1796027
#> [163]  25.5548794  24.0305780  19.4553621  22.4822588  31.1776490  41.8048826
#> [169]  33.6531750  30.6812262  31.6751564  32.1234260  29.1016314  13.1986883
#> [175]  11.3084720   3.2840254   3.5089460  -2.1791525  -2.9777318  -8.0126700
#> [181] -10.7853459 -13.3504094  -4.3307994   0.5338031  -1.0434405  -9.3695211
#> [187]  -2.2567375 -12.0099423 -20.0389063 -25.3093372 -31.9113459 -38.0471807
#> [193] -29.3070659 -30.0280798 -34.8888435 -30.3278498 -28.9859175 -37.2703743
#> [199] -41.0793274 -37.4552828