Simulation of a two-bubble process

The following data generating process is similar to sim_psy1, with the difference that there are two episodes of mildly explosive dynamics.

sim_psy2(n, te1 = 0.2 * n, tf1 = 0.2 * n + te1, te2 = 0.6 * n,
  tf2 = 0.1 * n + te2, c = 1, alpha = 0.6, sigma = 6.79)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te1
A scalar in (0, n) specifying the observation in which the first bubble originates.

tf1
A scalar in (te1, n) specifying the observation in which the first bubble collapses.

te2
A scalar in (tf1, n) specifying the observation in which the second bubble originates.

tf2
A scalar in (te2, n) specifying the observation in which the second bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

Value

A numeric vector of length n.

Details

The data generating process is described by:

$$X_t = X_{t-1}1\{t \in N_0\}+ \delta_T X_{t-1}1\{t \in B_1 \cup B_2\} + \left(\sum_{k=\tau_{1f}+1}^t \epsilon_k + X^*_{\tau_{1f}}\right) 1\{t \in N_1\} $$

$$ + \left(\sum_{l=\tau_{2f}+1}^t \epsilon_l + X^*_{\tau_{2f}}\right) 1\{t \in N_2\} + \epsilon_t 1\{t \in N_0 \cup B_1 \cup B_2\}$$

where the autoregressive coefficient \(\delta_T\) is given:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\), \(X_{\tau_{1f}} = X_{\tau_{1e}} + X^*\) and \(X_{\tau_{2f}} = X_{\tau_{2e}} + X^*\). We use the notation \(N_0 = [1, \tau_{1e})\), \(B_1 = [\tau_{1e}, \tau_{1f}]\), \(N_1 = (\tau_{1f}, \tau_{2e})\), \(B_2 = [\tau_{2e}, \tau_{2f}]\), \(N_2 = (\tau_{2f}, \tau]\), where \(\tau\) is the last observation of the sample. After the collapse of the first bubble, \(X_t\) resumes a martingale path until time \(\tau_{2e}-1\), and a second episode of exuberance begins at \(\tau_{2e}\). The expansion process lasts until \(\tau_{2f}\) and collapses to a value of \(X^*_{\tau_{2f}}\). The process then continues on a martingale path until the end of the sample period \(\tau\). The expansion duration of the first bubble is assumed to be longer than that of the second bubble, i.e. \(\tau_{1f}-\tau_{1e}>\tau_{2f}-\tau_{2e}\).

For further details the user can refer to Phillips et al., (2015) p. 1055.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination dates 20/60 and termination dates 40/70 respectively
sim_psy2(n = 100)

#>   [1] 100.00000  99.81588 102.84878 100.61410 101.38171 108.51956 110.73973
#>   [8] 116.08372 119.08191 112.16083 111.08327  99.96328 119.24189 125.77265
#>  [15] 129.22490 130.98706 125.73565 113.47611 114.68734 118.35359 128.53846
#>  [22] 135.81632 143.90132 167.21092 186.73815 206.67233 214.40354 224.93387
#>  [29] 237.45570 254.92187 270.77459 289.87070 306.65931 327.28273 344.95843
#>  [36] 366.12133 393.52858 424.23556 448.07929 476.43966 118.70158 122.35893
#>  [43] 125.88751 109.43594 103.96123 100.67025  96.23586  94.24120  92.36044
#>  [50]  93.28707 100.53341 100.54582 101.49705  97.96219 114.02605 102.50699
#>  [57] 107.14174 116.38675 129.32819 132.91967 134.08917 153.64233 167.00874
#>  [64] 187.43835 186.13451 191.90088 198.59242 205.10593 216.04536 223.65653
#>  [71] 133.59245 138.48497 145.27815 148.86327 156.76416 157.93976 161.86381
#>  [78] 161.07611 167.21781 167.10519 160.84279 153.64461 141.50953 134.26764
#>  [85] 126.59610 129.22565 137.85074 126.67360 125.83432 119.23688 126.49259
#>  [92] 142.21995 130.79324 131.71583 135.10023 144.15444 143.99858 142.69814
#>  [99] 139.70636 126.52421

# 200 periods with bubble origination dates 40/120 and termination dates 80/140 respectively
sim_psy2(n = 200)

#>   [1] 100.0000000 109.6126486 113.6622023 109.2288600 108.1822176 105.1698636
#>   [7] 101.6510562 103.0505232 114.1824136 103.5578649 107.0315604  96.5772597
#>  [13]  99.0406416 104.2739386  95.5217222  90.6791324  89.9839492  85.5866654
#>  [19]  83.0522719  83.6181995  89.8460980  88.0878050  92.6848098  89.8323258
#>  [25]  96.9428506  86.4415972  79.3392186  63.9077910  61.7577243  51.7343899
#>  [31]  51.0527565  45.9859050  47.3686605  50.7463485  43.8593422  51.6390549
#>  [37]  50.1831515  61.5557180  62.3458034  74.6274937  76.1067205  71.7686973
#>  [43]  76.9487359  80.2995925  85.7989492  82.7862900  95.9725253  93.2536521
#>  [49]  99.8267972 107.5135431 116.8539292 121.3033085 141.7958984 141.6035733
#>  [55] 151.6946755 158.5360262 169.9801797 180.6290216 184.9531118 196.7862229
#>  [61] 210.2283049 213.4135271 221.4254425 228.1643609 237.1544892 239.5781280
#>  [67] 256.1151885 273.1587686 273.6645374 286.4725153 306.0110536 314.3225290
#>  [73] 320.0050220 328.4903661 332.9122127 343.3708815 367.4318386 379.2531465
#>  [79] 395.0272070 423.5787486  81.6305706  77.6875857  66.8705387  63.5499398
#>  [85]  58.4391059  62.9570280  67.9521858  72.9143134  69.3242471  61.0926274
#>  [91]  56.8835130  62.1858070  62.8392067  52.2376384  56.9795994  57.2389485
#>  [97]  58.8322944  65.7671148  75.7832038  75.4545788  71.4167541  68.1982528
#> [103]  65.9822597  57.9316169  51.2958306  46.6604508  46.4395878  56.4988446
#> [109]  56.6906933  52.0188616  45.9436373  42.3234090  46.2878073  37.7129880
#> [115]  43.7777963  37.6859203  41.4517407  33.5481584  33.6748372  37.3093378
#> [121]  36.8597528  42.7506014  50.8223512  64.9706888  78.9608758  82.4374696
#> [127]  96.8383604  91.7198376  96.8526460 103.7464160 104.6692322  98.2002149
#> [133] 111.9100067 123.7018888 124.8830163 126.5052101 130.9413115 132.4109229
#> [139] 136.2398638 147.3517466  54.4548264  56.2883507  50.2699864  45.0449195
#> [145]  48.1279667  43.3297640  36.8833688  29.6425087  23.6235052  28.1632469
#> [151]  14.1244851  25.6277392  24.3391984  27.2057123  22.7031565  18.2029791
#> [157]  18.9722332  25.3748710  17.1190911  16.7981770  15.4941918  19.7864805
#> [163]  18.7550459  18.8582564  24.8926529  24.9628272  27.3049177  21.1044362
#> [169]  14.8277731  12.3013689  10.2793215  20.9701215   7.5214286   6.3926170
#> [175]   7.2089008   2.7157249  -0.8872024  -2.9338354  -7.0234633  -9.1840962
#> [181]  -7.0914451  -1.6686110  10.2177896  12.0193821  14.1728621  13.1629421
#> [187]  15.3508088  13.1494288   7.8304194   7.7076078   6.5587484   5.4401434
#> [193]  10.2856396  11.8962889  -3.8747519  -3.3380412  -6.2496364  -8.7803670
#> [199]  -9.0853644   0.4297547