Simulation of a two-bubble process

The following data generating process is similar to sim_psy1, with the difference that there are two episodes of mildly explosive dynamics.

sim_psy2(
  n,
  te1 = 0.2 * n,
  tf1 = 0.2 * n + te1,
  te2 = 0.6 * n,
  tf2 = 0.1 * n + te2,
  c = 1,
  alpha = 0.6,
  sigma = 6.79,
  seed = NULL
)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te1
A scalar in (0, n) specifying the observation in which the first bubble originates.

tf1
A scalar in (te1, n) specifying the observation in which the first bubble collapses.

te2
A scalar in (tf1, n) specifying the observation in which the second bubble originates.

tf2
A scalar in (te2, n) specifying the observation in which the second bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute. Results between seeds in parallell and non-parallel differ.

Value

A numeric vector of length n.

Details

The data generating process is described by:

$$X_t = X_{t-1}1\{t \in N_0\}+ \delta_T X_{t-1}1\{t \in B_1 \cup B_2\} + \left(\sum_{k=\tau_{1f}+1}^t \epsilon_k + X^*_{\tau_{1f}}\right) 1\{t \in N_1\} $$

$$ + \left(\sum_{l=\tau_{2f}+1}^t \epsilon_l + X^*_{\tau_{2f}}\right) 1\{t \in N_2\} + \epsilon_t 1\{t \in N_0 \cup B_1 \cup B_2\}$$

where the autoregressive coefficient \(\delta_T\) is given:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\), \(X_{\tau_{1f}} = X_{\tau_{1e}} + X^*\) and \(X_{\tau_{2f}} = X_{\tau_{2e}} + X^*\). We use the notation \(N_0 = [1, \tau_{1e})\), \(B_1 = [\tau_{1e}, \tau_{1f}]\), \(N_1 = (\tau_{1f}, \tau_{2e})\), \(B_2 = [\tau_{2e}, \tau_{2f}]\), \(N_2 = (\tau_{2f}, \tau]\), where \(\tau\) is the last observation of the sample. After the collapse of the first bubble, \(X_t\) resumes a martingale path until time \(\tau_{2e}-1\), and a second episode of exuberance begins at \(\tau_{2e}\). The expansion process lasts until \(\tau_{2f}\) and collapses to a value of \(X^*_{\tau_{2f}}\). The process then continues on a martingale path until the end of the sample period \(\tau\). The expansion duration of the first bubble is assumed to be longer than that of the second bubble, i.e. \(\tau_{1f}-\tau_{1e}>\tau_{2f}-\tau_{2e}\).

For further details the user can refer to Phillips et al., (2015) p. 1055.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination dates 20/60 and termination dates 40/70 respectively
sim_psy2(n = 100)

#>   [1] 100.00000  96.41738  99.67369 109.29872 108.75490 106.07837 109.64487
#>   [8] 104.24047 105.90468 100.61309 104.65263 112.18200 105.77924 110.53457
#>  [15] 107.65103  99.88913 101.47419 100.18927  88.89471 107.51112 108.74790
#>  [22] 118.21442 122.55913 125.43246 136.26335 136.91768 154.53359 173.55890
#>  [29] 182.81410 191.46503 211.00406 228.87308 243.52418 274.97171 309.87876
#>  [36] 337.66217 350.47159 375.30266 405.32932 427.69444 118.25978 122.31853
#>  [43] 125.48325 120.42899 126.94098 127.97738 122.75337 132.21239 138.45015
#>  [50] 145.34938 148.97956 142.27356 150.33867 146.82456 145.06214 142.83458
#>  [57] 143.33319 141.65020 132.32167 140.39598 152.11661 163.08337 161.84930
#>  [64] 176.42243 180.79626 198.74803 200.04369 219.93674 233.07352 260.67494
#>  [71] 132.67554 130.77624 124.70511 125.63331 120.54648 124.06505 122.75908
#>  [78] 122.95470 125.38953 125.19266 132.98107 135.51774 137.71321 132.07874
#>  [85] 141.54714 140.24656 142.09532 134.75086 118.93424 115.20228 114.70946
#>  [92] 121.71870 123.17949 119.82292 130.09034 126.15578 137.52760 130.73095
#>  [99] 139.03310 146.34721

# 200 periods with bubble origination dates 40/120 and termination dates 80/140 respectively
sim_psy2(n = 200)

#>   [1] 100.00000  95.84489  99.28652 102.41038 112.48940 118.47793 114.83372
#>   [8] 123.56263 127.55413 118.66953 120.82010 128.92844 135.12810 129.78604
#>  [15] 126.99583 130.23041 131.49278 122.49795 130.45835 132.03019 135.31025
#>  [22] 131.67543 141.03297 132.18783 136.49851 143.28617 140.99286 140.40869
#>  [29] 128.73792 122.42918 127.95404 131.52839 138.39958 144.04465 146.86341
#>  [36] 142.37347 146.09990 142.43759 141.27434 141.94914 143.73950 160.90830
#>  [43] 173.31739 182.82840 187.68996 201.61412 218.09147 226.84731 241.63366
#>  [50] 259.61409 268.57553 277.62565 287.23858 293.74029 305.32598 317.53321
#>  [57] 335.60277 350.41996 373.81100 381.80447 391.98052 398.08359 412.72634
#>  [64] 430.19826 448.04620 446.67304 465.40522 484.11050 508.26294 526.52988
#>  [71] 552.75742 571.83923 594.49531 606.21117 621.02866 639.31140 676.77358
#>  [78] 700.60180 740.40404 761.38557 136.57482 133.14940 135.87780 142.47358
#>  [85] 138.53767 149.43005 150.96441 147.46853 141.87501 144.14569 143.51089
#>  [92] 145.57547 142.33998 140.70148 146.29749 135.73467 136.36955 133.87796
#>  [99] 132.99927 135.76820 131.80306 130.48953 128.65931 129.15945 131.58508
#> [106] 135.32249 135.58324 124.65422 117.52669 131.46000 132.65800 140.31921
#> [113] 143.27287 146.99945 151.39542 157.36019 159.74211 160.08079 165.75553
#> [120] 170.74263 172.47474 179.66569 179.08142 188.99698 193.13382 205.87865
#> [127] 214.03577 214.84144 222.97061 227.44430 225.93512 238.70708 257.46921
#> [134] 257.21451 259.43175 280.97396 301.17492 312.29033 320.95121 337.02021
#> [141] 179.85463 179.59928 183.29080 195.10688 192.90382 189.86522 192.69042
#> [148] 183.89150 188.52657 191.90062 201.67591 209.87388 211.49339 209.46428
#> [155] 218.86586 214.11239 226.74812 238.99677 229.78923 227.65207 220.73126
#> [162] 216.57409 214.61131 224.84372 227.34056 227.20452 220.55085 211.03866
#> [169] 220.85031 217.97487 212.99271 208.18210 218.19798 223.93556 232.72178
#> [176] 234.74629 231.99171 231.04920 229.53780 241.40095 240.84618 241.68910
#> [183] 244.35238 246.92185 248.05018 255.88168 255.98183 256.12741 250.51648
#> [190] 250.83832 248.84812 257.04794 257.26523 258.49013 259.01405 262.80555
#> [197] 261.67261 260.12463 261.77233 255.46793