Simulation of a two-bubble process

The following data generating process is similar to sim_psy1, with the difference that there are two episodes of mildly explosive dynamics.

sim_psy2(n, te1 = 0.2 * n, tf1 = 0.2 * n + te1, te2 = 0.6 * n,
  tf2 = 0.1 * n + te2, c = 1, alpha = 0.6, sigma = 6.79,
  seed = NULL)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te1
A scalar in (0, n) specifying the observation in which the first bubble originates.

tf1
A scalar in (te1, n) specifying the observation in which the first bubble collapses.

te2
A scalar in (tf1, n) specifying the observation in which the second bubble originates.

tf2
A scalar in (te2, n) specifying the observation in which the second bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) shold be initialized. Either NULL or an integer will be used in a call to `set.seed` before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute.

Value

A numeric vector of length n.

Details

The data generating process is described by:

$$X_t = X_{t-1}1\{t \in N_0\}+ \delta_T X_{t-1}1\{t \in B_1 \cup B_2\} + \left(\sum_{k=\tau_{1f}+1}^t \epsilon_k + X^*_{\tau_{1f}}\right) 1\{t \in N_1\} $$

$$ + \left(\sum_{l=\tau_{2f}+1}^t \epsilon_l + X^*_{\tau_{2f}}\right) 1\{t \in N_2\} + \epsilon_t 1\{t \in N_0 \cup B_1 \cup B_2\}$$

where the autoregressive coefficient \(\delta_T\) is given:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\), \(X_{\tau_{1f}} = X_{\tau_{1e}} + X^*\) and \(X_{\tau_{2f}} = X_{\tau_{2e}} + X^*\). We use the notation \(N_0 = [1, \tau_{1e})\), \(B_1 = [\tau_{1e}, \tau_{1f}]\), \(N_1 = (\tau_{1f}, \tau_{2e})\), \(B_2 = [\tau_{2e}, \tau_{2f}]\), \(N_2 = (\tau_{2f}, \tau]\), where \(\tau\) is the last observation of the sample. After the collapse of the first bubble, \(X_t\) resumes a martingale path until time \(\tau_{2e}-1\), and a second episode of exuberance begins at \(\tau_{2e}\). The expansion process lasts until \(\tau_{2f}\) and collapses to a value of \(X^*_{\tau_{2f}}\). The process then continues on a martingale path until the end of the sample period \(\tau\). The expansion duration of the first bubble is assumed to be longer than that of the second bubble, i.e. \(\tau_{1f}-\tau_{1e}>\tau_{2f}-\tau_{2e}\).

For further details the user can refer to Phillips et al., (2015) p. 1055.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination dates 20/60 and termination dates 40/70 respectively
sim_psy2(n = 100)

#>   [1] 100.00000 102.42563 106.16304 106.42379  95.49477  88.36724 102.30055
#>   [8] 103.49856 111.15976 114.11343 117.84000 122.23598 128.20074 130.58266
#>  [15] 130.92134 136.59608 134.68317 129.30766 129.31889 129.41503 140.04138
#>  [22] 145.14673 159.01000 168.62972 171.16540 181.15103 187.77283 188.64335
#>  [29] 203.91275 225.60408 228.86619 234.81668 260.37529 285.30853 301.88848
#>  [36] 316.59732 339.28181 369.80105 392.87856 421.35904 141.23112 139.02806
#>  [43] 135.98946 138.81466 130.01573 134.65081 138.02485 147.80015 155.99811
#>  [50] 157.61763 155.58852 164.99010 160.23663 172.87235 185.12100 175.91347
#>  [57] 173.77631 166.85550 162.69832 171.00112 192.02297 206.63564 219.53743
#>  [64] 226.73564 231.52949 255.94967 269.22356 281.22826 294.16195 322.73819
#>  [71] 176.73869 185.52492 187.54943 184.79485 183.85234 182.34094 194.20409
#>  [78] 193.64931 194.49224 197.15552 199.72499 200.85332 208.68482 208.78496
#>  [85] 208.93055 203.31962 203.64146 201.65126 209.85108 210.06837 211.29327
#>  [92] 211.81719 215.60869 214.47575 212.92776 214.57547 208.27107 218.56068
#>  [99] 213.04992 221.95329
#> attr(,"seed")
#>   [1]       10403         570 -1788416688 -1917108519   406657915 -1738333394
#>   [7] -1253475118  1835410024  -211365254  1728304943  1725048761  -204105152
#>  [13] -1440852909  2111847184  1557887008  1041099242  1271837027  -898457198
#>  [19]  -615671617   679294759   863619794  2120171559  -374110121 -1843783261
#>  [25]    71513931  1852099239  1847714155  1791702448  1255258784  -412438645
#>  [31]  -511927276  1456167919  -580164974   682243157  -473165673  -915475388
#>  [37]   604150156   378544087 -1673807493   605919651   661722851 -1786691336
#>  [43]  -570574134  1525447155  -565222496  1902237407   402349331 -1145229557
#>  [49]  1151647738  -880035509  2052236732  -783528998  1515209315  -267522408
#>  [55]  1687829820  1907561870  -699074372  1284757624   343347296  1892149219
#>  [61]  -269664767 -1238665930  1120455403  -392119316   735130093  1319543126
#>  [67] -1326623128   321009315   -19462619   437109152  1990284116 -1282333214
#>  [73]  1616502344  -769305227   846632451   950765310  1154747701    11901694
#>  [79]   726734030   904734400 -1332044103 -1264427356  -526220487  -950991583
#>  [85] -1415067820  1415337465 -1639141627 -1723386147   162866989  1021561793
#>  [91]   262092777  -114534382 -1049365882 -1417368495  -710811622   816591749
#>  [97]    22311556 -1310782421   955138385   568699178 -1647313798  1286487429
#> [103]  -186708287   703361433  1842030685   408919898    42617584   891423285
#> [109]  -765253290   786810465 -1591914115 -1498064963   910849348   589322297
#> [115]   916461882 -1565821632  1079344649   916441014  -463556770   -59339068
#> [121]  1551082114 -1182122274  1003888162   292365249  1327650091   864659512
#> [127]    55513085   908931046  1553720627  -407288528  1935882950  1674764301
#> [133]  -757011265  2057712330  2118707126   895126980  -434670666 -2124062285
#> [139]  1435466853 -1466441508  1006107479  1384153276   243039180  1790274974
#> [145] -1648220585 -1282963602  1622543539   173653963   -20059626  -260156509
#> [151] -1415061445   667518919  1540243871  1239749899  1709685359  -134330388
#> [157]  -929765260  1059851423 -1979291616 -1180049197   445102654  1147619273
#> [163] -1075399101   913706560   541283160   557824011    94381583  1659396231
#> [169]  1229717655  2031336188 -1099826666  -308006521  -647088124  1492016683
#> [175]  1423212639   106653055   407745238  2095606959  1777204752  2123115502
#> [181]   191240135  1461850724   589965936 -1206710622   877560456  -717832484
#> [187]  1880205940  1017063127  -179072587 -1715597950  1438196447 -2071318008
#> [193] -1645711647   377143762 -2076362548 -1220246257  1766118993  1069253812
#> [199]  1057456592  -852003354   392898140  1210961281   868163719  -806768062
#> [205]   922886097  1232656114  1644144050   585520572  -748565803 -1388244088
#> [211]  1726262453  2095871005  1841042864 -1414099507  -422025967  -272482519
#> [217]   -33024423   313688253   764368501   266129702 -1379115726  -475794467
#> [223]  2091774750  -409915391   445227944   532365993 -1634532467     6837655
#> [229]  1689324633   543364226 -2128909344  1504579714 -2002666302  1520712079
#> [235]   813931602 -1057868484 -1915065838  1836377692  -695662121   414840759
#> [241] -1362245744 -1338127563  -853363590  1289105979  1173588283 -1400451287
#> [247] -1478146060  -915501855 -1614217735 -1802554161  2103649146 -1682470348
#> [253]  1668666006  -930737028 -1346494472   389256448 -1453712159  -415015583
#> [259]  1315376572  1429229610  1074738471   170082856  -690449788  2026071634
#> [265]  1752318066   709023729  1249160394  1293173982  1005479586  -301010068
#> [271]  1225479779  1972806417  -396470858 -1282665797  1666142420   419388687
#> [277]  1219586019    31429041   498841634 -1300948595  1335537921  1086519839
#> [283]  -565823708   915754063   312138655   982840608   509784434 -1933245297
#> [289] -1110917064 -1849748262   -51764539   993823276 -1274114612  1064951256
#> [295]   260879247  1589379751 -1028782000 -1775699847  2070191094  1624136928
#> [301]  2043875468 -1834196749  -360815436  -531171715 -1335954350  2037867453
#> [307] -1159439212  1504511782 -2116380731 -1054289386   824658186  -555750975
#> [313]  1627978250 -1924491058  -374746423  -913722323  1946375627  1909215276
#> [319]  -313926172  -933636373 -1962531064  1420902292   219388363  1783276318
#> [325]  1613650256   -36170230  2094309901   989985409  -891823850  1902909541
#> [331]  -152415136   482264920  -985893378 -1801430775 -1804711014  1493118025
#> [337] -1592221070   -89626163  1590954832 -1530490932  1999448703  1020950440
#> [343]   346214134   209435027   593359940  -366946228  -308022591  -565481885
#> [349]  -636791505  1719914952  1523107662   392998647  1159591240  -125635710
#> [355]   172290153  1734313212   272021908  -104029788   733069843   326074343
#> [361] -1217349668  -427931751 -1881273982 -1774575504  1444119540 -1092780349
#> [367]   331212528 -1109332687  -544209338   574363517 -2091711988    88073154
#> [373] -1379783403  1563488818   -88507918 -1263280623  -335481010 -1973178978
#> [379]  1100219457 -2043000067   623307571   255001444 -2005392548  -498679781
#> [385]  1037417984  1399738288  -159742061  1074755778 -1343090640   219047674
#> [391]  1144533445  -306688555  -994474482 -2100547755 -1983243784  1693427432
#> [397]  -229469326  1043602117  2056356370 -1991111303  1458172586  1581312292
#> [403]  -427918554   979505415 -1915947272  1278540258 -2126082667 -1418962303
#> [409] -1288966338   825083327  -520508483  1722063362  2094421094  2027538600
#> [415]   599438947  1617352391   245848056 -1981652177  -995515657  1535014387
#> [421]  1454654635   480699086 -1703906080   944431149  1538200886 -1322035894
#> [427]   885779094  -965929807   657857201 -1124898057  1400249401  1245874729
#> [433]  1251783926  1360378452 -1555106550 -1070265907  1148412502    71360732
#> [439]  -968255607 -1716302009   743778696  -776920359 -1343888643 -1579333728
#> [445]  -541469422   944061300   216688717  1669523171    54846888 -1620731823
#> [451]  -578935711 -1859446047   145087065  2143533764    70201499   184210148
#> [457]  1318117803  1686067707 -1235411585   238397890  1766530186 -1069169701
#> [463]  1129995978  -277558451   697605791 -2138396878 -1357970144  -359887921
#> [469]   962759284 -1714678082  -916221982  -593474573   603170725  -718178907
#> [475] -1778053900  -484067191   173312019   347740980 -1371166306  -624799490
#> [481]   115302456   735837942  2066403337  1675302333 -1501828250  -823389724
#> [487]  1931313359   447981769   168533884   881899188 -1466868984 -1651566607
#> [493]  1956034975 -1840192516 -1212756685  1025360660 -1830027186   172713759
#> [499]  -370995631  -281036708  1077792903   699092035  -648729727  -393472850
#> [505] -1618825110   942565380  -151149917  1386321074 -1129299798  -967118197
#> [511]  1716019185  -639742389  1788056081   709556557  -592170952    55316732
#> [517]   631779599  1870030395 -1717189568  1445697494  2076263743  -881559043
#> [523]  1219876973 -1800139208 -1186646636   394592321  -597371564  1918919583
#> [529]  1242870197  1157613864  1939364274 -1411655307  1928481994 -1531897728
#> [535]   521498228 -1849616363   162473455   604187195   849795134   -85466913
#> [541]  1144131317   133759782 -1588393680 -1903455500  2103103454 -1427619976
#> [547]  -521219633   691494955 -1841178012 -1235214170 -1467836327 -1254231973
#> [553] -1026004214  -865269458   972452798  -205589849   214083113 -2131618710
#> [559]  -999565419 -1513387834  1922189388  1386664557  -446889853 -1570316318
#> [565]  -588317939   485929229 -1605530133   383018176  -197143998    21117940
#> [571]   299913358 -1793368373 -1176237898 -2128491973 -1261265894 -1625286221
#> [577]  1948329404  1053068770 -1457834866 -1359455188  -550824888  1239947502
#> [583] -2085773898  1259520991 -1722005979 -1000100949 -1528010830  -924186833
#> [589]   296776538  1577279107   358253489 -1887169531   652458182 -2058420675
#> [595]  -592081014  -642585057  -761095889   812168067  1079230012   344291934
#> [601] -1233357053 -1576708981 -1247250521  1845011770  -336625399  1498657682
#> [607] -1619725681  1333125396   263369211  -470997661  1065338289 -1581372653
#> [613]  1593289763  -213361281 -1006044725  -535976050  2026190076 -1383977856
#> [619]  1139512257 -1493508512  1114258995   767580007  -902250648 -1327087205
#> [625]  2066529015  1881339066
#> attr(,"class")
#> [1] "sim"     "numeric"

# 200 periods with bubble origination dates 40/120 and termination dates 80/140 respectively
sim_psy2(n = 200)

#>   [1] 100.00000 105.53770 106.35272 112.34007 100.12661  93.62823  90.40937
#>   [8]  77.77684  79.49940  78.04361  61.95008  61.97220  66.51977  68.05988
#>  [15]  64.04110  61.28833  70.89092  72.53156  62.78052  60.64471  57.38183
#>  [22]  63.05999  62.70454  69.16833  79.31607  76.97886  81.36367  70.94441
#>  [29]  75.01619  71.59159  68.82868  65.69893  80.73815  90.83467  88.42354
#>  [36]  94.37297  92.91518 106.63640  94.15189  99.24545  98.19437  94.68741
#>  [43] 106.76329 117.86990 130.86853 140.61322 150.08451 150.76355 155.95766
#>  [50] 162.33077 166.43561 170.62909 186.40082 199.73536 210.90021 231.30168
#>  [57] 225.88608 234.03386 248.33338 253.27191 253.18014 271.96068 293.49264
#>  [64] 297.76965 314.32994 337.49065 348.52637 364.85844 371.70705 390.42955
#>  [71] 418.02057 435.17425 450.29558 474.02697 489.85325 510.75325 532.41737
#>  [78] 560.34038 601.15298 626.81760  95.71500 100.97337 109.31345 122.08277
#>  [85] 121.23238 116.07813 112.61995 111.56044 112.46068 105.84665  97.35270
#>  [92]  97.03808 104.76094 113.34656 109.62060 114.25404 107.94087 101.31962
#>  [99] 101.13549 104.16839 101.93372 102.70132 109.83917 112.05935 117.40333
#> [106] 120.40152 113.48044 112.40288 101.28289 120.56151 127.09226 130.54451
#> [113] 132.30668 127.05526 114.79573 116.00695 112.43692 115.15419 114.32182
#> [120] 118.59634 137.76326 152.47496 166.97394 168.61574 172.63720 178.15315
#> [127] 188.05297 195.64942 205.80523 212.87142 223.00729 229.31611 238.25951
#> [134] 252.48425 268.87158 277.14035 288.76554 301.13415 317.32702 334.06518
#> [141] 102.14477  96.67006  93.37908  88.94469  86.95003  85.06927  85.99590
#> [148]  93.24224  93.25465  94.20588  90.67102 106.73488  95.21582  99.85057
#> [155] 109.09558 122.03702 117.46844 110.25127 121.34398 125.01622 134.90829
#> [162] 121.77789 115.79997 110.38338 104.36655 102.36467  96.34431  97.01709
#> [169] 101.90961 108.70279 112.28791 120.18879 121.36439 125.28845 124.50075
#> [176] 130.64245 130.52983 124.26743 117.06925 104.93416  97.69228  90.02074
#> [183]  92.65029 101.27537  90.09823  89.25896  82.66152  89.91723 105.64459
#> [190]  94.21788  95.14047  98.52486 107.57908 107.42322 106.12277 103.13100
#> [197]  89.94885  99.56150 103.61105  99.17771
#> attr(,"seed")
#>   [1]       10403         144  1451696842 -2080387128  1383168179 -1126513329
#>   [7]  1288765491 -1266704443 -1497521494   723988118 -1836673119  2133620168
#>  [13]  1339938871  -829155411  2040688375 -1069958136  1575000161  -686503197
#>  [19] -1387201141   387691153   801469411  1769526787 -1954821635 -1919062063
#>  [25] -1298385086   879244426   221215477  2127022182   462336592  1148560540
#>  [31]   872417689   159624184  1907277826   -73882451   787869707   413005872
#>  [37] -1017838240   878418280   187921859  -826058248   243286176 -1464382740
#>  [43]  1842201441    57841752   805927273  1024523435   621502714 -1108636945
#>  [49] -2071439602  1908619923  1260700430  -144411578 -1486655715 -1355597057
#>  [55] -1443721178   596616711   428496632   236239019 -1416523318  -529609100
#>  [61]  1063969632  1311885525   979760180 -1515506525  -417950864  1529673214
#>  [67]  1072521411 -1879593646  -846444574 -1805669238  -213640898  1229874512
#>  [73]   701893851   156267580 -1057650548 -1735323168  -714360614 -2085779821
#>  [79]  -103601175 -1961789552  1324037222  1193040093   543318705  1393329042
#>  [85]  -654482219  -318490796 -1234158068  2074724399  1234519259  -304436252
#>  [91]   639993024  1790076479  -128298941 -1035883771  1557249772  1950122205
#>  [97]   330814774  1925799236   770643841 -1676935565  -183187198  1250179566
#> [103]   583086159  -271518602   629441518  -670036231  1796579691 -1229773631
#> [109]  -162251797  2138815170 -1728868429  -409634296  1321365064  2098746220
#> [115]  -930351893 -1817076278   822418070  -291805289  2104973214  1267851342
#> [121]  1941141844 -1001469103  -946489623  -984031371  -760173087  -803541172
#> [127] -1885169205   234911698   609854169 -1512749001  -629038266   -93554891
#> [133]   967626829  1287400041 -1317625961  1106900241  2097879174  -321232743
#> [139] -2054494085  -849581925   222521957   899212760  -809675248  -740578047
#> [145]  1580908689  1208008630   313523662  1111243477  1272948342   741708045
#> [151]  -472181993   648956532   194648828    58324401 -1899834451 -1500373200
#> [157]   183746494  1022707630   912437777  1099004726  1003030353  1116748907
#> [163]   -17788954  1655955936  -617496905  2122687323   169500542  1631603609
#> [169]   545221747 -1082822432  -116470564 -1735864768  -197286850 -1568497538
#> [175]   255745195 -1245523491   643648691 -2145098990 -1603924924  -392520064
#> [181]   357926096 -1198644810 -1292772483 -2062789364  1668499328   -73054068
#> [187]  -866787221   989675040 -2000900758  1463305730  -894493141  -968035083
#> [193]   356400746   390340567  1294919261  -884382927 -1701626777 -1019850910
#> [199]   -14068500 -1726543359 -1921587812 -1065050464   863085439  1962264628
#> [205] -2034679822 -2057853570  2089521124  1532482748   267598422  1963856906
#> [211]  -401904699   963427107  1866388314  -719632410  1331135576 -1191503920
#> [217]   588885502  1271390766  -404500707  1847794277  1163617713  1035823694
#> [223]   205190847   254734311 -1156233492  1551470193  1775950735  -797144060
#> [229]  1611890427  1443534650  -282540151  1445045970  1216375127  1418776090
#> [235] -1811511973  -531589213  1570594232   550151061 -1434960180  1488665599
#> [241]   260195048   523449546  1062868426  -435033053  1482628520   -22913423
#> [247]  -358333378  -114704960 -1170735941  -171288512 -2142852405 -1020862967
#> [253]   271269556  1525708041   895353574 -1775563521 -1361265933  1395887943
#> [259]   589005805  -928912818   -24574535    93960521  1694255897  -143907495
#> [265] -1200471473   772350886  -397837417   326595569 -1615842470 -1874798449
#> [271] -1598981873    69620658   -35764951  1420609680   681350327   970478624
#> [277]  -400276844  1173954335  2002143327  1504085826   375060271 -2033087564
#> [283]  -506613505 -1982886936   324977211 -1532966541  2131033324  1636729724
#> [289]   116277944  -585607799  -130715211   -49162858    76552978 -1591506469
#> [295]   966859998  -321174582  1178820713 -1896459067  2032749344   356484253
#> [301] -1727073126 -1245005098  1907546625 -1924873485  1713276818  -459209309
#> [307]  -406326269  -686895269   663237078  2029110836  -445921363 -1468458544
#> [313]  -341738701  1538025399 -2040159514 -1432592671  -717454350   276916786
#> [319] -1751761515 -1109834809 -1474743089 -1018117418  1092211026   597752222
#> [325]  1277515329  -896666024   411284460  -796973687 -1422228221   431233663
#> [331] -1584424998   119191825   189305506   555910310  1136919098  -422024494
#> [337] -1691926789  -140932581  -168411666 -1403237464  1668638904 -1025862768
#> [343]   210039260   548135092  -611644495 -1834082783 -1111693408 -2017565924
#> [349] -1223744587  -353249970  1474705617  -264349883  -329542003  1401442592
#> [355]  1034925484   742587923 -1724704539   374469264  1792243737   571771811
#> [361] -1500792454  -157705386   182773033   369285500  1790734119  -866903293
#> [367]  1252306652  1653303622   -41001287  -837495018 -1857790752  -774674748
#> [373]   647546543   804471095 -2123537150  1038081489  1761888194   437887464
#> [379] -1128538155 -1073619014    81818627 -1963015677  1320800226  1410949310
#> [385]   356671606 -1806221049   562480983   741347145    68620054  1172580315
#> [391]  -848857131 -1184006736 -1494547410  1325355906   322410733   425076170
#> [397]  2036845917 -1002041899    86603811 -1622581909 -1918708500  1006723384
#> [403]  1912301953   549140431  -432136221   969800273   606339871  1287680847
#> [409]  2031121334   461749884  -161605619  1561882035  -943084328  1149242245
#> [415] -1815681636 -1880160618  -893270198   838041487   855342803 -1486050080
#> [421]   420966064  1075239981   241476600 -1220233476  -609229369 -1521704537
#> [427]   589901958  1206212635  1207278340 -1010688580  -925090305 -1544067703
#> [433] -1377994668  1831662945 -1572478843  1841990525  1999529060  1378291934
#> [439]  -323113889   957630366  1691051605 -2119318023  -401230848  1287997303
#> [445]   400916372  2023640036  -957687349  1157175141 -1425598791   -74623920
#> [451]  -558640568  1716460319  1670927123  -221908195  -721296010  -564936975
#> [457]   -46298344  1829698025   289108531  2082858770  -306536664  -173299138
#> [463]  1323451098 -1872780760   546304012   -33571748 -1258959161  1395853266
#> [469]   332742549  1542181435   290027781 -1867612507  1295144060 -1582906812
#> [475]   296997211  1712654445    -6859302 -1755021820 -1371015562  1397001896
#> [481]  1327336050 -1849435587  1053701886 -1109307584   876971957 -1013058731
#> [487]  1559269109   -74659321   533848595  1890903709   547561086  1529982559
#> [493]   656495192   714780113  1308201595 -1765395587  1282268127   222842616
#> [499]  1941540764    80223669  -646848914 -1656835672  1302224503 -1478157663
#> [505]   435039005  -719312431   821447963   134403962  2096382510   199307224
#> [511]  1009261586   -14697492   680330506  1365173488   539527938 -1926042912
#> [517]  1940641355  -186149995 -1778114947 -1602746030   496043514  -990691273
#> [523]  -427735850   451017379  1547461690  1461395850  -184089203   621550751
#> [529] -1749721534  1216280280    27369110   526885367  -151418397 -1471331527
#> [535]  1189540750   -94956546  -201792522   -63394894   201480544   280022166
#> [541]  1481696804 -1765005698  -304506981 -1416502243    99972494  2129257901
#> [547]  1883915021 -1590044163 -1738485115 -1940902495  -533286548   759784388
#> [553] -1397001521  1148750387  1890118661  1951622856   432869450  1521667279
#> [559]  1442807154  1912714884 -1854317233  -396217692 -1209334493  1817768866
#> [565]   552231490  2094549956  -415860280  1500152217 -2105831830    74766747
#> [571] -1288765682 -2022109206  1261849571  -405427795  1361501978  -853560981
#> [577]  -174877761  1131127446 -1488815789 -2083644247   930897239  1014594423
#> [583]  -792514781   211127080  -311912550   950356928  -546752277  -296517417
#> [589]  1083213000 -1723788242 -2000133215   969782084  1476121794   620652825
#> [595] -1755664746   232235761   254295560 -1318222850  -611197973 -1532080655
#> [601]  -403088467  1118957070  1260621644  -973391975  1991276577   705238605
#> [607] -1480904912 -1708024799   570482797 -2012307995  -573651480  -483326280
#> [613]   888921959  -628487315  1547099278  2020971880  1394946459   899288810
#> [619] -1441787264  1591127016 -1055405842   908838745  -668092968 -1644617479
#> [625]   -61376632   778758470
#> attr(,"class")
#> [1] "sim"     "numeric"