Simulation of a two-bubble process

The following data generating process is similar to sim_psy1, with the difference that there are two episodes of mildly explosive dynamics.

sim_psy2(n, te1 = 0.2 * n, tf1 = 0.2 * n + te1, te2 = 0.6 * n,
  tf2 = 0.1 * n + te2, c = 1, alpha = 0.6, sigma = 6.79,
  seed = NULL)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te1
A scalar in (0, n) specifying the observation in which the first bubble originates.

tf1
A scalar in (te1, n) specifying the observation in which the first bubble collapses.

te2
A scalar in (tf1, n) specifying the observation in which the second bubble originates.

tf2
A scalar in (te2, n) specifying the observation in which the second bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute.

Value

A numeric vector of length n.

Details

The data generating process is described by:

$$X_t = X_{t-1}1\{t \in N_0\}+ \delta_T X_{t-1}1\{t \in B_1 \cup B_2\} + \left(\sum_{k=\tau_{1f}+1}^t \epsilon_k + X^*_{\tau_{1f}}\right) 1\{t \in N_1\} $$

$$ + \left(\sum_{l=\tau_{2f}+1}^t \epsilon_l + X^*_{\tau_{2f}}\right) 1\{t \in N_2\} + \epsilon_t 1\{t \in N_0 \cup B_1 \cup B_2\}$$

where the autoregressive coefficient \(\delta_T\) is given:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\), \(X_{\tau_{1f}} = X_{\tau_{1e}} + X^*\) and \(X_{\tau_{2f}} = X_{\tau_{2e}} + X^*\). We use the notation \(N_0 = [1, \tau_{1e})\), \(B_1 = [\tau_{1e}, \tau_{1f}]\), \(N_1 = (\tau_{1f}, \tau_{2e})\), \(B_2 = [\tau_{2e}, \tau_{2f}]\), \(N_2 = (\tau_{2f}, \tau]\), where \(\tau\) is the last observation of the sample. After the collapse of the first bubble, \(X_t\) resumes a martingale path until time \(\tau_{2e}-1\), and a second episode of exuberance begins at \(\tau_{2e}\). The expansion process lasts until \(\tau_{2f}\) and collapses to a value of \(X^*_{\tau_{2f}}\). The process then continues on a martingale path until the end of the sample period \(\tau\). The expansion duration of the first bubble is assumed to be longer than that of the second bubble, i.e. \(\tau_{1f}-\tau_{1e}>\tau_{2f}-\tau_{2e}\).

For further details the user can refer to Phillips et al., (2015) p. 1055.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination dates 20/60 and termination dates 40/70 respectively
sim_psy2(n = 100)

#>   [1] 100.00000 102.42563 106.16304 106.42379  95.49477  88.36724 102.30055
#>   [8] 103.49856 111.15976 114.11343 117.84000 122.23598 128.20074 130.58266
#>  [15] 130.92134 136.59608 134.68317 129.30766 129.31889 129.41503 140.04138
#>  [22] 145.14673 159.01000 168.62972 171.16540 181.15103 187.77283 188.64335
#>  [29] 203.91275 225.60408 228.86619 234.81668 260.37529 285.30853 301.88848
#>  [36] 316.59732 339.28181 369.80105 392.87856 421.35904 141.23112 139.02806
#>  [43] 135.98946 138.81466 130.01573 134.65081 138.02485 147.80015 155.99811
#>  [50] 157.61763 155.58852 164.99010 160.23663 172.87235 185.12100 175.91347
#>  [57] 173.77631 166.85550 162.69832 171.00112 192.02297 206.63564 219.53743
#>  [64] 226.73564 231.52949 255.94967 269.22356 281.22826 294.16195 322.73819
#>  [71] 176.73869 185.52492 187.54943 184.79485 183.85234 182.34094 194.20409
#>  [78] 193.64931 194.49224 197.15552 199.72499 200.85332 208.68482 208.78496
#>  [85] 208.93055 203.31962 203.64146 201.65126 209.85108 210.06837 211.29327
#>  [92] 211.81719 215.60869 214.47575 212.92776 214.57547 208.27107 218.56068
#>  [99] 213.04992 221.95329

# 200 periods with bubble origination dates 40/120 and termination dates 80/140 respectively
sim_psy2(n = 200)

#>   [1] 100.00000 105.53770 106.35272 112.34007 100.12661  93.62823  90.40937
#>   [8]  77.77684  79.49940  78.04361  61.95008  61.97220  66.51977  68.05988
#>  [15]  64.04110  61.28833  70.89092  72.53156  62.78052  60.64471  57.38183
#>  [22]  63.05999  62.70454  69.16833  79.31607  76.97886  81.36367  70.94441
#>  [29]  75.01619  71.59159  68.82868  65.69893  80.73815  90.83467  88.42354
#>  [36]  94.37297  92.91518 106.63640  94.15189  99.24545  98.19437  94.68741
#>  [43] 106.76329 117.86990 130.86853 140.61322 150.08451 150.76355 155.95766
#>  [50] 162.33077 166.43561 170.62909 186.40082 199.73536 210.90021 231.30168
#>  [57] 225.88608 234.03386 248.33338 253.27191 253.18014 271.96068 293.49264
#>  [64] 297.76965 314.32994 337.49065 348.52637 364.85844 371.70705 390.42955
#>  [71] 418.02057 435.17425 450.29558 474.02697 489.85325 510.75325 532.41737
#>  [78] 560.34038 601.15298 626.81760  95.71500 100.97337 109.31345 122.08277
#>  [85] 121.23238 116.07813 112.61995 111.56044 112.46068 105.84665  97.35270
#>  [92]  97.03808 104.76094 113.34656 109.62060 114.25404 107.94087 101.31962
#>  [99] 101.13549 104.16839 101.93372 102.70132 109.83917 112.05935 117.40333
#> [106] 120.40152 113.48044 112.40288 101.28289 120.56151 127.09226 130.54451
#> [113] 132.30668 127.05526 114.79573 116.00695 112.43692 115.15419 114.32182
#> [120] 118.59634 137.76326 152.47496 166.97394 168.61574 172.63720 178.15315
#> [127] 188.05297 195.64942 205.80523 212.87142 223.00729 229.31611 238.25951
#> [134] 252.48425 268.87158 277.14035 288.76554 301.13415 317.32702 334.06518
#> [141] 102.14477  96.67006  93.37908  88.94469  86.95003  85.06927  85.99590
#> [148]  93.24224  93.25465  94.20588  90.67102 106.73488  95.21582  99.85057
#> [155] 109.09558 122.03702 117.46844 110.25127 121.34398 125.01622 134.90829
#> [162] 121.77789 115.79997 110.38338 104.36655 102.36467  96.34431  97.01709
#> [169] 101.90961 108.70279 112.28791 120.18879 121.36439 125.28845 124.50075
#> [176] 130.64245 130.52983 124.26743 117.06925 104.93416  97.69228  90.02074
#> [183]  92.65029 101.27537  90.09823  89.25896  82.66152  89.91723 105.64459
#> [190]  94.21788  95.14047  98.52486 107.57908 107.42322 106.12277 103.13100
#> [197]  89.94885  99.56150 103.61105  99.17771