Simulation of a two-bubble process

The following data generating process is similar to sim_psy1, with the difference that there are two episodes of mildly explosive dynamics.

sim_psy2(
  n,
  te1 = 0.2 * n,
  tf1 = 0.2 * n + te1,
  te2 = 0.6 * n,
  tf2 = 0.1 * n + te2,
  c = 1,
  alpha = 0.6,
  sigma = 6.79,
  seed = NULL
)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te1
A scalar in (0, n) specifying the observation in which the first bubble originates.

tf1
A scalar in (te1, n) specifying the observation in which the first bubble collapses.

te2
A scalar in (tf1, n) specifying the observation in which the second bubble originates.

tf2
A scalar in (te2, n) specifying the observation in which the second bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute. Results between seeds in parallel and non-parallel differ.

Value

A numeric vector of length n.

Details

The data generating process is described by:

$$X_t = X_{t-1}1\{t \in N_0\}+ \delta_T X_{t-1}1\{t \in B_1 \cup B_2\} + \left(\sum_{k=\tau_{1f}+1}^t \epsilon_k + X^*_{\tau_{1f}}\right) 1\{t \in N_1\} $$

$$ + \left(\sum_{l=\tau_{2f}+1}^t \epsilon_l + X^*_{\tau_{2f}}\right) 1\{t \in N_2\} + \epsilon_t 1\{t \in N_0 \cup B_1 \cup B_2\}$$

where the autoregressive coefficient \(\delta_T\) is given:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\), \(X_{\tau_{1f}} = X_{\tau_{1e}} + X^*\) and \(X_{\tau_{2f}} = X_{\tau_{2e}} + X^*\). We use the notation \(N_0 = [1, \tau_{1e})\), \(B_1 = [\tau_{1e}, \tau_{1f}]\), \(N_1 = (\tau_{1f}, \tau_{2e})\), \(B_2 = [\tau_{2e}, \tau_{2f}]\), \(N_2 = (\tau_{2f}, \tau]\), where \(\tau\) is the last observation of the sample. After the collapse of the first bubble, \(X_t\) resumes a martingale path until time \(\tau_{2e}-1\), and a second episode of exuberance begins at \(\tau_{2e}\). The expansion process lasts until \(\tau_{2f}\) and collapses to a value of \(X^*_{\tau_{2f}}\). The process then continues on a martingale path until the end of the sample period \(\tau\). The expansion duration of the first bubble is assumed to be longer than that of the second bubble, i.e. \(\tau_{1f}-\tau_{1e}>\tau_{2f}-\tau_{2e}\).

For further details the user can refer to Phillips et al., (2015) p. 1055.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination dates 20/60 and termination dates 40/70 respectively
sim_psy2(n = 100)

#>   [1] 100.00000 102.52303 104.93642  98.25507  99.71308  99.16930  89.51141
#>   [8]  97.03574 104.35861 101.36931  96.08554  83.73731  76.10480  83.30576
#>  [15]  73.26539  62.73029  68.00956  75.26427  74.01771  89.26838  93.45292
#>  [22] 105.66774 115.12433 113.69804 115.55964 107.39635 113.31437 130.29694
#>  [29] 128.72180 146.80580 151.02161 158.48475 170.77716 177.42753 186.61560
#>  [36] 197.47674 215.46907 227.21111 256.21007 279.78873  94.31597  86.11460
#>  [43]  88.34049  84.71350  93.43150  93.62795  90.94130  86.22314  76.09192
#>  [50]  85.89842  83.54217  83.26881  91.73345  82.59206  78.69527  73.90181
#>  [57]  61.29003  59.61310  57.33464  59.25894  66.11809  67.16595  71.80398
#>  [64]  71.29231  61.09989  63.37366  73.31289  71.32859  72.34757  70.64191
#>  [71]  49.11179  47.98513  58.72852  65.61767  57.83309  51.65521  72.76656
#>  [78]  65.50430  60.75055  59.40312  64.45230  64.11171  73.02796  73.61199
#>  [85]  69.33530  73.78322  73.91491  78.64258  92.28281  91.89592  93.52834
#>  [92]  93.37971  87.73852  97.26132  92.34625 101.38089  95.87194 108.06024
#>  [99] 112.72243 113.32931

# 200 periods with bubble origination dates 40/120 and termination dates 80/140 respectively
sim_psy2(n = 200)

#>   [1]  100.000000  102.203679  102.687929  104.270328  117.757825  111.295980
#>   [7]  107.502799  109.239779  107.894842  110.806320  111.286850  100.852970
#>  [13]  103.935524  109.872690  101.061052   93.734878   86.029247   86.182798
#>  [19]   86.037077   84.144212   83.979850   89.581738   84.487997   88.128866
#>  [25]   77.450114   70.752272   84.220321   71.649461   65.471999   52.221722
#>  [31]   46.787845   34.094879   28.998499   24.986027   19.947173   24.656120
#>  [37]   24.252360   11.596512    2.942597   -9.033134  -12.862511  -25.191351
#>  [43]  -25.965140  -27.889186  -33.210744  -33.502331  -39.394377  -43.307915
#>  [49]  -40.880474  -35.600793  -44.786557  -40.425090  -33.865682  -39.932913
#>  [55]  -49.027345  -48.590743  -57.457898  -56.081310  -55.454956  -57.047932
#>  [61]  -61.145584  -70.321307  -70.571880  -66.618430  -65.104096  -71.096502
#>  [67]  -84.419890  -83.745734  -78.885015  -81.099707  -74.926636  -75.351378
#>  [73]  -76.271668  -89.355620  -98.502502  -99.411544 -111.359540 -117.839319
#>  [79] -119.638783 -124.734210  -12.710559   -6.763786   -1.733495   -2.852571
#>  [85]   -7.951619  -16.461976  -23.498020  -23.697908  -25.564776  -22.268561
#>  [91]  -16.042253  -22.732738  -32.857916  -27.292884  -20.411867  -13.257887
#>  [97]  -13.742888   -7.275789   -9.883404  -12.719994  -15.913115  -27.389439
#> [103]  -33.809368  -36.138419  -33.704459  -33.377257  -25.703387  -29.286011
#> [109]  -26.029694  -16.404670  -16.948487  -19.625016  -16.058517  -21.462916
#> [115]  -19.798712  -25.090293  -21.050760  -13.521391  -19.924151  -15.998211
#> [121]  -19.547721  -28.123351  -27.709001  -30.147379  -42.696906  -31.466738
#> [127]  -38.323338  -37.313662  -41.981055  -48.588258  -47.694239  -57.622939
#> [133]  -51.044667  -43.894638  -47.417493  -52.275230  -46.992922  -44.393560
#> [139]  -46.031378  -31.865361    1.559291    9.790678    1.295052    4.012858
#> [145]   10.359523    7.150095   17.898754   21.957506   25.122220   20.067964
#> [151]   26.579950   27.616355   22.392342   31.851360   38.089121   44.988352
#> [157]   48.618527   41.912529   49.977638   46.463527   44.701113   42.473554
#> [163]   42.972160   41.289170   31.960645   31.686018   34.548262   35.917110
#> [169]   24.393176   28.754310   21.996633   28.540929   17.296436   24.567583
#> [175]   23.827294   36.722770   29.002331   27.103034   21.031898   21.960103
#> [181]   16.873270   20.391842   19.085872   19.281491   21.716323   21.519447
#> [187]   29.307862   31.844531   34.040004   28.405531   37.873932   36.573351
#> [193]   38.422112   31.077648   15.261026   11.529066   11.036247   18.045486
#> [199]   19.506276   16.149707