Simulation of a two-bubble process

The following data generating process is similar to sim_psy1, with the difference that there are two episodes of mildly explosive dynamics.

sim_psy2(
  n,
  te1 = 0.2 * n,
  tf1 = 0.2 * n + te1,
  te2 = 0.6 * n,
  tf2 = 0.1 * n + te2,
  c = 1,
  alpha = 0.6,
  sigma = 6.79,
  seed = NULL
)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te1
A scalar in (0, n) specifying the observation in which the first bubble originates.

tf1
A scalar in (te1, n) specifying the observation in which the first bubble collapses.

te2
A scalar in (tf1, n) specifying the observation in which the second bubble originates.

tf2
A scalar in (te2, n) specifying the observation in which the second bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute. Results between seeds in parallel and non-parallel differ.

Value

A numeric vector of length n.

Details

The data generating process is described by:

$$X_t = X_{t-1}1\{t \in N_0\}+ \delta_T X_{t-1}1\{t \in B_1 \cup B_2\} + \left(\sum_{k=\tau_{1f}+1}^t \epsilon_k + X^*_{\tau_{1f}}\right) 1\{t \in N_1\} $$

$$ + \left(\sum_{l=\tau_{2f}+1}^t \epsilon_l + X^*_{\tau_{2f}}\right) 1\{t \in N_2\} + \epsilon_t 1\{t \in N_0 \cup B_1 \cup B_2\}$$

where the autoregressive coefficient \(\delta_T\) is given:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\), \(X_{\tau_{1f}} = X_{\tau_{1e}} + X^*\) and \(X_{\tau_{2f}} = X_{\tau_{2e}} + X^*\). We use the notation \(N_0 = [1, \tau_{1e})\), \(B_1 = [\tau_{1e}, \tau_{1f}]\), \(N_1 = (\tau_{1f}, \tau_{2e})\), \(B_2 = [\tau_{2e}, \tau_{2f}]\), \(N_2 = (\tau_{2f}, \tau]\), where \(\tau\) is the last observation of the sample. After the collapse of the first bubble, \(X_t\) resumes a martingale path until time \(\tau_{2e}-1\), and a second episode of exuberance begins at \(\tau_{2e}\). The expansion process lasts until \(\tau_{2f}\) and collapses to a value of \(X^*_{\tau_{2f}}\). The process then continues on a martingale path until the end of the sample period \(\tau\). The expansion duration of the first bubble is assumed to be longer than that of the second bubble, i.e. \(\tau_{1f}-\tau_{1e}>\tau_{2f}-\tau_{2e}\).

For further details the user can refer to Phillips et al., (2015) p. 1055.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination dates 20/60 and termination dates 40/70 respectively
sim_psy2(n = 100)

#>   [1] 100.00000  94.05141  89.21631  85.10883  87.36180  84.75311  97.42885
#>   [8]  97.81048  81.68089  81.27534  84.15345  93.78032 104.08058 110.54905
#>  [15] 107.09405 100.76904  94.69854  99.46478  97.45804 105.99134 119.71713
#>  [22] 129.19924 134.02392 146.00175 167.70949 185.19233 200.34304 199.50495
#>  [29] 226.83694 236.91273 258.24313 271.63658 299.77451 316.73458 329.25128
#>  [36] 352.08281 369.22297 388.28121 415.56271 443.57307 103.76806 101.97231
#>  [43] 104.21494 102.37326  98.58102  95.27229 101.50060 104.63595 120.98416
#>  [50] 125.64724 117.14173 118.00861 111.97878 118.31724 124.67779 128.00838
#>  [57] 127.46833 132.49601 134.15529 142.20552 140.59473 148.21237 151.53516
#>  [64] 165.35802 181.10176 180.96501 201.23740 207.51398 218.57734 231.60336
#>  [71] 138.75021 134.38675 141.52913 134.87420 138.79783 140.66323 142.59114
#>  [78] 137.17170 136.48266 141.50297 128.29109 125.24386 126.53498 116.44772
#>  [85] 106.04413 102.30194 101.18126  95.76091 102.05287  87.68353  81.05123
#>  [92]  81.28421  84.38245  83.74963  84.16520  75.85541  74.65354  75.73050
#>  [99]  93.71664  94.76601

# 200 periods with bubble origination dates 40/120 and termination dates 80/140 respectively
sim_psy2(n = 200)

#>   [1] 100.00000  99.14710 103.89107  94.42266  91.71705  94.85671  86.84824
#>   [8]  96.29239  92.19383  97.01266  91.42792  93.11657  98.75998 105.33542
#>  [15] 107.05335 105.17357  94.68564 100.08333  90.56239  80.59489  83.45554
#>  [22]  88.74520  79.60683  88.56466  91.00935  77.93857  75.65313  82.51860
#>  [29]  85.05811  89.55742  93.20100  92.70932  96.92722 108.56851 104.56064
#>  [36]  99.76472  96.44020  91.04831  93.85503 100.76013 100.54120 111.57002
#>  [43] 115.52094 122.51060 124.90092 120.45614 119.31496 129.90290 126.40690
#>  [50] 138.92867 145.53472 162.28925 179.81993 177.38353 191.38909 206.92400
#>  [57] 233.31249 231.00675 218.64107 229.82715 238.21105 252.59161 263.04644
#>  [64] 276.86073 292.79933 307.63718 315.22490 334.16561 345.67690 362.82766
#>  [71] 379.17316 382.21436 394.65711 410.01294 435.33170 453.74021 475.32674
#>  [78] 503.87600 530.73588 557.01286  90.69357  91.61828 100.41125  91.64774
#>  [85]  85.01328  95.74488  94.39557  95.60570 102.00655  97.09545  94.40333
#>  [92]  96.54782  87.68895  75.40327  69.16774  71.39528  72.73869  64.97175
#>  [99]  64.53375  59.70129  46.47006  48.54109  38.28212  34.96432  48.14422
#> [106]  48.74260  45.54080  47.96131  64.51511  70.23293  75.92722  74.71996
#> [113]  72.42611  85.21702  84.43260  77.14951  78.32875  82.41226  83.06954
#> [120]  84.18367  90.19338  83.49672  69.55013  75.48328  75.53894  78.27247
#> [127]  77.52769  82.45688  86.77385  91.80930  87.05322  86.17201  86.74435
#> [134]  86.34957  86.48496  87.91876  97.96160 105.74630 116.17306 117.09258
#> [141]  79.75471  83.73587  72.93457  77.03303  82.23282  87.65312  87.02574
#> [148]  80.69530  82.74829  93.53818  99.74328  96.50952  86.09987  74.27014
#> [155]  75.55369  83.76472  82.37976  71.80450  74.32781  77.44809  71.58984
#> [162]  76.15899  72.13050  71.21454  70.89992  73.28998  84.30883  78.91319
#> [169]  77.34810  67.54665  63.98698  72.58509  67.72228  68.71010  75.23211
#> [176]  70.65397  78.29054  95.08178 106.89776  96.50392 106.12550 103.43088
#> [183] 106.75794 111.63738  99.49764 107.69148  97.02105 104.51427 116.25333
#> [190] 115.46569 114.96872 117.02030 118.01511 128.23602 128.74305 131.49273
#> [197] 132.38881 137.79400 140.23630 145.09624