Simulation of a two-bubble process

The following data generating process is similar to sim_psy1, with the difference that there are two episodes of mildly explosive dynamics.

sim_psy2(
  n,
  te1 = 0.2 * n,
  tf1 = 0.2 * n + te1,
  te2 = 0.6 * n,
  tf2 = 0.1 * n + te2,
  c = 1,
  alpha = 0.6,
  sigma = 6.79,
  seed = NULL
)

Arguments

n
A strictly positive integer specifying the length of the simulated output series.

te1
A scalar in (0, n) specifying the observation in which the first bubble originates.

tf1
A scalar in (te1, n) specifying the observation in which the first bubble collapses.

te2
A scalar in (tf1, n) specifying the observation in which the second bubble originates.

tf2
A scalar in (te2, n) specifying the observation in which the second bubble collapses.

c
A positive scalar determining the autoregressive coefficient in the explosive regime.

alpha
A positive scalar in (0, 1) determining the value of the expansion rate in the autoregressive coefficient.

sigma
A positive scalar indicating the standard deviation of the innovations.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute. Results between seeds in parallel and non-parallel differ.

Value

A numeric vector of length n.

Details

The data generating process is described by:

$$X_t = X_{t-1}1\{t \in N_0\}+ \delta_T X_{t-1}1\{t \in B_1 \cup B_2\} + \left(\sum_{k=\tau_{1f}+1}^t \epsilon_k + X^*_{\tau_{1f}}\right) 1\{t \in N_1\} $$

$$ + \left(\sum_{l=\tau_{2f}+1}^t \epsilon_l + X^*_{\tau_{2f}}\right) 1\{t \in N_2\} + \epsilon_t 1\{t \in N_0 \cup B_1 \cup B_2\}$$

where the autoregressive coefficient \(\delta_T\) is given:

$$\delta_T = 1 + cT^{-a}$$

with \(c>0\), \(\alpha \in (0,1)\), \(\epsilon \sim iid(0, \sigma^2)\), \(X_{\tau_{1f}} = X_{\tau_{1e}} + X^*\) and \(X_{\tau_{2f}} = X_{\tau_{2e}} + X^*\). We use the notation \(N_0 = [1, \tau_{1e})\), \(B_1 = [\tau_{1e}, \tau_{1f}]\), \(N_1 = (\tau_{1f}, \tau_{2e})\), \(B_2 = [\tau_{2e}, \tau_{2f}]\), \(N_2 = (\tau_{2f}, \tau]\), where \(\tau\) is the last observation of the sample. After the collapse of the first bubble, \(X_t\) resumes a martingale path until time \(\tau_{2e}-1\), and a second episode of exuberance begins at \(\tau_{2e}\). The expansion process lasts until \(\tau_{2f}\) and collapses to a value of \(X^*_{\tau_{2f}}\). The process then continues on a martingale path until the end of the sample period \(\tau\). The expansion duration of the first bubble is assumed to be longer than that of the second bubble, i.e. \(\tau_{1f}-\tau_{1e}>\tau_{2f}-\tau_{2e}\).

For further details the user can refer to Phillips et al., (2015) p. 1055.

References

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 5 6(4), 1043-1078.

See also

Examples

# 100 periods with bubble origination dates 20/60 and termination dates 40/70 respectively
sim_psy2(n = 100)

#>   [1] 100.00000  90.17027  91.34197  90.33051  89.93001  94.49605 101.42274
#>   [8] 108.23794 114.59584 123.62262 124.33159 129.90024 126.11976 128.68428
#>  [15] 131.22399 125.94629 139.26821 127.39008 129.25347 126.84196 125.07322
#>  [22] 132.83727 151.18435 166.19384 169.01062 187.92752 207.81295 214.58099
#>  [29] 224.86784 236.96389 248.18119 239.31156 263.91941 282.39386 286.63471
#>  [36] 310.20082 342.85277 346.57668 375.51947 399.57277 128.48469 141.77915
#>  [43] 126.39804 131.30911 127.72803 127.46845 118.08101 131.71600 145.12159
#>  [50] 147.88506 146.55051 148.78536 136.20339 135.98301 139.41999 129.74897
#>  [57] 124.03674 118.64626 128.87573 132.05696 138.71302 153.42122 155.19537
#>  [64] 157.48945 163.26518 176.40362 194.29686 191.39921 197.65231 203.96336
#>  [71] 142.41590 142.82198 140.52458 132.78933 123.53168 115.62176 111.61892
#>  [78] 119.49130 115.51002 109.80370 105.77223 107.81413 109.25775 111.81194
#>  [85] 115.14139 118.76900 115.05172 117.97863 121.81148 129.99649 140.26551
#>  [92] 139.06706 144.22806 142.57058 135.39730 136.84582 131.92334 128.86793
#>  [99] 134.82533 131.40912

# 200 periods with bubble origination dates 40/120 and termination dates 80/140 respectively
sim_psy2(n = 200)

#>   [1] 100.00000  96.46814  85.41520  85.67073  74.63623  70.73488  76.85253
#>   [8]  74.67971  71.20998  78.14650  79.57001  83.31588  88.02518  79.36219
#>  [15]  86.90485  89.13389  90.98039  94.51475 112.83219 100.80063 118.48463
#>  [22] 119.49647 132.46772 121.49820 118.54290 117.42436 125.40744 123.61098
#>  [29] 123.09889 117.71723 115.46321 119.69958 117.93424 128.75905 131.54299
#>  [36] 135.99074 140.15168 134.60412 132.07603 143.53221 146.93639 156.96625
#>  [43] 143.43177 155.90309 163.32492 177.81432 193.81809 210.17453 215.70624
#>  [50] 212.15856 232.42017 242.59743 236.89087 242.83614 257.99015 264.72932
#>  [57] 277.54133 290.23445 310.42399 334.01653 337.07466 351.27629 361.87267
#>  [64] 374.42051 397.89975 410.82815 430.90238 467.45573 495.96714 521.80267
#>  [71] 538.21071 567.10404 591.00696 611.37421 622.55139 650.39172 676.69604
#>  [78] 705.91886 737.89452 764.70857 145.30818 140.47137 143.74452 138.04742
#>  [85] 137.77054 138.03083 138.09728 139.58017 146.39225 146.94877 147.43599
#>  [92] 150.65580 146.62876 148.29867 156.12874 159.33124 160.36623 170.78040
#>  [99] 169.80099 164.10521 165.77213 167.21625 172.14885 164.65436 169.95594
#> [106] 163.35072 163.17831 153.93878 149.24388 134.09238 127.22983 135.43756
#> [113] 142.27822 149.15152 159.55121 161.08995 168.71629 180.05698 182.36063
#> [120] 182.42417 194.13780 203.04449 217.97918 228.59067 228.33882 224.37116
#> [127] 230.13313 243.93353 266.02279 272.38886 290.40690 291.94114 304.03770
#> [134] 323.15071 329.68128 344.96313 357.17118 382.89813 390.56639 407.47030
#> [141] 183.10116 180.40603 182.45419 175.42824 175.03957 181.89467 177.84708
#> [148] 177.80071 181.64109 186.42826 177.37410 176.42481 178.42865 174.47122
#> [155] 179.15336 171.07136 164.83676 167.36409 157.16565 159.61058 152.71344
#> [162] 140.14640 139.16219 130.39499 119.21241 118.31637 120.99102 122.18884
#> [169] 119.41038 120.85979 121.11296 124.11535 131.00515 157.66345 154.54276
#> [176] 143.26906 139.63406 141.34741 143.02346 141.84505 148.42931 150.47675
#> [183] 157.17148 151.41070 152.10394 164.88849 169.72718 159.81103 158.33285
#> [190] 158.96791 162.17623 161.60151 168.58682 171.21982 170.94212 166.59058
#> [197] 162.23500 148.13651 152.85231 146.08320