Summarizing radf models

summary method for radf models that consist of radf_obj and radf_cv.

# S3 method for radf_obj
summary(object, cv = NULL, ...)

Arguments

object
An object of class radf_obj. The output of radf().

cv
An object of class radf_cv. The output of radf_mc_cv(), radf_wb_cv() or radf_sb_cv().

...
Further arguments passed to methods. Not used.

Value

Returns a list of summary statistics, which include the estimated ADF, SADF, and GSADF test statistics and the corresponding critical values

Examples

# \donttest{
# Simulate bubble processes, compute the test statistics and critical values
rsim_data <- radf(sim_data)

# Summary, diagnostics and datestamp (default)
summary(rsim_data)

#> Using 'radf_crit' for 'cv'.

#> 
#> -- Summary (minw = 19, lag = 0) ------------------------------------------------ Monte Carlo (nrep = 2000) --
#> 
#> psy1 :
#> # A tibble: 3 x 5
#>   name  tstat   `90`    `95`  `99`
#>   <fct> <dbl>  <dbl>   <dbl> <dbl>
#> 1 adf   -2.46 -0.413 -0.0812 0.652
#> 2 sadf   1.95  0.988  1.29   1.92 
#> 3 gsadf  5.19  1.71   1.97   2.57 
#> 
#> psy2 :
#> # A tibble: 3 x 5
#>   name  tstat   `90`    `95`  `99`
#>   <fct> <dbl>  <dbl>   <dbl> <dbl>
#> 1 adf   -2.86 -0.413 -0.0812 0.652
#> 2 sadf   7.88  0.988  1.29   1.92 
#> 3 gsadf  7.88  1.71   1.97   2.57 
#> 
#> evans :
#> # A tibble: 3 x 5
#>   name  tstat   `90`    `95`  `99`
#>   <fct> <dbl>  <dbl>   <dbl> <dbl>
#> 1 adf   -5.83 -0.413 -0.0812 0.652
#> 2 sadf   5.28  0.988  1.29   1.92 
#> 3 gsadf  5.99  1.71   1.97   2.57 
#> 
#> div :
#> # A tibble: 3 x 5
#>   name  tstat   `90`    `95`  `99`
#>   <fct> <dbl>  <dbl>   <dbl> <dbl>
#> 1 adf   -1.95 -0.413 -0.0812 0.652
#> 2 sadf   1.11  0.988  1.29   1.92 
#> 3 gsadf  1.34  1.71   1.97   2.57 
#> 
#> blan :
#> # A tibble: 3 x 5
#>   name  tstat   `90`    `95`  `99`
#>   <fct> <dbl>  <dbl>   <dbl> <dbl>
#> 1 adf   -5.15 -0.413 -0.0812 0.652
#> 2 sadf   3.93  0.988  1.29   1.92 
#> 3 gsadf 11.0   1.71   1.97   2.57 
#> 

#Summary, diagnostics and datestamp (wild bootstrap critical values)

wb <- radf_wb_cv(sim_data)

#>  1/5 2/5 3/5 4/5 5/5

summary(rsim_data, cv = wb)

#> 
#> -- Summary (minw = 19, lag = 0) --------------------------------------------- Wild Bootstrap (nboot = 500) --
#> 
#> psy1 :
#> # A tibble: 3 x 5
#>   name  tstat   `90`   `95`   `99`
#>   <fct> <dbl>  <dbl>  <dbl>  <dbl>
#> 1 adf   -2.46 -0.539 -0.369 0.0285
#> 2 sadf   1.95  1.42   1.97  2.69  
#> 3 gsadf  5.19  2.72   3.20  4.49  
#> 
#> psy2 :
#> # A tibble: 3 x 5
#>   name  tstat   `90`   `95`   `99`
#>   <fct> <dbl>  <dbl>  <dbl>  <dbl>
#> 1 adf   -2.86 -0.638 -0.479 -0.261
#> 2 sadf   7.88  2.94   3.59   5.35 
#> 3 gsadf  7.88  3.84   5.04   7.22 
#> 
#> evans :
#> # A tibble: 3 x 5
#>   name  tstat   `90`   `95`    `99`
#>   <fct> <dbl>  <dbl>  <dbl>   <dbl>
#> 1 adf   -5.83 -0.589 -0.373 -0.0745
#> 2 sadf   5.28  5.22   7.08  13.2   
#> 3 gsadf  5.99  8.37   9.85  14.9   
#> 
#> div :
#> # A tibble: 3 x 5
#>   name  tstat   `90`  `95`  `99`
#>   <fct> <dbl>  <dbl> <dbl> <dbl>
#> 1 adf   -1.95 -0.364 0.167 0.772
#> 2 sadf   1.11  0.928 1.22  1.93 
#> 3 gsadf  1.34  1.72  2.03  2.45 
#> 
#> blan :
#> # A tibble: 3 x 5
#>   name  tstat   `90`    `95`   `99`
#>   <fct> <dbl>  <dbl>   <dbl>  <dbl>
#> 1 adf   -5.15 -0.221 -0.0883  0.470
#> 2 sadf   3.93  3.12   3.89    6.49 
#> 3 gsadf 11.0   6.40   7.67   10.1  
#> 

# }