Wild Bootstrap Critical values

wb_cv performs the Harvey et al. (2016) wild bootstrap re-sampling scheme, which is asymptotically robust to non-stationary volatility, to generate critical values for the recursive unit root tests. wb_dist computes the distribution.

wb_cv(data, minw = NULL, nboot = 1000L, dist_rad = FALSE, seed = NULL)

wb_distr(data, minw = NULL, nboot = 1000L, dist_rad = FALSE, seed = NULL)

Arguments

data
A univariate or multivariate numeric ts object, data.frame or matrix. The estimation process cannot handle NA values.

minw
A positive integer. The minimum window size, which defaults to \((0.01 + 1.8/\sqrt(T))*T\).

nboot
A positive integer indicating the number of bootstraps. Default is 1000 repetitions.

dist_rad
Logical. If TRUE then the Rademacher distribution will be used.

seed
An object specifying if and how the random number generator(rng) should be initialized. Either NULL or an integer will be used in a call to set.seed before simulation. If set, the value is save as "seed" attribute of the returned value. The default, NULL will note change the rng state, and return .Random.seed as the "seed" attribute. Results between seeds in parallel and non-parallel differ.

Value

A list that contains the critical values for ADF, BADF, BSADF and GSADF t-statistics.

Details

This approach involves applying a wild bootstrap re-sampling scheme to construct the bootstrap analogue of the Phillips et al. (2015) test which is asymptotically robust to non-stationary volatility.

References

Harvey, D. I., Leybourne, S. J., Sollis, R., & Taylor, A. M. R. (2016). Tests for explosive financial bubbles in the presence of non-stationary volatility. Journal of Empirical Finance, 38(Part B), 548-574.

Phillips, P. C. B., Shi, S., & Yu, J. (2015). Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500. International Economic Review, 56(4), 1043-1078.

See also

Examples

# \donttest{
# Default minimum window
wb <- wb_cv(sim_data)

# Change the minimum window and the number of bootstraps
wb <- wb_cv(im_data, nboot = 500, minw = 20)

#> Error in parse_dt(x): object 'im_data' not found

tidy(wdist)

#> Error in tidy(wdist): object 'wdist' not found

# Simulate distribution
wdist <- wb_distr(im_data)

#> Error in parse_dt(x): object 'im_data' not found

autoplot(wdist)

#> Error in autoplot(wdist): object 'wdist' not found

# }